+++ .github/CONTRIBUTING.md
@@ -1,0 +1,37 @@
+<!---
+Derived from the CONTRIBUTING.md file for dplyr https://github.com/tidyverse/dplyr/blob/main/.github/CONTRIBUTING.md
+-->
+
+# Contributing to hudr
+
+This outlines how to propose changes to hudr...
+
+### Fixing typos
+
+Typos or grammatical errors in the package may be edited using
+the GitHub web interface.
+
+### Prerequisites
+
+Before making substantial pull request, you should file an issue and
+get confirmation from a team member. If you've found a
+bug, create an associated issue and illustrate the bug.
+
+### Pull request process
+
+*  You should create a Git branch for each pull request (PR).  
+The `README` should contain badges for any continuous integration services used
+by the package.  
+*  Our team use [roxygen2](https://cran.r-project.org/package=roxygen2), with
+[Markdown syntax](https://cran.r-project.org/web/packages/roxygen2/vignettes/rd-formatting.html), for documentation.  
+*  Our team use [testthat](https://cran.r-project.org/package=testthat). Contributions
+with test cases included are easier to accept.  
+*  For user-facing changes, add a bullet to the top of `NEWS.md` below the current
+development version header describing the changes made followed by your GitHub
+username, and links to relevant issue(s)/PR(s).
+
+### Code of Conduct
+
+Please note that this project is released with a [Contributor Code of
+Conduct](). By participating in this project you agree to
+abide by its terms.
