### 2022-04-01 120ebe8
Adding Sys.getenv('HUD_KEY') == '' checks to tests and outputed vignettes to not run when it does not have access to a key. Currently still issues in rmdcheck because cannot find pdflatex program.
