+++ tests/testthat/test-test_decomp_cw.R
@@ -1,4 +1,5 @@
 test_that("Query hud_cw_zip_tract()", {
+  skip_if(Sys.getenv("HUD_KEY") == "")
   res <- hud_cw_zip_tract(zip = '35213', year = c('2010'), quarter = c('1'))
   expect_true(nrow(res) >= 1)
 
@@ -10,6 +11,7 @@ test_that("Query hud_cw_zip_tract()", {
 })
 
 test_that("Query hud_cw_zip_county()", {
+  skip_if(Sys.getenv("HUD_KEY") == "")
   res <- hud_cw_zip_county(zip = 35213, year = c('2020'), quarter = c('2'))
   expect_true(nrow(res) >= 1)
 
@@ -21,6 +23,7 @@ test_that("Query hud_cw_zip_county()", {
 })
 
 test_that("Query hud_cw_zip_cbsa()", {
+  skip_if(Sys.getenv("HUD_KEY") == "")
   res <- hud_cw_zip_cbsa(zip = 35213, year = c('2020'), quarter = c('2'))
   expect_true(nrow(res) >= 1)
 
@@ -32,6 +35,7 @@ test_that("Query hud_cw_zip_cbsa()", {
 })
 
 test_that("Query hud_cw_zip_cbsadiv()", {
+  skip_if(Sys.getenv("HUD_KEY") == "")
   res <- hud_cw_zip_cbsadiv(zip = 22031, year = c('2020'), quarter = c('2'))
   expect_true(nrow(res) >= 1)
 
@@ -43,6 +47,7 @@ test_that("Query hud_cw_zip_cbsadiv()", {
 })
 
 test_that("Query hud_cw_zip_cd()", {
+  skip_if(Sys.getenv("HUD_KEY") == "")
   res <- hud_cw_zip_cd(zip = 35213, year = c('2020'), quarter = c('2'))
   expect_true(nrow(res) >= 1)
 
@@ -55,6 +60,7 @@ test_that("Query hud_cw_zip_cd()", {
 
 
 test_that("Query hud_cw_tract_zip()", {
+  skip_if(Sys.getenv("HUD_KEY") == "")
   res <- hud_cw_tract_zip(tract = 48201223100, year = c('2017'), quarter = c('1'))
   expect_true(nrow(res) >= 1)
 
@@ -67,6 +73,7 @@ test_that("Query hud_cw_tract_zip()", {
 
 
 test_that("Query hud_cw_county_zip()", {
+  skip_if(Sys.getenv("HUD_KEY") == "")
   res <- hud_cw_county_zip(county = '22031', year = c('2010'), quarter = c('1'))
   expect_true(nrow(res) >= 1)
 
@@ -78,6 +85,7 @@ test_that("Query hud_cw_county_zip()", {
 })
 
 test_that("Query hud_cw_cbsa_zip()", {
+  skip_if(Sys.getenv("HUD_KEY") == "")
   res <- hud_cw_cbsa_zip(cbsa = '10140', year = c('2017'), quarter = c('2'))
   expect_true(nrow(res) >= 1)
 
@@ -89,6 +97,7 @@ test_that("Query hud_cw_cbsa_zip()", {
 })
 
 test_that("Query hud_cw_cbsadiv_zip()", {
+  skip_if(Sys.getenv("HUD_KEY") == "")
   res <- hud_cw_cbsadiv_zip(cbsadiv = 10380, year = c('2017'), quarter = c('4'))
   expect_true(nrow(res) >= 1)
 
@@ -100,6 +109,7 @@ test_that("Query hud_cw_cbsadiv_zip()", {
 })
 
 test_that("Query hud_cw_cd_zip", {
+  skip_if(Sys.getenv("HUD_KEY") == "")
[…]
