+++ R/hudexchange.R
@@ -1,12 +1,11 @@
 #' @import rio
 #' @import zoo
 
-# A R interface for accessing HUD EXCHANGE (US Department of Housing and Urban Development)
-# DRGR Datasets at https://drgr.hud.gov/public/index.html#
-# The HUD EXCHANGE has three main datasets:
-# CDBG-DR (Community Development Block Grant - Disaster Recovery)
-# NSP (Neighborhood Stabilization Program)
-# RIF (Rural Innovation Fund)
+# A R interface for accessing HUD EXCHANGE (US Department of Housing and Urban
+# Development) DRGR Datasets at https://drgr.hud.gov/public/index.html# The HUD
+# EXCHANGE has three main datasets: CDBG-DR (Community Development Block Grant -
+# Disaster Recovery) NSP (Neighborhood Stabilization Program) RIF (Rural
+# Innovation Fund)
 
 # Currently there is no data for RIF (Rural Innovation Fund) and therefore
 # there are no functions to grab that data. A main issue with these datasets
@@ -17,14 +16,15 @@
 #' @name hud_cdbg
 #' @title hud_cdbg
 #' @description This will grab data from
-#'   https://drgr.hud.gov/public/data_downloads.html?programName=DR%20CDBG
-#'   and format it into a dataframe. It will unroll each column to replace
-#'   NA values with the first non-NA value above it.
-#' @param file The specific file needed. Currently you can specify 1 -> 4. Will default to 1:
-#'   1) DR CDBG CDBG-DR Financial Report by Appropriation (xlsx - 0.1 MB - 1/1/2022)
-#'   2) DR CDBG CDBG-DR Financial Report by Grantee (xlsx - 0.1 MB - 1/1/2022)
-#'   3) DR CDBG CDBG-DR Financial Report Monthly Summary (xlsx - 0.1 MB - 1/1/2022)
-#'   4) DR CDBG CDBG-DR Performance by Activity (xlsx - 2.2 MB - 1/1/2022)
+#'   https://drgr.hud.gov/public/data_downloads.html?programName=DR%20CDBG and
+#'   format it into a dataframe. It will unroll each column to replace NA values
+#'   with the first non-NA value above it.
+#' @param file The specific file needed. Currently you can specify 1 -> 4. Will
+#'   default to 1: 1) DR CDBG CDBG-DR Financial Report by Appropriation (xlsx -
+#'   0.1 MB - 1/1/2022) 2) DR CDBG CDBG-DR Financial Report by Grantee (xlsx -
+#'   0.1 MB - 1/1/2022) 3) DR CDBG CDBG-DR Financial Report Monthly Summary
+#'   (xlsx - 0.1 MB - 1/1/2022) 4) DR CDBG CDBG-DR Performance by Activity (xlsx
+#'   - 2.2 MB - 1/1/2022)
 #' @returns A dataframe.
 #' @export
 #' @examples
@@ -45,7 +45,12 @@ hud_cdbg <- function(file = 1) {
   if(as.integer(file) > 4 || as.integer(file) < 1) stop("File number out of range.")
 
   if(as.integer(file) == 1) {
-    data <- suppressMessages(import("https://drgr.hud.gov/public/downloads/DR-CDBG/CDBG-DR%20Financial%20Report%20by%20Appropriation.xlsx"))
+    data <- suppressMessages(import(paste("https://drgr.hud.gov/public/",
+                                          "downloads/DR-CDBG/CDBG-DR",
+                                          "%20Financial",
+                                          "%20Report",
+                                          "%20by",
+                                          "%20Appropriation.xlsx", sep = "")))
 
     # Rename first two rows to their 'appropriate' names.
[…]
