+++ R/hudinputcheckcleansing.R
@@ -9,21 +9,42 @@
 #' @returns The cleansed input arguments.
 #' @noRd
 chas_input_check_cleansing <- function(query, year, key) {
-  if(!is.vector(year) || !is.vector(key)) stop("Make sure all inputs are of type vector. Check types with typeof([variable]). If list try unlist([variable]); if matrix try as.vector([variable])")
-  if(key == "") stop("Did you forget to set the key? Please go to https://www.huduser.gov/hudapi/public/register?comingfrom=1 to and sign up and get a token. Then save this to your environment using Sys.setenv('HUD_KEY' = YOUR_KEY)")
+  if(!is.vector(year) || !is.vector(key)) {
+    stop(paste("Make sure all inputs are of type vector. ",
+               "Check types with typeof([variable]). ",
+               "If list try unlist([variable]); ",
+               "if matrix try as.vector([variable])", sep = ""))
+  }
+  if(key == "") {
+    stop(paste("Did you forget to set the key?",
+               "Please go to https://www.huduser.gov/",
+               "hudapi/public/register?comingfrom=1 to",
+               "and sign up and get a token. Then save",
+               "this to your environment using",
+               "Sys.setenv('HUD_KEY' = YOUR_KEY)", sep = " "))
+  }
+
   if(length(key) != 1) stop("There seems to be multiple keys specified.")
-  
+
   year <- unique(paste(trimws(as.character(year), which = "both")))
   key <- paste(trimws(as.character(key), which = "both"))
-  
-  for(i in 1:length(year)) {
-    if(!any(year %in% c("2014-2018","2013-2017","2012-2016","2011-2015","2010-2014",
-                        "2009-2013","2008-2012","2007-2011","2006-2010"))) stop("One of the years does not fall in the correct range of values.")
+
+  for(i in seq_len(length(year))) {
+    if(!any(year %in% c("2014-2018","2013-2017","2012-2016","2011-2015",
+                        "2010-2014","2009-2013","2008-2012","2007-2011",
+                        "2006-2010"))) {
+      stop("One of the years does not fall in the correct range of values.")
+    }
   }
-  
+
   if(!missing(query)) {
-    if(!is.vector(query)) stop("Make sure all inputs are of type vector. Check types with typeof([variable]). If list try unlist([variable]); if matrix try as.vector([variable])")
-    
+    if(!is.vector(query)) {
+      stop(paste("Make sure all inputs are of type vector. ",
+                 "Check types with typeof([variable]). ",
+                 "If list try unlist([variable]); ",
+                 "if matrix try as.vector([variable])", sep = ""))
+    }
+
     query <- paste(trimws(as.character(query), which = "both"))
     return(list(query, year, key))
   }
@@ -44,76 +65,116 @@ chas_input_check_cleansing <- function(query, year, key) {
 #' @returns The cleansed input arguments.
 #' @noRd
 cw_input_check_cleansing <- function(query, year, quarter, key) {
-  if(!is.vector(query) || !is.vector(year) || !is.vector(quarter) || !is.vector(key)) stop("Make sure all inputs are of type vector. Check types with typeof([variable]). If list try unlist([variable]); if matrix try as.vector([variable])")
-  if(key == "") stop("Did you forget to set the key? Please go to https://www.huduser.gov/hudapi/public/register?comingfrom=1 to and sign up and get a token. Then save this to your environment using Sys.setenv('HUD_KEY' = YOUR_KEY)")
+  if(!is.vector(query) || !is.vector(year) ||
+     !is.vector(quarter) || !is.vector(key)) {
[…]
