### 2022-04-06 0327117
deleted hudaddingleadingzeros.R and renamed it to more broad name. Added functions inside the new hudmischelpers file to house functions useful for parsing geoid data. They are not fully implemented. Hopefully this time fmr and il don't return mix of vectors and list in certain types of queries. Added an very experimental function for calling OS to open wewebsite docs quickly for reference. Fixed tests in test-test_misc to accomodate for change in the hud_metropolitan function to be able to filter by stateid.
