### 2022-04-06 3457e3d
updating documentation for R package. Website and vignettes need to be corrected to use the new hud_metropolitan() function.
