+++ man/hud_cw_zip_countysub.Rd
@@ -13,19 +13,20 @@ hud_cw_zip_countysub(
 )
 }
 \arguments{
-\item{zip}{5 digit USPS ZIP code of the data to retrieve. E.g. 22031 for type 1 to 5 and 11 .}
+\item{zip}{5 digit USPS ZIP code of the data to retrieve. E.g. 22031 for type
+1 to 5 and 11 .}
 
-\item{year}{Gets the year that this data was recorded.
-Can specify multiple years. Default is the
-previous year.}
+\item{year}{Gets the year that this data was recorded. Can specify multiple
+years. Default is the previous year.}
 
 \item{quarter}{Gets the quarter of the year that this data was recorded.
 Defaults to the first quarter of the year.}
 
-\item{minimal}{Return just the crosswalked GEOIDs if true. Otherwise, return all fields.}
+\item{minimal}{Return just the crosswalked GEOIDs if true. Otherwise, return
+all fields.}
 
-\item{key}{The API key for this user. You must go to HUD and sign up for
-an account and request for an API key.}
+\item{key}{The API key for this user. You must go to HUD and sign up for an
+account and request for an API key.}
 }
 \value{
 This function returns a dataframe containing CROSSWALK data for
@@ -34,9 +35,9 @@ oth-ratio, tot-ratio. For more details on these measurements, visit
 https://www.huduser.gov/portal/dataset/uspszip-api.html
 }
 \description{
-This function queries the Crosswalks API provided by
-US Department of Housing and Urban Development. This
-returns the crosswalk for zip to countysub.
+This function queries the Crosswalks API provided by US
+Department of Housing and Urban Development. This returns the crosswalk for
+zip to countysub.
 }
 \examples{
 \dontrun{
@@ -44,7 +45,8 @@ library(hudr)
 
 hud_cw_zip_countysub(zip = 35213, year = c('2017'), quarter = c('1'))
 
-hud_cw_zip_countysub(zip = '35213', year = c('2010'), quarter = c('1'), minimal = TRUE)
+hud_cw_zip_countysub(zip = '35213', year = c('2010'), quarter = c('1'),
+   minimal = TRUE)
 }
 }
 \keyword{API}
