+++ man/hud_chas_county.Rd
@@ -7,7 +7,11 @@
 hud_chas_county(county, year = c("2014-2018"), key = Sys.getenv("HUD_KEY"))
 }
 \arguments{
-\item{county}{The county to query for. Must supply a geoid. 2 digit state fip + 3 digit county fip. hud_counties() will show an extra 99999 at the end. Just remove that.}
+\item{county}{The county to query for. Must supply a geoid. 2 digit state fip
+\itemize{
+\item 3 digit county fip. hud_counties() will show an extra 99999 at the end.
+Just remove that.
+}}
 
 \item{year}{The years to query for.
 \itemize{
