+++ man/hud_cw.Rd
@@ -13,7 +13,8 @@ hud_cw(
 )
 }
 \arguments{
-\item{type}{Must be a number between 1 and 12 depending on the Crosswalk type. You can also supply the string name.
+\item{type}{Must be a number between 1 and 12 depending on the Crosswalk
+type. You can also supply the string name.
 \enumerate{
 \item zip-tract
 \item zip-county
@@ -29,7 +30,8 @@ hud_cw(
 \item countysub-zip (Available 2nd Quarter 2018 onwards)
 }}
 
-\item{query}{5 digit USPS ZIP code of the data to retrieve. E.g. 22031 for type 1 to 5 and 11 .
+\item{query}{5 digit USPS ZIP code of the data to retrieve.
+E.g. 22031 for type 1 to 5 and 11 .
 or
 11 digit unique 2000 or 2010 Census tract GEOID consisting of
 state FIPS + county FIPS + tract code. Eg: 51059461700  for type 6
@@ -37,7 +39,8 @@ or
 5 digit unique 2000 or 2010 Census county GEOID consisting of
 state FIPS + county FIPS. Eg: 51600 for type 7
 or
-5 digit CBSA code for Micropolitan and Metropolitan Areas Eg: 10380 for type 8
+5 digit CBSA code for Micropolitan and Metropolitan Areas
+Eg: 10380 for type 8
 or
 5-digit CBSA Division code which only applies to Metropolitan Areas.
 Eg: 35614 for type 9
@@ -73,29 +76,41 @@ library(hudr)
 
 Sys.setenv("HUD_KEY" = "q3r2rjimd129fj121jid")
 
-hud_cw(type = 1, query = '35213', year = c('2010', '2011'), quarter = c('1'))
+hud_cw(type = 1, query = '35213', year = c('2010', '2011'),
+   quarter = c('1'))
 
-hud_cw(type = '2', query = '35213', year = c('2016', '2020'), quarter = c('2'))
+hud_cw(type = '2', query = '35213', year = c('2016', '2020'),
+   quarter = c('2'))
 
-hud_cw(type = 3, query = 35213, year = c('2012', '2011'), quarter = c('3'))
+hud_cw(type = 3, query = 35213, year = c('2012', '2011'),
+   quarter = c('3'))
 
-hud_cw(type = 4, query = '22031', year = c('2017', '2019'), quarter = c('4'))
+hud_cw(type = 4, query = '22031', year = c('2017', '2019'),
+   quarter = c('4'))
 
-hud_cw(type = '5', query = '35213', year = c(2011, '2012'), quarter = c('1','2'))
+hud_cw(type = '5', query = '35213', year = c(2011, '2012'),
+   quarter = c('1','2'))
 
-hud_cw(type = 6, query = '48201223100', year = c('2017', '2010'), quarter = c('1','2','3'))
+hud_cw(type = 6, query = '48201223100', year = c('2017', '2010'),
+   quarter = c('1','2','3'))
 
-hud_cw(type = 7, query = '22031', year = c('2010', '2011'), quarter = c('1','2','3','4'))
+hud_cw(type = 7, query = '22031', year = c('2010', '2011'),
+   quarter = c('1','2','3','4'))
 
-hud_cw(type = 8, query = '10140', year = c('2010', '2011'), quarter = c('2','1'))
+hud_cw(type = 8, query = '10140', year = c('2010', '2011'),
+   quarter = c('2','1'))
 
-hud_cw(type = 9, query = '10380', year = c('2017'), quarter = c('4'))
+hud_cw(type = 9, query = '10380', year = c('2017'),
+   quarter = c('4'))
 
[…]
