+++ man/hud_chas_place.Rd
@@ -12,9 +12,11 @@ hud_chas_place(
 )
 }
 \arguments{
-\item{state}{The state name, abbreviation, or fips code. Make sure if state fips is 1 digit number, do not include leading 0.}
+\item{state}{The state name, abbreviation, or fips code. Make sure if state
+fips is 1 digit number, do not include leading 0.}
 
-\item{place}{The place(Usually just cities) to query for. Must supply as a 5 digit geoid.}
+\item{place}{The place(Usually just cities) to query for. Must supply as a 5
+digit geoid.}
 
 \item{year}{The years to query for.
 \itemize{
@@ -32,7 +34,8 @@ hud_chas_place(
 \item{key}{The key obtain from HUD USER website.}
 }
 \value{
-Returns a dataframe with CHAS data for a particular place (Usually just cities).
+Returns a dataframe with CHAS data for a particular place (Usually
+just cities).
 }
 \description{
 Returns CHAS data for place.
