+++ man/remove_leading_zeros.Rd
@@ -1,0 +1,30 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/hudmischelpers.R
+\name{remove_leading_zeros}
+\alias{remove_leading_zeros}
+\title{remove_leading_zeros}
+\usage{
+remove_leading_zeros(input)
+}
+\arguments{
+\item{input}{A vector or column of character vectors that need leading zeros
+removed.}
+}
+\value{
+A vector with leading zeros removed.
+}
+\description{
+Remove leading zeros from character vector.
+}
+\examples{
+\dontrun{
+library(hudr)
+zip <- c(02102, 11032, 01232)
+
+zip <- hud_remove_leading_zeros("zip", zip)
+
+cbsa <- c(02102, 11032, 01232)
+
+cbsa <- hud_remove_leading_zeros("cbsa", cbsa)
+}
+}
