### 2022-04-06 cec2e86
Paraphrasing quoted paragraphs in vignettes and rebuilding site. Added documentation for the omni functions.
