+++ R/hudwebsitedocs.R
@@ -5,6 +5,7 @@
 #' @param website The websites available.
 #'   1) "github-pages"
 #'   2) "github"
+#' @export
 hudr_website <- function(website = c("github-pages", "github")) {
   github_pages <- "https://etam4260.github.io/hudr/"
   github <- "https://github.com/etam4260/hudr"
