+++ README.Rmd
@@ -83,7 +83,7 @@ such data.
 Please cite this package using:
 
 Tam, E., Reilly, A., & Ghaedi, H. (2022). hudr: A R interface for accessing HUD
-(US Department of Housing and Urban Development) APIs (Version 0.1.0.9000).
+<br/>&nbsp;&nbsp;&nbsp;&nbsp; (US Department of Housing and Urban Development) APIs (Version 0.1.0.9000).
 https://github.com/etam4260/hudr
 
 ## Available Data
@@ -131,7 +131,9 @@ https://etam4260.github.io/hudr/. Select [Setup] in the navigation bar.
 
 ## Disclaimers
 
-* License: GPL >= 2 * To get citation information for hudr in R, type
+* License: GPL >= 2 
+
+* To get citation information for hudr in R, type
 citation(package = 'hudr') 
 
 * This interface uses the HUD User Data API but is
