### 2022-04-07 3c41288
fixing bullet point. Fixing bug with fmr/il county and metro queries not allowing multiple queries. Noticed that hud_fmr for small areas/metro, returns differently structured data depending on if it is a small area of not: not fixed yet so metro areas that are small areas in hud_fmr will not play nicely with those that are not small areas [ci skip]
