+++ R/hudcheckisnotlist.R
@@ -9,7 +9,7 @@ check_is_not_list <- function(dataframe) {
   if(typeof(dataframe) != "list") return(FALSE)
 
   for(i in seq_len(ncol(dataframe))) {
-    if(typeof(dataframe[,i]) != "character" && typeof(dataframe[,i]) != "integer" && typeof(dataframe[,i]) != "numeric" && typeof(dataframe[,i]) != "double") {
+    if(typeof(dataframe[[i]]) != "character" && typeof(dataframe[[i]]) != "integer" && typeof(dataframe[[i]]) != "numeric" && typeof(dataframe[[i]]) != "double") {
       return(FALSE)
     }
   }
