### 2022-04-20 5381d63
renaming some functions to be more intuitive and fixing some NOTEs in rmdcheck
