+++ README.Rmd
@@ -83,8 +83,7 @@ such data.
 Please cite this package using:
 
 Tam, E., Reilly, A., & Ghaedi, H. (2022). hudr: A R interface for accessing HUD
-<br/>&nbsp;&nbsp;&nbsp;&nbsp; (US Department of Housing and Urban Development) APIs (Version 0.1.0.9000).
-https://github.com/etam4260/hudr
+<br/>&nbsp;&nbsp; (US Department of Housing and Urban Development) APIs (Version 0.1.0.9000). https://github.com/etam4260/hudr
 
 ## Available Data
 
@@ -152,7 +151,6 @@ and introduce some pull requests!
 ## References
 
 Disaster Recovery Grant Reporting System. (2022). HUD Exchange. Retrieved
-<br />&nbsp;&nbsp;&nbsp;&nbsp; February 19, 2022, from https://www.hudexchange.info/programs/drgr/
+<br />&nbsp;&nbsp;February 19, 2022, from https://www.hudexchange.info/programs/drgr/
 
-HUD User Home Page: HUD USER. HUD User Home Page | HUD USER. (n.d.). Retrieved        
-<br />&nbsp;&nbsp;&nbsp;&nbsp; February 24, 2022, from https://www.huduser.gov/portal/home.html 
+HUD User Home Page: HUD USER. HUD User Home Page | HUD USER. (n.d.). Retrieved        <br />&nbsp;&nbsp;February 24, 2022, from https://www.huduser.gov/portal/home.html 
