### 2022-04-21 #524 @jhollist
Couple of thoughts.

As for the license stuff.  You will need to make sure that DESCRIPTION knows that you have an additional file.  If you add "+ file LICENSE" to "License: GPL (>= 2)" that should take care of that note.  I believe the two other files just need to be added to .Rbuildignore.    The root folder is a good spot for both (might event be required for codemeta.json).

And no need to apologize!  I understand the need to juggle multiple demands on time!
