### 2022-04-21 47e7a87
adding documentation to help users understand function syntax
