+++ vignettes/Setup.Rmd
@@ -106,31 +106,31 @@ these function calls should easily help you determine what you are querying
 and what data is returned.
 
 The general syntax for these functions within the package follow this pattern
-where *,^, and $ are placeholders:
+where {1},{2}, and {3} are placeholders:
 
 
-hud_*_^_$
+hud\_{1}\_{2}\_{3}
 
 
-*: This symbol represents the dataset to query for. So for example if you want
+{1}: This symbol represents the dataset to query for. So for example if you want
 fair markets rent data, this will be 'fmr'. If you want data from the crosswalk
 files, this will be 'cw'.
 
-^: This symbol represents the geographic resolution to query for. 
+{2}: This symbol represents the geographic resolution to query for. 
 For example, if the function requires that you input state(s), then this 
 will be named as 'state.' If the function requires that you input
 county(s) then this will be 'county'.
 
-$: This symbol represented the output geographic resolution. So for example,
+{3}: This symbol represented the output geographic resolution. So for example,
 if we want data at a zip code level, then this will have 'zip'.
 
-For those that do not have the ^ and $, these are omni functions which 
-are capable of performing all queries under *, but these functions are
+For those that do not have the {2} and {3}, these are omni functions which 
+are capable of performing all queries under {1}, but these functions are
 harder to understand.
 
-For those that do not have the $, it inherits the ^. This means that 
-if ^ is requires a state input, then the resolution of the output data is also
-as a state.
+For those that do not have the {3}, it inherits the {2}. This means that 
+if {2} is a state input, then the data described by the output data is also
+a state.
 
 A quick example:
 
@@ -148,7 +148,8 @@ The fourth part is the 'zip' meaning the data will be returned
 at a zip code resolution.
 
 If you ever get confused or need help you can easily revisit this website 
-through:
+through where you can check the all the function definitions in the
+REFERENCE tab.
 
 ```{r, eval=FALSE}
 rhud_website()
