### 2022-04-21 90305ec
fixing bug with non-small area fmr query
