### 2022-04-22 e4173a2
fixed bug in opening website docs fromR. Fixed most lintr warnings.
