+++ R/hudcheckisnotlist.R
@@ -6,10 +6,13 @@
 #' @noRd
 #' @noMd
 check_is_not_list <- function(dataframe) {
-  if(typeof(dataframe) != "list") return(FALSE)
+  if (typeof(dataframe) != "list") return(FALSE)
 
-  for(i in seq_len(ncol(dataframe))) {
-    if(typeof(dataframe[,i]) != "character" && typeof(dataframe[,i]) != "integer" && typeof(dataframe[,i]) != "numeric" && typeof(dataframe[,i]) != "double") {
+  for (i in seq_len(ncol(dataframe))) {
+    if (typeof(dataframe[, i]) != "character" &&
+       typeof(dataframe[, i]) != "integer" &&
+       typeof(dataframe[, i]) != "numeric" &&
+       typeof(dataframe[, i]) != "double") {
       return(FALSE)
     }
   }
