+++ tests/testthat/test-test_cw.R
@@ -7,88 +7,83 @@ test_that("Crosswalk All Types", {
   # to add least have 1 row or more in data frame returned.
 
   # Lets try a ZIP code in Alabama for now for 1 -> 5 as well as 11.
-  zip_tract <- hud_cw(type = 1, query = '35213',
-                      year = c('2010', '2011'), quarter = c('1'))
+  zip_tract <- hud_cw(type = 1, query = "35213",
+                      year = c("2010", "2011"), quarter = c("1"))
 
   expect_true(nrow(zip_tract) >= 1)
 
-  zip_county <- hud_cw(type = '2', query = '35213',
-                       year = c('2016', '2020'), quarter = c('2'))
+  zip_county <- hud_cw(type = "2", query = "35213",
+                       year = c("2016", "2020"), quarter = c("2"))
 
   expect_true(nrow(zip_county) >= 1)
 
   zip_cbsa <- hud_cw(type = 3, query = 35213,
-                     year = c('2012', '2011'), quarter = c('3'))
+                     year = c("2012", "2011"), quarter = c("3"))
 
   expect_true(nrow(zip_cbsa) >= 1)
 
-  zip_cbsadiv <- hud_cw(type = 4, query = '22031',
-                        year = c('2017', '2019'), quarter = c('4'))
+  zip_cbsadiv <- hud_cw(type = 4, query = "22031",
+                        year = c("2017", "2019"), quarter = c("4"))
 
   expect_true(nrow(zip_cbsadiv) >= 1)
 
-  zip_cd <- hud_cw(type = '5', query = '35213',
-                   year = c(2011, '2012'), quarter = c('1','2'))
+  zip_cd <- hud_cw(type = "5", query = "35213",
+                   year = c(2011, "2012"), quarter = c("1", "2"))
 
   expect_true(nrow(zip_cd) >= 1)
 
-  # A census tract code consists of STATE+COUNTY+TRACT. So in this
-  # STATE = 48
-  # COUNTY = 201
-  # TRACT = 223100
-  tract_zip <- hud_cw(type = 6, query = '48201223100',
-                      year = c('2017', '2010'),
-                      quarter = c('1','2','3'))
+  tract_zip <- hud_cw(type = 6, query = "48201223100",
+                      year = c("2017", "2010"),
+                      quarter = c("1", "2", "3"))
 
   expect_true(nrow(tract_zip) >= 1)
 
   # Testing county to zip cross_walk. Assuming this to be the most popular.
-  county_zip <- hud_cw(type = 7, query = '22031',
-                       year = c('2010', '2011'), quarter = c('1','2','3','4'))
+  county_zip <- hud_cw(type = 7, query = "22031",
+                       year = c("2010", "2011"),
+                       quarter = c("1", "2", "3", "4"))
 
   expect_true(nrow(county_zip) >= 1)
 
   # Try all counties in MD
-  county_zip <- hud_cw(type = 7, query = substr(hud_state_counties("md")$fips_code,1,5),
-                       year = c('2010'), quarter = c('1'))
+  county_zip <- hud_cw(type = 7,
+                       query = substr(hud_state_counties("md")$fips_code, 1, 5),
+                       year = c("2010"), quarter = c("1"))
   expect_true(nrow(county_zip) >= 1)
 
   # A core based Statistical Area to zip crosswalk.
   # CBSA defines Micropolitan and Metropolitan
-  cbsa_zip <- hud_cw(type = 8, query = '10140',
-                     year = c('2010', '2011'), quarter = c('2','1'))
+  cbsa_zip <- hud_cw(type = 8, query = "10140",
+                     year = c("2010", "2011"), quarter = c("2", "1"))
 
[…]
