+++ tests/testthat/setup.R
@@ -6,7 +6,7 @@ library(httptest)
 with_mock_dir("good_key", {
   test_that("HUD Good", {
 
-    #va <- hud_fmr("VA", year = c(2021))
-    #expect_true(nrow(va) > 1)
+    # va <- hud_fmr("VA", year = c(2021))
+    # expect_true(nrow(va) > 1)
   })
 })
