### 2022-04-27 3094ed7
added a loading bar. chas data now returns NA for missing values instead of 0.
