+++ man/hud_nation_states_territories.Rd
@@ -29,3 +29,4 @@ hud_nation_states_territories()
 }
 }
 \keyword{States}
+\keyword{Territories}

[…]
