### 2022-04-27 7c308aa
adding files to build ignore, adding description for return values in hud_fmr functions.
