+++ DESCRIPTION
@@ -22,10 +22,10 @@ Suggests:
     httptest,
     knitr,
     rmarkdown,
-    testthat (>= 3.0.0)
+    testthat (>= 3.0.0),
+    devtools
 Imports: 
     httr,
-    devtools,
     zoo,
     rio
 Config/testthat/edition: 3
