### 2022-04-29 0b68732
fixing formatting in download bar. fixed formatting in error and warning messages to be more user friendly.
