### 2022-05-03 #524 @jhollist
@etam4260 Two things came up in the most recent checks

1.) A couple of functions are missing examples.  Add these examples into the docs
2.) The test coverage.  I see that most of your tests skip without the key hence the low number above.  Looks like your current suite is at about 72%.  While we shoot for 75%, this is close enough to get the review process started.  I would just ask that after and/or during the review that you revisit the tests and make sure you have adequate coverage.

I think we are close enough.  I will start looking for potential reviewers!
