### 2022-05-04 2ba1624
[no ci] adding documentation for crosswalk and rhud_website functions. Forming vignette for the crosswalk function andadding seealso to families of function. Fixed some bugs in crosswalk functions.
