+++ R/hudwebsitedocs.R
@@ -1,3 +1,5 @@
+#' @importFrom curl has_internet
+
 #' @name rhud_website
 #' @title rhud_website
 #' @description Quickly get documentation for the hudr package by opening up
@@ -14,6 +16,9 @@
 #' rhud_website()
 #' }
 rhud_website <- function(website = c("github-pages", "github")) {
+  if(!curl::has_internet()) stop("You currently do not have internet access.")
+  # Check for internet
+
   github_pages <- "https://etam4260.github.io/hudr/"
   github <- "https://github.com/etam4260/hudr"
 
