+++ README.Rmd
@@ -65,8 +65,8 @@ R users may want to be able to extract specific bits of the data into memory.
 
 Please cite this package using:
 
-Tam, E., Reilly, A., & Ghaedi, H. (2022). rhud: A R interface for accessing HUD
-<br/>&nbsp;&nbsp; (US Department of Housing and Urban Development) APIs (Version 0.2.0.9000). https://github.com/etam4260/rhud
+| Tam, E., Reilly, A., & Ghaedi, H. (2022). rhud: A R interface for accessing HUD
+|       (US Department of Housing and Urban Development) APIs (Version 0.2.0.9000). https://github.com/etam4260/rhud
 
 ## Available Data
 
@@ -128,5 +128,5 @@ and introduce some pull requests!
 
 ## References
 
-HUD User Home Page: HUD USER. HUD User Home Page | HUD USER. (n.d.). Retrieved 
-<br />&nbsp;&nbsp;February 24, 2022, from https://www.huduser.gov/portal/home.html 
+| HUD User Home Page: HUD USER. HUD User Home Page | HUD USER. (n.d.). Retrieved 
+|         February 24, 2022, from https://www.huduser.gov/portal/home.html 
