### 2022-05-04 5193e8e
adding curl to dependencies
