+++ man/hud_cw.Rd
@@ -76,7 +76,7 @@ US Department of Housing and Urban Development.
 }
 \examples{
 \dontrun{
-library(hudr)
+library(rhud)
 
 Sys.setenv("HUD_KEY" = "q3r2rjimd129fj121jid")
 

[…]
