### 2022-05-17 #524 @jhollist
@etam4260 Just a quick update.  I have been looking for reviewers.  I have one so far (Thank you, @rtaph!) and have another invite out that I am waiting on.

Have you had a chance to add those examples to the few functions that were missing them?
