### 2022-05-17 8a59f2d
added ROpensci review badge. Website docs now refer to converting to different geographies using the word crosswalk instead of convert. Added sample case study vignette for FMR and IL
