+++ README.md
@@ -13,6 +13,9 @@ maturing](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www
 [![R-CMD-check](https://github.com/etam4260/rhud/workflows/R-CMD-check/badge.svg)](https://github.com/etam4260/rhud/actions)
 [![Codecov test
 coverage](https://codecov.io/gh/etam4260/rhud/branch/main/graph/badge.svg)](https://codecov.io/gh/etam4260/rhud?branch=main)
+[![CodeFactor](https://www.codefactor.io/repository/github/etam4260/rhud/badge/main)](https://www.codefactor.io/repository/github/etam4260/rhud/overview/main)
+[![Status at rOpenSci Software Peer
+Review](https://badges.ropensci.org/524_status.svg)](https://github.com/ropensci/software-review/issues/524)
 <br/> <br/> [![devel
 version](https://img.shields.io/badge/devel%20version-0.2.0.9000-yellow)]()
 
