+++ vignettes/Comparing-FMR-and-IL.Rmd
@@ -1,0 +1,145 @@
+---
+title: "Comparing-FMR-and-IL"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Comparing-FMR-and-IL}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+<i>Sample provided by <b>Shuyu Jin</b> (geojsy@umd.edu)</i>
+
+## Loading in the necessary libraries
+
+These are some the packages needed to demonstrate this example.
+They might take a while to download if you don't have them already.
+
+```{r message=FALSE}
+library(rhud)
+library(tigris)
+library(leaflet)
+library(tidyverse)
+library(sf)
+```
+
+To improve performance, use tigris caching.
+
+```{r message=FALSE}
+options(tigris_use_cache = TRUE)
+```
+
+## Reading in data
+
+Read in the Fair Markets Rent for counties in the year 2021 as well as the shape
+files needed to map.
+
+```{r message=FALSE}
+md_counties_2021_fmr = hud_fmr_state_counties(state = 'MD', year = '2021')
+```
+
+Get county data for the state of Maryland.
+
+```{r message=FALSE}
+md_counties = hud_state_counties("MD")
+```
+
+Query for every county in Maryland and get their Income Limits data.
+
+```{r message=FALSE}
+md_counties_2021_il = hud_il(md_counties$fips_code)
+```
+
+Download the shape files for counties in the US.
+
+```{r message=FALSE}
+md_county_shapefile = tigris::counties(state = NULL, cb = FALSE,
+                                       resolution = "500k", year = 2021)
+```
+
+Filter the counties to those only in Maryland.
+
+```{r message=FALSE}
+md_county_shapefile = md_county_shapefile[md_county_shapefile$STATEFP == 24, ]
+
+md_county_shapefile$fullFips = paste0(as.character(md_county_shapefile$STATEFP),
+                                 as.character(md_county_shapefile$COUNTYFP),
+                                 "99999")
+```
+
+## Merging of shapefiles and data
+
+Merge the shape file with the FMR and IL data.
+
+```{r message=FALSE, warning=FALSE}
+md_county_fmr_merged = tigris::geo_join(md_county_shapefile,
+                                        md_counties_2021_fmr,
+                                        by_sp = "fullFips",
+                                        by_df = "fips_code",
+                                        how = "inner")
+```
+
+
+## Forming components of Leaflet map
+
+Create the pallete for coloring the overlayed shape file. The scale used
+corresponds to Efficiency data.
+
+```{r message=FALSE}
+pal <- colorNumeric(
+  palette = "Greens",
+  domain = as.numeric(md_county_fmr_merged$Efficiency))
+```
+
+Create the county level, efficiency, one bedroom, two bedroom, three bedroom
+four bedroom label to appear when hovering using the FMR overlay.
+
+```{r message=FALSE}
+label <- paste0(
+  "County: ", md_county_fmr_merged$NAME, "<br/>",
+  "Efficiency: ", scales::dollar(as.numeric(md_county_fmr_merged$Efficiency)), "<br/>",
+  "One Bedroom: ", scales::dollar(as.numeric(md_county_fmr_merged$`One-Bedroom`)), "<br/>",
+  "Two Bedroom: ", scales::dollar(as.numeric(md_county_fmr_merged$`Two-Bedroom`)), "<br/>",
[…]

[…]
