### 2022-05-17 ac0a7bb
fixing type in hudmisc, build should ignore Rproj, not evaluating sample FMR-IL vignette as dependencies are not available. Fixing vignette index.
