### 2022-05-17 #524 @etam4260
@jhollist Yes, I was able to add the documentation to those functions. As for the testing suite, it is at 70% now. Still working to get over that 75% mark. I have also added the under review badge to the README.

@rtaph Thanks for taking the time to review the package
