### 2022-05-17 #524 @jhollist
Sounds good!  Since the code might change some as a result of the review, it might make sense to wait until after review to worry too much about the code coverage, unless there are a few functions that would be easy to add to the tests at this point.

Still working on that second reviewer!
