### 2022-05-25 50e6688
updating description file for dependencies in upcoming vignetters. Fixed data returned from IL with two copies of very low instead of low
