+++ README.Rmd
@@ -28,6 +28,9 @@ knitr::opts_chunk$set(
 
 <!-- badges: end -->
 
+Are you a python developer? Check out [hudpy][https://github.com/etam4260/hudpy]
+instead.
+
 ## Housing and Urban Development in R
 
 * This interface uses the HUD User Data API but is not endorsed
@@ -68,7 +71,8 @@ R users may want to be able to extract specific bits of the data into memory.
 Please cite this package using:
 
 | Tam, E., Reilly, A., & Ghaedi, H. (2022). rhud: A R interface for accessing HUD
-|       (US Department of Housing and Urban Development) APIs (Version 0.2.0.9000). https://github.com/etam4260/rhud
+|       (US Department of Housing and Urban Development) APIs (Version 0.2.0.9000).                     
+|        https://github.com/etam4260/rhud
 
 ## Available Data
 
