+++ tests/testthat/test-test_decomp_cw.R
@@ -107,10 +107,6 @@ test_that("Query hud_cw_county_zip()", {
                            quarter = c("1"), minimal = TRUE)
   expect_true(is.vector(res) && length(res) >= 1)
 
-
-  # It looks like HUD only provides support for querying all counties in a state
-  # but, not any other smaller geographies.
-
   all_md_counties <- hud_state_counties("md")
   all_md <- hud_cw_county_zip(county = substr(all_md_counties$fips_code, 1, 5),
                               year = c("2010"), quarter = c("1"))
