+++ tests/testthat/test-test_cw_a_dataset.R
@@ -6,23 +6,6 @@ test_that("test simple crosswalk for zip", {
 
   # Crosswalk zip to every other geoid...
 
-  ### Crosswalk zip to county in quarter 1, 2018
-  expect_error(crosswalk(sample, "zip", "zip", "county", year = 2018,
-                         quarter = 1), NA)
-  expect_error(crosswalk(sample, "zip", "zip", "county", "mes", "res",
-                         2018, 1), NA)
-
-  # Use business
-  expect_error(crosswalk(sample, "zip", "zip", "county", "mes", "bus",
-                         2018, 1), NA)
-  # Use total
-  expect_error(crosswalk(sample, "zip", "zip", "county", "mes", "tot",
-                         2018, 1), NA)
-  # Use other
-  expect_error(crosswalk(sample, "zip", "zip", "county", "mes", "oth",
-                         2018, 1), NA)
-
-
   ### Crosswalk zip to tract in quarter 1, 2018
   expect_error(crosswalk(sample, "zip", "zip", "tract", year = 2018,
                          quarter = 1), NA)
@@ -89,11 +72,30 @@ test_that("test simple crosswalk for zip", {
   # Use other
   expect_error(crosswalk(sample, "zip", "zip", "cd", "mes", "oth",
                          2018, 1), NA)
+
+
+
+  ### Crosswalk zip to cd in quarter 1, 2018
+  expect_error(crosswalk(sample, "zip", "zip", "cd",
+                         year = 2018, quarter = 1), NA)
+  expect_error(crosswalk(sample, "zip", "zip", "cd", "mes", "res",
+                         2018, 1), NA)
+
+  # Use business
+  expect_error(crosswalk(sample, "zip", "zip", "cd", "mes", "bus",
+                         2018, 1), NA)
+  # Use total
+  expect_error(crosswalk(sample, "zip", "zip", "cd", "mes", "tot",
+                         2018, 1), NA)
+  # Use other
+  expect_error(crosswalk(sample, "zip", "zip", "cd", "mes", "oth",
+                         2018, 1), NA)
 })
 
 
 test_that("test awkward crosswalk",{
   sample <- data.frame(mes = c(1232, 1232, 1232), zip = c(21206, 21206, 21206))
+
   expect_error(crosswalk(sample, "zip", "zip", "county", year = 2018,
                          quarter = 1), NA)
   expect_error(crosswalk(sample, "zip", "zip", "county", "mes", "res",
