+++ R/hudfmr.R
@@ -3,7 +3,8 @@
 #' @description This function queries for a state and returns the
 #'   FMR calculation
 #'   at a metroarea resolution for all metroareas in this state.
-#' @param state The state to query for.
+#' @param state The state to query for. Can be abbreviation, fip code, or
+#'   full name.
 #' @param year Gets the year that this data was recorded.
 #'   Can specify multiple years. Default is the
 #'   previous year.
@@ -17,6 +18,17 @@
 #' * [rhud::hud_fmr_county_zip()]
 #' @export
 #' @returns A data frame with fair markets rent for metro areas in states.
+#' \dontrun{
+#' library(rhud)
+#'
+#' Sys.setenv("HUD_KEY" = "q3r2rjimd129fj121jid")
+#'
+#' hud_fmr_state_metroareas("VA", year = c(2021))
+#'
+#' hud_fmr_state_metroareas("Alabama", year = c(2021))
+#'
+#' hud_fmr_state_metroareas("24", year = c(2021)))
+#' }
 hud_fmr_state_metroareas <- function(state,
                                      year = format(Sys.Date() - 365, "%Y"),
                                      key = Sys.getenv("HUD_KEY")) {
@@ -90,7 +102,8 @@ hud_fmr_state_metroareas <- function(state,
 #' @description This function queries for a state and returns the
 #'   FMR calculation
 #'   at a county resolution for all counties in this state.
-#' @param state The state to query for.
+#' @param state The state to query for. Can be abbreviation, fip code, or
+#'   full name.
 #' @param year Gets the year that this data was recorded.
 #'   Can specify multiple years. Default is the
 #'   previous year.
@@ -104,6 +117,17 @@ hud_fmr_state_metroareas <- function(state,
 #' * [rhud::hud_fmr_county_zip()]
 #' @export
 #' @returns A data frame with fair markets rent for counties in states.
+#' \dontrun{
+#' library(rhud)
+#'
+#' Sys.setenv("HUD_KEY" = "q3r2rjimd129fj121jid")
+#'
+#' hud_fmr_state_counties("VA", year = c(2021))
+#'
+#' hud_fmr_state_counties("Alabama", year = c(2021))
+#'
+#' hud_fmr_state_counties("24", year = c(2021)))
+#' }
 hud_fmr_state_counties <- function(state, year = format(Sys.Date() - 365, "%Y"),
                                    key = Sys.getenv("HUD_KEY")) {
 
@@ -179,7 +203,7 @@ hud_fmr_state_counties <- function(state, year = format(Sys.Date() - 365, "%Y"),
 #'    a small area, it will return only single
 #'    measurement for that county. If the county is considered a small area,
 #'    it will return data at a zip code level.
-#' @param county A county to query for.
+#' @param county A county to query for. Must be provided as a 5 digit fipcode.
 #' @param year Gets the year that this data was recorded.
 #'   Can specify multiple years. Default is the
 #'   previous year.
@@ -193,6 +217,17 @@ hud_fmr_state_counties <- function(state, year = format(Sys.Date() - 365, "%Y"),
 #' * [rhud::hud_fmr_county_zip()]
 #' @export
 #' @returns A data frame with fair markets rent for zip codes in counties.
+#' \dontrun{
+#' library(rhud)
+#'
+#' Sys.setenv("HUD_KEY" = "q3r2rjimd129fj121jid")
+#'
+#' hud_fmr_county_zip("5100199999", year = c(2021))
+#'
+#' hud_fmr_county_zip("5100199999", year = c("2021"))
+#'
+#' hud_fmr_county_zip(5151099999, year = c(2021))
+#' }
[…]
