+++ R/hudcw.R
@@ -1,4 +1,3 @@
-
 #' @name hud_cw_zip_tract
 #' @title hud_cw_zip_tract
 #' @description This function queries the Crosswalks API provided by US
@@ -14,6 +13,8 @@
 #'   all fields. This does not remove duplicates.
 #' @param key The API key for this user. You must go to HUD and sign up for an
 #'   account and request for an API key.
+#' @param to_tibble If TRUE, return the data in a tibble format
+#'   rather than a data frame.
 #' @keywords Crosswalks API
 #' @seealso
 #' * [rhud::crosswalk()]
@@ -46,9 +47,20 @@
 #' }
 hud_cw_zip_tract <- function(zip, year = format(Sys.Date() - 365, "%Y"),
                              quarter = 1, minimal = FALSE,
-                             key = Sys.getenv("HUD_KEY")) {
+                             key = Sys.getenv("HUD_KEY"),
+                             to_tibble) {
+
   if (!curl::has_internet()) stop("\nYou currently do not have internet access.",
                                   call. = FALSE)
+
+  if (!is.null(getOption("rhud_use_tibble")) && missing(to_tibble)) {
+    to_tibble = getOption("rhud_use_tibble")
+    message(paste("Outputted in tibble format",
+                  "because it was set using `options(rhud_use_tibble = TRUE)`\n"))
+  } else if (missing(to_tibble)) {
+    to_tibble = FALSE
+  }
+
   primary_geoid <- "zip"
   secondary_geoid <- "tract"
 
@@ -72,11 +84,11 @@ hud_cw_zip_tract <- function(zip, year = format(Sys.Date() - 365, "%Y"),
   if (!minimal) {
     return(cw_do_query_calls(urls, all_queries$query, all_queries$year,
                              all_queries$quarter, primary_geoid,
-                             secondary_geoid, key))
+                             secondary_geoid, key, to_tibble))
   }
   return(cw_do_query_calls(urls, all_queries$query, all_queries$year,
                            all_queries$quarter, primary_geoid,
-                           secondary_geoid, key)$tract)
+                           secondary_geoid, key, to_tibble)$tract)
 }
 
 
@@ -97,6 +109,8 @@ hud_cw_zip_tract <- function(zip, year = format(Sys.Date() - 365, "%Y"),
 #'   all fields. This does not remove duplicates.
 #' @param key The API key for this user. You must go to HUD and sign up for an
 #'   account and request for an API key.
+#' @param to_tibble If TRUE, return the data in a tibble format
+#'   rather than a data frame.
 #' @keywords Crosswalks API
 #' @seealso
 #' * [rhud::hud_cw_zip_tract()]
@@ -128,9 +142,21 @@ hud_cw_zip_tract <- function(zip, year = format(Sys.Date() - 365, "%Y"),
 #' }
 hud_cw_zip_county <- function(zip, year = format(Sys.Date() - 365, "%Y"),
                               quarter = 1, minimal = FALSE,
-                              key = Sys.getenv("HUD_KEY")) {
+                              key = Sys.getenv("HUD_KEY"),
+                              to_tibble) {
+
   if (!curl::has_internet()) stop("\nYou currently do not have internet access.",
                                   call. = FALSE)
+
+
+  if (!is.null(getOption("rhud_use_tibble")) && missing(to_tibble)) {
+    to_tibble = getOption("rhud_use_tibble")
+    message(paste("Outputted in tibble format",
+                  "because it was set using `options(rhud_use_tibble = TRUE)`\n"))
+  } else if (missing(to_tibble)) {
+    to_tibble = FALSE
+  }
+
   primary_geoid <- "zip"
   secondary_geoid <- "county"
 
[…]
