+++ R/huduser.R
@@ -1,5 +1,6 @@
 #' @import httr
-#'
+#' @import tibble
+
 # An R interface for accessing HUD USER
 # (US Department of Housing and Urban Development) APIs.
 
@@ -59,6 +60,8 @@
 #'   all fields. This does not remove duplicates.
 #' @param key The API key for this user. You must go to HUD and sign up for
 #'   an account and request for an API key.
+#' @param to_tibble If TRUE, return the data in a tibble format
+#'   rather than a data frame.
 #' @keywords Crosswalks API
 #' @export
 #' @returns This function returns a dataframe containing CROSSWALK data for
@@ -123,10 +126,19 @@
 #'    quarter = c("4", "4"))
 #' }
 hud_cw <- function(type, query, year = format(Sys.Date() - 365, "%Y"),
-                   quarter = 1, minimal = FALSE, key = Sys.getenv("HUD_KEY")) {
+                   quarter = 1, minimal = FALSE, key = Sys.getenv("HUD_KEY"),
+                   to_tibble) {
 
   if (!curl::has_internet()) stop("\nYou currently do not have internet access.",
                                   call. = FALSE)
+
+  if (!is.null(getOption("rhud_use_tibble")) && missing(to_tibble)) {
+    to_tibble = getOption("rhud_use_tibble")
+    message(paste("Outputted in tibble format",
+                  "because it was set using `options(rhud_use_tibble = TRUE)`\n"))
+  } else if (missing(to_tibble)) {
+    to_tibble = FALSE
+  }
 
   args <- cw_input_check_cleansing(query, year, quarter, key)
 
@@ -222,9 +234,9 @@ hud_cw <- function(type, query, year = format(Sys.Date() - 365, "%Y"),
 
   if (!minimal) return(cw_do_query_calls(urls, all_queries$query,
                                          all_queries$year,
-                           all_queries$quarter, lhgeoid, rhgeoid, key))
+                           all_queries$quarter, lhgeoid, rhgeoid, key, to_tibble))
   return(cw_do_query_calls(urls, all_queries$query, all_queries$year,
-                           all_queries$quarter, lhgeoid, rhgeoid, key)[[1]])
+                           all_queries$quarter, lhgeoid, rhgeoid, key, to_tibble)[[1]])
 }
 
 
@@ -243,6 +255,8 @@ hud_cw <- function(type, query, year = format(Sys.Date() - 365, "%Y"),
 #'   previous year.
 #' @param key The API key for this user. You must go to HUD and sign up
 #'   for an account and request for an API key.
+#' @param to_tibble If TRUE, return the data in a tibble format
+#'   rather than a data frame.
 #' @keywords Fair Markets Rent API
 #' @export
 #' @returns This function returns a dataframe containing fair markets rent data
@@ -267,10 +281,18 @@ hud_cw <- function(type, query, year = format(Sys.Date() - 365, "%Y"),
 #' hud_fmr("METRO47900M47900", year=c(2018))
 #' }
 hud_fmr <- function(query, year = format(Sys.Date() - 365, "%Y"),
-                    key = Sys.getenv("HUD_KEY")) {
+                    key = Sys.getenv("HUD_KEY"), to_tibble) {
 
   if (!curl::has_internet()) stop("\nYou currently do not have internet access.",
                                   call. = FALSE)
+
+  if (!is.null(getOption("rhud_use_tibble")) && missing(to_tibble)) {
+    to_tibble = getOption("rhud_use_tibble")
+    message(paste("Outputted in tibble format",
+                  "because it was set using `options(rhud_use_tibble = TRUE)`\n"))
+  } else if (missing(to_tibble)) {
+    to_tibble = FALSE
+  }
 
   args <- fmr_il_input_check_cleansing(query, year, key)
   query <- args[[1]]
[…]
