+++ R/zzz.R
@@ -12,7 +12,8 @@
                               "-------------------------------------",
                               "\n* To begin using this package, you must obtain a key ",
                               "from HUD User website.",
-                              "\n* To enable caching of rhud data, ",
+                              "\n* (Soon to be implemented)",
+                              " To enable caching of rhud data, ",
                               "use `options(rhud_use_cache = TRUE)`\n",
                               "* To get tibbles instead of dataframes, ",
                               "use `options(rhud_use_tibble = TRUE)`\n",
