### 2022-05-29 412c950
readme now has list of available functions shown as well as year ranges that can be queried for.
