### 2022-05-29 6f732df
bolding some dates for hud_cw table in read_me to make it more noticable
