+++ README.Rmd
@@ -88,15 +88,15 @@ also provide miscellaneous supplemental APIs under them.
   | `hud_cw_zip_tract()` | 2010-2021 | 
   | `hud_cw_zip_county()` | 2010-2021 |
   | `hud_cw_zip_cbsa()` | 2010-2021 |
-  | `hud_cw_zip_cbsadiv()` | 2017-2021 |
-  | `hud_cw_zip_countysub()` | 2018-2021 |
+  | `hud_cw_zip_cbsadiv()` | <b>2017-2021<b> |
+  | `hud_cw_zip_countysub()` | <b>2018-2021<b> |
   | `hud_cw_zip_cd()` | 2010-2021 |
   | `hud_cw_tract_zip()` | 2010-2021 |
   | `hud_cw_county_zip()` | 2010-2021 |
   | `hud_cw_cbsa_zip()` | 2010-2021 |
-  | `hud_cw_cbsadiv_zip()` | 2017-2021 |
+  | `hud_cw_cbsadiv_zip()` | <b>2017-2021<b> |
   | `hud_cw_cd_zip()` | 2010-2021 |
-  | `hud_cw_countysub_zip()` | 2018-2021 |
+  | `hud_cw_countysub_zip()` | <b>2018-2021<b> |
   | `hud_cw()` | 2010-2021 |
   | `crosswalk()` | 2010-2021 | 
 
