+++ README.Rmd
@@ -150,7 +150,7 @@ also provide miscellaneous supplemental APIs under them.
   
   | Caching |
   |------------------------------------------|
-  | `rhud_use_cache()` |
+  | `rhud_cache_dir()` |
   | `rhud_clear_cache()` | 
 
   - Documentation
