### 2022-05-30 d83db7c
updating news for 0.4.0 update
