### 2022-05-30 d1cf15f
Adding on load now initializes the cache directory. Now allow user to specify cache directory using rhud_cache_dir. Allow user to clear the cache. Adding some todo items for the loading bar. All API calls are now memoized.
