+++ R/hudcache.R
@@ -1,30 +1,158 @@
 # File adapted from the tigris package:
 # https://github.com/walkerke/tigris/blob/master/R/helpers.R
 
+#' @import R.cache
+
 #' @name rhud_cache_dir
 #' @title rhud_cache_dir
 #' @description Set the caching directory to store data retrieved using the
-#' rhud API calls. By default, rhud uses a non-persistent temporary directory given
-#' for an R session. However, it is possible that a user might want
-#' to save their queried data in a custom location. This function allows users
-#' to set the cache directory. The user can also store the result in the
-#' .RProfile so that rhud will remember the set preference directory throughout
-#' sessions.
+#' rhud API calls. By default, rhud uses a non-persistent temporary directory
+#' given for an R session. However, it is possible that a user might want
+#' to save their queried data in a custom location which is persistent.
+#' This function allows users to set the cache directory. The user can also
+#' store the preferred path in the .RProfile so that rhud will remember the set
+#' preference directory throughout sessions. Make sure the path is valid and
+#' and is not in a sensitive location.
 #'
 #' Windows users: please note that you'll need to use double-backslashes or
 #' forward slashes when specifying your cache directory's path in R.
 #'
-#' @param path The full path to the desired cache directory
-#' @param in_wkdir Set the .Rprofile in current working directly to have
-#'   have rhud cache to path.
-#' @param in_home Set the .Rprofile in the HOME directory to have
-#'   have rhud cache to path.
+#' @param path The full path to the desired cache directory. Only one can be
+#'  set at a time. If no path is specified, it will use the temp directory
+#'  for R. The temp directory is not persistent.
+#' @in_wkdir Store the path as environment variable in the working directory
+#'   .Rprofile so rhud will cache to this specified path.
+#' @in_home Store the path as environment variable in the HOME directory
+#'   .Rprofile so rhud will cache to this specified path.
 #' @export
 #' @examples
 #' \dontrun{
 #' rhud_cache_dir("./an/example/path", in_wkdir = TRUE, in_home = TRUE)
 #'
 #' }
-rhud_cache_dir <- function(path, in_wkdir = FALSE, in_home = FALSE) {
+rhud_cache_dir <- function(path = "NULL",
+                           in_wkdir = FALSE,
+                           in_home = FALSE
+                           ) {
 
+  # Set for the rhud cache path in the current session.
+
+  # TODO: First check for valid path formatting by regex
+
+
+  if (!is.character(path) || length(path) != 1) {
+    stop(paste("Make sure argument key is of type",
+               "character and is of vector length 1",
+               sep = ""))
+  }
+
+
+  if(!is.logical(in_wkdir) ||
+     !is.logical(in_home) ||
+     length(in_wkdir) != 1  ||
+     length(in_home) != 1) {
+    stop("Make sure argument in_wkdir and in_home are of type logical.")
+  }
+
+  Sys.setenv("RHUD_CACHE_DIR" = path)
+
+  if (in_wkdir) {
+
+    # Set the key in the Rprofile working direct. If not made, make one and set.
+    if (any(list.files(all.files = TRUE) == ".Rprofile")) {
+      # Check the file if it contains a call to set hud key, regex for it.
+      rprof <- readLines("./.Rprofile")
+
+      all_occur <- grep("^Sys\\.setenv\\(\"RHUD_CACHE_DIR\" = \".*\"\\)",
+                        rprof)
+
+      if (any(all_occur))  {
+
[…]

[…]
