### 2022-05-30 5c6ef56
Correcting readme for 0.4.0 update. Making sure to directly search for R.cache functions. Loading bar has now errored out number and current url on it. Updating hud_il to be memoized. Adding tests for keys and cache.
