+++ tests/testthat/test-test_key.R
@@ -1,0 +1,3 @@
+test_that("test", {
+  expect_equal(2 * 2, 4)
+})
