### 2022-05-30 1c6a717
removing rhud-package file and moved info to zzz.R. Making md file is named rhud-package so description shows up when user types in console
