+++ man/rhud-exports.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/rhud-package.R
+% Please edit documentation in R/zzz.R
 \name{rhud-exports}
 \alias{rhud-exports}
 \title{rhud exported operators}
