### 2022-05-30 f8d6475
updated news to include description of pkg when typing in console
