### 2022-05-30 b6a7700
attempt 2 at trying to remove download bar from appearing when using hud_nation_states_territories from first validating inputs.
