+++ R/hudchas.R
@@ -114,7 +114,7 @@ hud_chas_state <- function(state, year = c("2014-2018"),
   if (all(nchar(state) > 2)) state <- capitalize(tolower(state))
 
   if (is.null(pkg.env$state)) {
-    pkg.env$state <- hud_nation_states_territories(key = Sys.getenv("HUD_KEY"))
+    pkg.env$state <- supressMessages(hud_nation_states_territories(key = Sys.getenv("HUD_KEY")))
   }
 
   for (i in seq_len(length(state))) {
@@ -217,7 +217,7 @@ hud_chas_county <- function(county, year = c("2014-2018"),
   # Grab all possible states if the package doesn't have a state dataset to
   # query from.
   if (is.null(pkg.env$state)) {
-    pkg.env$state <- hud_nation_states_territories(key = Sys.getenv("HUD_KEY"))
+    pkg.env$state <- supressMessages(hud_nation_states_territories(key = Sys.getenv("HUD_KEY")))
   }
 
   # Check if first two numbers of county code inputted is a valid state.
@@ -306,7 +306,7 @@ hud_chas_state_mcd <- function(state, year = c("2014-2018"),
   if (all(nchar(state) > 2)) state <- capitalize(tolower(state))
 
   if (is.null(pkg.env$state)) {
-    pkg.env$state <- hud_nation_states_territories(key = Sys.getenv("HUD_KEY"))
+    pkg.env$state <- supressMessages(hud_nation_states_territories(key = Sys.getenv("HUD_KEY")))
   }
 
   for (i in seq_len(length(state))) {
@@ -408,15 +408,11 @@ hud_chas_state_place <- function(state, year = c("2014-2018"),
   key <- args[[2]]
 
   if (is.null(pkg.env$state)) {
-    pkg.env$state <- hud_nation_states_territories(key = Sys.getenv("HUD_KEY"))
+    pkg.env$state <- supressMessages(hud_nation_states_territories(key = Sys.getenv("HUD_KEY")))
   }
 
   if (all(nchar(state) == 2)) state <- toupper(state)
   if (all(nchar(state) > 2)) state <- capitalize(tolower(state))
-
-  if (is.null(pkg.env$state)) {
-    pkg.env$state <- hud_nation_states_territories(key = Sys.getenv("HUD_KEY"))
-  }
 
   for (i in seq_len(length(state))) {
     if (!any(as.character(state[i]) == pkg.env$state)) {
