+++ R/hudchas.R
@@ -114,7 +114,7 @@ hud_chas_state <- function(state, year = c("2014-2018"),
   if (all(nchar(state) > 2)) state <- capitalize(tolower(state))
 
   if (is.null(pkg.env$state)) {
-    pkg.env$state <- supressMessages(hud_nation_states_territories(key = Sys.getenv("HUD_KEY")))
+    pkg.env$state <- suppressMessages(hud_nation_states_territories(key = Sys.getenv("HUD_KEY")))
   }
 
   for (i in seq_len(length(state))) {
@@ -217,7 +217,7 @@ hud_chas_county <- function(county, year = c("2014-2018"),
   # Grab all possible states if the package doesn't have a state dataset to
   # query from.
   if (is.null(pkg.env$state)) {
-    pkg.env$state <- supressMessages(hud_nation_states_territories(key = Sys.getenv("HUD_KEY")))
+    pkg.env$state <- suppressMessages(hud_nation_states_territories(key = Sys.getenv("HUD_KEY")))
   }
 
   # Check if first two numbers of county code inputted is a valid state.
@@ -306,7 +306,7 @@ hud_chas_state_mcd <- function(state, year = c("2014-2018"),
   if (all(nchar(state) > 2)) state <- capitalize(tolower(state))
 
   if (is.null(pkg.env$state)) {
-    pkg.env$state <- supressMessages(hud_nation_states_territories(key = Sys.getenv("HUD_KEY")))
+    pkg.env$state <- suppressMessages(hud_nation_states_territories(key = Sys.getenv("HUD_KEY")))
   }
 
   for (i in seq_len(length(state))) {
@@ -334,7 +334,7 @@ hud_chas_state_mcd <- function(state, year = c("2014-2018"),
   fip_code <- unlist(fip_code)
 
   # Get all MCDs in these states...
-  all_mcd_in_states <- hud_state_minor_civil_divisions(fip_code)
+  all_mcd_in_states <- suppressMessages(hud_state_minor_civil_divisions(fip_code))
   all_queries <- data.frame()
 
   for (i in year) {
@@ -408,7 +408,7 @@ hud_chas_state_place <- function(state, year = c("2014-2018"),
   key <- args[[2]]
 
   if (is.null(pkg.env$state)) {
-    pkg.env$state <- supressMessages(hud_nation_states_territories(key = Sys.getenv("HUD_KEY")))
+    pkg.env$state <- suppressMessages(hud_nation_states_territories(key = Sys.getenv("HUD_KEY")))
   }
 
   if (all(nchar(state) == 2)) state <- toupper(state)
@@ -439,7 +439,7 @@ hud_chas_state_place <- function(state, year = c("2014-2018"),
   fip_code <- unlist(fip_code)
 
   # Get all places in these states...
-  all_places_in_states <- hud_state_places(fip_code)
+  all_places_in_states <- suppressMessages(hud_state_places(fip_code))
   all_queries <- data.frame()
 
   for (i in year) {
