### 2022-05-31 3c1602c
bug fix on hud_chas with specify entire name of type
