### 2022-06-01 efa2291
Updating readme and news files to include new functions to determine geographic inclusion and queries to determine most recent released data from HUD APIs.
