### 2022-06-01 5efab69
Removed dependencies zoo and rio. Added WIP items into the NEWS.md. Renamed R files to be more intuitive. Some vignettes are updated to have parameters for functions displayed in a table format. Geoid in geoid functions are fully developed, but need to be tested.
