+++ R/hud_geo_is_infix_geo.R
@@ -1,0 +1,370 @@
+#' @name %z_in_trt%
+#' @title %z_in_trt%
+#' @description Given a zip code and a tract, determine if they overlap
+#'   using the crosswalk files. Overlap will be described if
+#'   any residential, business, other, or total addresses reside in both.
+#' @param zip The zip to determine overlap with tract
+#' @param tract The tract to determine overlap with zip
+#' @export
+#' @examples
+#' \dontrun{
+#'
+#' 71052 %z_in_trt% 22031950600
+#'
+#' }
+`%z_in_trt%` <- function(zip, tract) {
+
+  args <- hud_rec_cw_yr()
+
+  # TODO: We might want to allow using names also...
+
+  if (as.character(tract) %in%
+      as.character(suppressMessages(hud_cw_zip_tract(zip,
+                                                     minimal = TRUE,
+                                                     year = args[1],
+                                                     quarter = args[2])))) {
+    return(TRUE)
+  }
+  return(FALSE)
+}
+
+
+
+#' @name %z_in_cty%
+#' @title %z_in_cty%
+#' @description Given a zip code and a county, determine if they overlap
+#'   using the crosswalk files. Overlap will be described if
+#'   any residential, business, other, or total addresses reside in both.
+#' @param zip The zip to determine overlap with tract
+#' @param county The county to determine overlap with zip
+#' @export
+#' @examples
+#' \dontrun{
+#'
+#' 71052 %z_in_cty% 22031
+#'
+#' }
+`%z_in_cty%` <- function(zip, county) {
+
+  args <- hud_rec_cw_yr()
+
+  if (as.character(county) %in%
+      as.character(suppressMessages(hud_cw_zip_county(zip,
+                                                      minimal = TRUE,
+                                                      year = args[1],
+                                                      quarter = args[2]
+      )))) {
+    return(TRUE)
+  }
+  return(FALSE)
+
+}
+
+
+#' @name %z_in_cbsa%
+#' @title %z_in_cbsa%
+#' @description Given a zip code and a cbsa, determine if they overlap
+#'   using the crosswalk files. Overlap will be described if
+#'   any residential, business, other, or total addresses reside in both.
+#' @param zip The zip to determine overlap with cbsa
+#' @param cbsa The cbsa to determine overlap with zip
+#' @export
+#' @examples
+#' \dontrun{
+#'
+#' 71052 %z_in_cbsa% 43340
+#'
+#' }
+`%z_in_cbsa%` <- function(zip, cbsa) {
+
+  args <- hud_rec_cw_yr()
+
+  if (as.character(cbsa) %in%
+      as.character(suppressMessages(hud_cw_zip_cbsa(zip,
+                                                    minimal = TRUE,
+                                                    year = args[1],
+                                                    quarter = args[2]
+      )))) {
+    return(TRUE)
+  }
+  return(FALSE)
+}
+
+
+#' @name %z_in_cbsadiv%
+#' @title %z_in_cbsadiv%
+#' @description Given a zip code and a cbsadiv, determine if they overlap
+#'   using the crosswalk files. Overlap will be described if
+#'   any residential, business, other, or total addresses reside in both.
+#' @param zip The zip to determine overlap with cbsadiv
+#' @param cbsadiv The cbsadiv to determine overlap with zip
+#' @export
+#' @examples
+#' \dontrun{
+#'
+#' 71052 %z_in_cbsadiv% 43340
+#'
+#' }
+`%z_in_cbsadiv%` <- function(zip, cbsadiv) {
+
+  args <- hud_rec_cw_yr()
+
+  if (as.character(cbsadiv) %in%
+      as.character(suppressMessages(hud_cw_zip_cbsadiv(zip,
+                                                       minimal = TRUE,
[…]
