+++ R/hud_geo_is_in_geo.R
@@ -1,0 +1,660 @@
+#' @name z_in_trt
+#' @title z_in_trt
+#' @description Given a zip code and a tract, determine if they overlap
+#'   using the crosswalk files. Overlap will be described if
+#'   any residential, business, other, or total addresses reside in both.
+#' @param zip The zip to determine overlap with tract
+#' @param tract The tract to determine overlap with zip
+#' @param year The year of the crosswalk files.
+#' @param quarter The quarter of the crosswalk files.
+#' @seealso
+#' * [rhud::z_in_trt()]
+#' * [rhud::z_in_county()]
+#' * [rhud::z_in_cbsa()]
+#' * [rhud::z_in_cbsadiv()]
+#' * [rhud::z_in_ctysb()]
+#' * [rhud::z_cd()]
+#' * [rhud::trt_in_z()]
+#' * [rhud::cty_in_z()]
+#' * [rhud::cbsa_in_z()]
+#' * [rhud::cbsadiv_in_z()]
+#' * [rhud::cd_in_z()]
+#' * [rhud::ctysb_in_z()]
+#' @export
+#' @examples
+#' \dontrun{
+#'
+#' z_in_trt(zip = 71052, tract = 22031950600, year = 2019, quarter = 2)
+#'
+#' }
+z_in_trt <- function(zip, tract, year, quarter) {
+
+  if (missing(year)) {
+    args <- hud_rec_cw_yr()
+    year = args[1]
+  }
+
+  if (missing(quarter)) {
+    args <- hud_rec_cw_yr()
+    quarter = args[2]
+  }
+
+  if (as.character(tract) %in%
+      as.character(suppressMessages(hud_cw_zip_tract(zip,
+                                                     minimal = TRUE,
+                                                     year = year,
+                                                     quarter = quarter)))) {
+    return(TRUE)
+  }
+
+  return(FALSE)
+}
+
+
+
+
+
+#' @name z_in_cty
+#' @title z_in_cty
+#' @description Given a zip code and a county, determine if they overlap
+#'   using the crosswalk files. Overlap will be described if
+#'   any residential, business, other, or total addresses reside in both.
+#' @param zip The zip to determine overlap with county
+#' @param county The county to determine overlap with zip
+#' @param year The year of the crosswalk files.
+#' @param quarter The quarter of the crosswalk files.
+#' @seealso
+#' * [rhud::z_in_trt()]
+#' * [rhud::z_in_county()]
+#' * [rhud::z_in_cbsa()]
+#' * [rhud::z_in_cbsadiv()]
+#' * [rhud::z_in_ctysb()]
+#' * [rhud::z_cd()]
+#' * [rhud::trt_in_z()]
+#' * [rhud::cty_in_z()]
+#' * [rhud::cbsa_in_z()]
+#' * [rhud::cbsadiv_in_z()]
+#' * [rhud::cd_in_z()]
+#' * [rhud::ctysb_in_z()]
+#' @export
+#' @examples
+#' \dontrun{
+#'
+#' z_in_cty(zip = 71052, county = 22031, year = 2019, quarter = 2)
+#'
+#' }
+z_in_cty <- function(zip, county, year, quarter) {
+
+  if (missing(year)) {
+    args <- hud_rec_cw_yr()
+    year = args[1]
+  }
+
+  if (missing(quarter)) {
+    args <- hud_rec_cw_yr()
+    quarter = args[2]
+  }
+
+  if (as.character(county) %in%
+      as.character(suppressMessages(hud_cw_zip_county(zip,
+                                                      minimal = TRUE,
+                                                      year = year,
[…]
