+++ R/hud_set_user_agent.R
@@ -1,0 +1,42 @@
+#' @name hud_set_user_agent
+#' @title hud_set_user_agent
+#' @description Set a user agent when querying the HUD User APIs.
+#' @param user A character vector with the user agent.
+#' @param in_wkdir set the key in the user's .RProfile in this directory.
+#'   Is defaulted to false.
+#' @param in_home set the key in the user's HOME directory.
+#'   Is defaulted to false.
+#' @export
+#' @seealso
+#' * [rhud::hud_get_user_agent()]
+#' * [rhud::hud_set_user_agent()]
+#' @examples
+#' \dontrun{
+#'
+#' hud_get_key()
+#'
+#' }
+hud_set_user_agent <- function(user) {
+
+}
+
+
+
+#' @name hud_get_user_agent
+#' @title hud_get_user_agent
+#' @description Get the most recent user agent set.
+#' @returns A character vector with the user agent used for querying HUD User
+#'   APIs.
+#' @export
+#' @seealso
+#' * [rhud::hud_get_user_agent()]
+#' * [rhud::hud_set_user_agent()]
+#' @examples
+#' \dontrun{
+#'
+#' hud_get_key()
+#'
+#' }
+hud_get_user_agent <- function() {
+
+}
