+++ man/hud_fmr_state_metroareas.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/hudfmr.R
+% Please edit documentation in R/hud_fmr.R
 \name{hud_fmr_state_metroareas}
 \alias{hud_fmr_state_metroareas}
 \title{hud_fmr_state_metroareas}

[…]
