### 2022-06-01 861509a
Updating readme to include hud_set_user_agent and hud_get_user_agent.
