### 2022-06-02 7b3a6d8
implementing z_in_trt infix operator to allow for multiple zip code inputs to compare with tract
