### 2022-06-02 7e5422e
Fixing all geo infix geo functions to allow for multiple geoids onlhs and rhs
