+++ R/hud_geo_is_infix_geo.R
@@ -3,6 +3,10 @@
 #' @description Given a zip code and a tract, determine if they overlap
 #'   using the crosswalk files. Overlap will be described if
 #'   any residential, business, other, or total addresses reside in both.
+#'
+#'   This means that it is possible that certain geoids(lhs) which have boundaries
+#'   that intersect that of the queried geoids(rhs) that are excluded.
+#'
 #' @param zip The zip to determine overlap with tract
 #' @param tract The tract to determine overlap with zip
 #' @export
@@ -17,18 +21,35 @@
   args <- hud_rec_cw_yr()
 
   # TODO: We might want to allow using names also..
+  # There is a bit of overhead cost for doing individual queries because each
+  # zip will need individual calls to hud_cw_zip_tract... Could optimize by
+  # using internal functions...
+
+  # Need to validate tract..
+
+  tract <- unique(paste(trimws(as.character(tract), which = "both")))
+  if (FALSE %in% numbers_only(tract)) stop("\nTract inputs must only be numbers.",
+                                           call. = FALSE)
+  if (any(nchar(tract) != 11)) stop("\nTract inputs are not all of length 11.",
+                                 call. = FALSE)
+
   res <- c()
   for (i in seq_len(length(zip))) {
     if (any(as.character(suppressWarnings(suppressMessages(hud_cw_zip_tract(zip[i],
-                                                                        minimal = TRUE,
-                                                                        year = args[1],
-                                                                        quarter = args[2])))) %in% as.character(tract))) {
+                            minimal = TRUE,
+                            year = args[1],
+                            quarter = args[2])))) %in% as.character(tract))) {
+
+      # What if zip does not have any data (invalid zip). Might want to catch
+      # the warning and tell user... Furthermore, you could validate the
+      # tract... using hud_tract_... but that would more API calls which might
+      # not be worthwhile.
+
       res <- c(res, TRUE)
     } else {
       res <- c(res, FALSE)
     }
   }
-
 
   return(res)
 }
@@ -53,16 +74,31 @@
 
   args <- hud_rec_cw_yr()
 
-  if (as.character(county) %in%
-      as.character(suppressMessages(hud_cw_zip_county(zip,
-                                                      minimal = TRUE,
-                                                      year = args[1],
-                                                      quarter = args[2]
-      )))) {
-    return(TRUE)
-  }
-  return(FALSE)
+  # TODO: We might want to allow using names also..
 
+
+  county <- unique(paste(trimws(as.character(county), which = "both")))
+  if (FALSE %in% numbers_only(county)) stop("\nCounty inputs must only be numbers.",
+                                           call. = FALSE)
+  if (any(nchar(county) != 5)) stop("\nCounty inputs are not all of length 5.",
+                                    call. = FALSE)
+
+
+  res <- c()
+  for (i in seq_len(length(zip))) {
+    if (any(as.character(suppressWarnings(suppressMessages(hud_cw_zip_county(zip[i],
+                            minimal = TRUE,
+                            year = args[1],
+                            quarter = args[2])))) %in% as.character(county))) {
+
+
+      res <- c(res, TRUE)
+    } else {
+      res <- c(res, FALSE)
+    }
+  }
+
+  return(res)
 }
 
 
@@ -84,15 +120,27 @@
 
   args <- hud_rec_cw_yr()
 
-  if (as.character(cbsa) %in%
-      as.character(suppressMessages(hud_cw_zip_cbsa(zip,
-                                                    minimal = TRUE,
[…]
