+++ tests/testthat/test-test-infix_cw.R
@@ -1,0 +1,168 @@
+test_that("test %z_in_trt%", {
+  # When numerics have leading zeros, it might not work...
+
+  expect_equal(c(35213, 34321) %z_in_trt% c("01073010801"), c(TRUE, FALSE))
+
+  expect_equal(c(77032, 77396) %z_in_trt% c(24033800608), c(TRUE, TRUE))
+
+  expect_equal(c(20774, 20772) %z_in_trt% c(24033800608), c(TRUE, TRUE))
+
+
+  expect_error(c(20774, 1) %z_in_trt% c(24033800608))
+
+
+  expect_error(c(20774, 20772) %z_in_trt% c(240338008))
+})
+
+
+test_that("test %z_in_cty%", {
+  expect_equal(71052 %z_in_cty% 22031, TRUE)
+
+  expect_equal(c(71049, 71052, 71419, 71027) %z_in_cty% 22031, c(TRUE,
+                                                                 TRUE,
+                                                                 TRUE, TRUE))
+
+  expect_error(c(20774, 1) %z_in_cty% c(24033800))
+
+
+  expect_error(c(20774, 20772) %z_in_cty% c(240338008))
+})
+
+
+test_that("test %z_in_cbsa%", {
+
+  expect_equal(71052 %z_in_cbsa% 43340, TRUE)
+
+  expect_equal(98569 %z_in_cbsa% 10140, TRUE)
+
+  expect_equal(c(98520, 98541, 98526, 98571, 98559, 98583) %z_in_cbsa% 10140,
+               c(TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE))
+
+  expect_error(c(98520, 12) %z_in_cbsa% c(10140))
+
+  expect_error(c(20774, 12222) %z_in_cbsa% c(1014033))
+})
+
+
+test_that("test %z_in_cbsadiv%", {
+  expect_equal(35235 %z_in_cbsadiv% 13820, FALSE)
+
+  expect_equal(35071 %z_in_cbsadiv% 13820, FALSE)
+
+  # The zip to cbsadiv and cbsadiv to zip API might be broken for any
+  # years after 2017 and 4th quarter. Should report this to HUD...
+  expect_equal(c(35022, 35126, 35210, 35501, 35071, 35064) %z_in_cbsadiv% 13820,
+               c(FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE))
+
+  expect_error(c(98520, 12) %z_in_cbsadiv% c(1014220))
+
+  expect_error(c(20774, 122) %z_in_cbsadiv% c(1014033))
+})
+
+test_that("test %z_in_ctysb%", {
+  expect_equal(44214 %z_in_ctysb% 3910383426, TRUE)
+
+  expect_equal(44256 %z_in_ctysb% 3910383426, TRUE)
+
+  expect_equal(c(44273, 44256, 44254, 44217, 44215, 44251) %z_in_ctysb% 3910383426,
+               c(TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE))
+
+  expect_error(c(98520, 12) %z_in_ctysb% c(1014220))
+
+  expect_error(c(20774, 122) %z_in_ctysb% c(1014033))
+})
+
+test_that("test %z_in_cd%", {
+  expect_equal(70072 %z_in_cd% 2202, TRUE)
+
+  expect_equal(70117 %z_in_cd% 2202, TRUE)
+
+  expect_equal(c(70743, 70808) %z_in_cd% 2202,
+               c(TRUE, TRUE))
+
[…]
