### 2022-06-03 e55766c
Refactored all in infix geo in geo functions to encapsulate validation of rhs and custom warning messages for query calls. Tests are made and all should be passing.
