+++ R/hud_geo_is_infix_geo.R
@@ -4,8 +4,9 @@
 #'   using the crosswalk files. Overlap will be described if
 #'   any residential, business, other, or total addresses reside in both.
 #'
-#'   This means that it is possible that certain geoids(lhs) which have boundaries
-#'   that intersect that of the queried geoids(rhs) that are excluded.
+#'   This means that it is possible that certain geoids(lhs) are not counted even
+#'   though their boundaries intersect the queried geoids(rhs). This is likely
+#'   because addresses do not lie in their intersecting region.
 #'
 #' @param zip The zip to determine overlap with tract
 #' @param tract The tract to determine overlap with zip
@@ -26,25 +27,19 @@
   # using internal functions...
 
   # Need to validate tract..
-
-  tract <- unique(paste(trimws(as.character(tract), which = "both")))
-  if (FALSE %in% numbers_only(tract)) stop("\nTract inputs must only be numbers.",
-                                           call. = FALSE)
-  if (any(nchar(tract) != 11)) stop("\nTract inputs are not all of length 11.",
-                                 call. = FALSE)
+  tract <- geo_is_infix_rhs_cleansing(query = tract, "tract")
 
   res <- c()
+
   for (i in seq_len(length(zip))) {
-    if (any(as.character(suppressWarnings(suppressMessages(hud_cw_zip_tract(zip[i],
-                            minimal = TRUE,
-                            year = args[1],
-                            quarter = args[2])))) %in% as.character(tract))) {
 
-      # What if zip does not have any data (invalid zip). Might want to catch
-      # the warning and tell user... Furthermore, you could validate the
-      # tract... using hud_tract_... but that would more API calls which might
-      # not be worthwhile.
+    queried <- geo_is_infix_query_and_get_warnings(query = zip[i],
+                                                   f = hud_cw_zip_tract,
+                                                   year = args[1],
+                                                   quarter = args[2],
+                                                   querytype = "zip")
 
+    if (any(queried %in% as.character(tract))) {
       res <- c(res, TRUE)
     } else {
       res <- c(res, FALSE)
@@ -61,6 +56,11 @@
 #' @description Given a zip code and a county, determine if they overlap
 #'   using the crosswalk files. Overlap will be described if
 #'   any residential, business, other, or total addresses reside in both.
+#'
+#'   This means that it is possible that certain geoids(lhs) are not counted even
+#'   though their boundaries intersect the queried geoids(rhs). This is likely
+#'   because addresses do not lie in their intersecting region.
+#'
 #' @param zip The zip to determine overlap with tract
 #' @param county The county to determine overlap with zip
 #' @export
@@ -74,22 +74,18 @@
 
   args <- hud_rec_cw_yr()
 
-  # TODO: We might want to allow using names also..
-
-
-  county <- unique(paste(trimws(as.character(county), which = "both")))
-  if (FALSE %in% numbers_only(county)) stop("\nCounty inputs must only be numbers.",
-                                           call. = FALSE)
-  if (any(nchar(county) != 5)) stop("\nCounty inputs are not all of length 5.",
-                                    call. = FALSE)
-
+  county <- geo_is_infix_rhs_cleansing(query = county, "county")
 
   res <- c()
   for (i in seq_len(length(zip))) {
-    if (any(as.character(suppressWarnings(suppressMessages(hud_cw_zip_county(zip[i],
-                            minimal = TRUE,
-                            year = args[1],
-                            quarter = args[2])))) %in% as.character(county))) {
+
+    queried <- geo_is_infix_query_and_get_warnings(query = zip[i],
+                                                   f = hud_cw_zip_county,
[…]
