### 2022-06-03 9ff38eb
Added to NEWS.md added functions to get most recently released files for each program for year and quarter if quarter exists for the program. Added news about user agent access. Added additional details to new infix operators about their non-infix counterparts.
