### 2022-06-03 6ece1be
Adding all updated Rd files.
