+++ man/hud_set_user_agent.Rd
@@ -1,0 +1,34 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/hud_set_user_agent.R
+\name{hud_set_user_agent}
+\alias{hud_set_user_agent}
+\title{hud_set_user_agent}
+\usage{
+hud_set_user_agent(user)
+}
+\arguments{
+\item{user}{A character vector with the user agent.}
+
+\item{in_wkdir}{set the key in the user's .RProfile in this directory.
+Is defaulted to false.}
+
+\item{in_home}{set the key in the user's HOME directory.
+Is defaulted to false.}
+}
+\description{
+Set a user agent when querying the HUD User APIs.
+}
+\examples{
+\dontrun{
+library(rhud)
+
+hud_set_user_agent("im-the-user")
+
+}
+}
+\seealso{
+\itemize{
+\item \code{\link[=hud_get_user_agent]{hud_get_user_agent()}}
+\item \code{\link[=hud_set_user_agent]{hud_set_user_agent()}}
+}
+}
