### 2022-06-03 6ee2181
Updating README to have install and simple setup/example at the very top.
