+++ NAMESPACE
@@ -43,12 +43,14 @@ export(hud_fmr_metroarea_zip)
 export(hud_fmr_state_counties)
 export(hud_fmr_state_metroareas)
 export(hud_get_key)
+export(hud_get_user_agent)
 export(hud_il)
 export(hud_nation_states_territories)
 export(hud_rec_cw_yr)
 export(hud_rec_fmr_yr)
 export(hud_rec_il_yr)
 export(hud_set_key)
+export(hud_set_user_agent)
 export(hud_state_counties)
 export(hud_state_metropolitan)
 export(hud_state_minor_civil_divisions)
