### 2022-06-03 e889959
fixing documentation issues. adding tests for geo in geo non infix operators. Renamed rhud_cache_dir to rhud_set_cache_dir to be more clear-also added the getter function for this. Adding template test file for setting user agent.
