+++ man/rhud_set_cache_dir.Rd
@@ -1,13 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/hud_cache.R
-\name{rhud_cache_dir}
-\alias{rhud_cache_dir}
-\alias{rhud_clear_cache}
-\title{rhud_cache_dir}
+\name{rhud_set_cache_dir}
+\alias{rhud_set_cache_dir}
+\title{rhud_set_cache_dir}
 \usage{
-rhud_cache_dir(path = "NULL", in_wkdir = FALSE, in_home = FALSE)
-
-rhud_clear_cache()
+rhud_set_cache_dir(path = "NULL", in_wkdir = FALSE, in_home = FALSE)
 }
 \arguments{
 \item{path}{The full path to the desired cache directory. Only one can be
@@ -32,19 +29,10 @@ and is not in a sensitive location.
 
 Windows users: please note that you'll need to use double-backslashes or
 forward slashes when specifying your cache directory's path in R.
-
-Remove cached data from the caching directory that is used to
-store data retrieved using the rhud API calls. By default, rhud uses a
-non-persistent temporary directory given for an R session, but a user might
-have set another directory to use.
 }
 \examples{
 \dontrun{
-rhud_cache_dir("./an/example/path", in_wkdir = TRUE, in_home = TRUE)
-
-}
-\dontrun{
-rhud_cache_dir("./an/example/path", in_wkdir = TRUE, in_home = TRUE)
+rhud_set_cache_dir("./an/example/path", in_wkdir = TRUE, in_home = TRUE)
 
 }
 }
