+++ NAMESPACE
@@ -55,8 +55,8 @@ export(hud_state_counties)
 export(hud_state_metropolitan)
 export(hud_state_minor_civil_divisions)
 export(hud_state_places)
-export(rhud_cache_dir)
 export(rhud_clear_cache)
+export(rhud_set_cache_dir)
 export(rhud_website)
 export(trt_in_z)
 export(z_in_cbsa)

[…]
