### 2022-06-03 9b53455
Adding the updated caching functions in readme
