+++ README.Rmd
@@ -216,7 +216,8 @@ also provide miscellaneous supplemental APIs under them.
   
   | Caching |
   |------------------------------------------|
-  | `rhud_cache_dir()` |
+  | `rhud_set_cache_dir()` |
+  | `rhud_get_cache_dir()``|
   | `rhud_clear_cache()` | 
 
   - Utilities
