### 2022-06-03 131052c
Fixing package onLoad to show rhud_set_cache_dir()
