### 2022-06-03 78e981a
Spacing in NEWS.md
