+++ NEWS.md
@@ -22,7 +22,7 @@ rhud 0.4.0.9000 (6/10/2022)
     non-infix counterparts are also available.
     (WIP)
     
-  * %z_in_trt%`, `%z_in_cty%`, `%z_in_ctysb%`, `%z_in_cd%`, `%z_in_cbsa%`, `%z_in_cbsadiv%`
+  * %trt_in_z%`, `%cty_in_z%`, `%ctysb_in_z%`, `%cd_in_z%`, `%cbsa_in_z%`, `%cbsadiv_in_z%`
     are now available to help determine whether a tract, cd, county, countysub,
     cbsa, or cbsadiv overlaps a zip code using the crosswalk files. Currently 
     supports only numeric identifiers, but may in the future provide support 
@@ -31,11 +31,11 @@ rhud 0.4.0.9000 (6/10/2022)
 ### MINOR IMPROVEMENTS:
 
   * Type checking added for using hud_set_key() as a way of reducing 
-  latent errors.
+    latent errors.
   * Download bar now shows the current working url as well as the 
-  number of failed downloads. 
+    number of failed downloads. 
   * Using the hud_set_key method does not default to writing to the working 
-  directory .Rprofile.
+    directory .Rprofile.
   * Introduction message to the package is now more informative.
   * Functions now check if user enters years below what is allowed. (WIP)
   * State queries for hud_fmr no longer does double queries.
@@ -43,27 +43,27 @@ rhud 0.4.0.9000 (6/10/2022)
 ### BUG FIXES
 
   * Attempt two at removing download bar from appearing when first
-  loading in the  hud_nation_state_territories() in package environment when
-  validating state inputs. 
+    loading in the hud_nation_state_territories() in package environment when
+    validating state inputs. 
   
   * Fixed some bugs caused by hud_set_key not properly setting to the .Rprofile.
-  (WIP)
+    (WIP)
   
   * Fix some issues with numerics being passed with less digits because of
-  leading 0 truncation. (WIP)
+    leading 0 truncation. (WIP)
   
 ### DOCUMENTATION FIXES
   
   * Added documentation to tell user that queries will be a combination of the 
-  geoid, year, and quarter inputs. (WIP)
+    geoid, year, and quarter inputs. (WIP)
   * Included in README.md how to setup the key and provide a simple example.
   * Improved sample case studies in the website documentation. (WIP)
   * Added diagram for census geographies to the crosswalk website documentation
   to better communicate relationship among geographies. (WIP)
   * Introduced some iconography into documentation to help 
-  overall user experience.  (WIP)
+    overall user experience.  (WIP)
   * All functions are listed in the README.md page along with the range of years
-  that they can be queried for.
+    that they can be queried for.
   * Description of package now shows up when typing it in console.
   
 rhud 0.3.0.9000 (5/30/2022)
@@ -71,7 +71,7 @@ rhud 0.3.0.9000 (5/30/2022)
 
 ### NEW FEATURES:
 
-  * Added options() for the rhud package, allowing a user to allow caching and 
+  * Added options() for the rhud package, allowing a user to allow using
   tibbles. 
    
   * User is now given a `to_tibble` option for rhud functions that return a
@@ -81,26 +81,26 @@ rhud 0.3.0.9000 (5/30/2022)
 ### MINOR IMPROVEMENTS:
 
   * User is now presented with a message when attaching the package 
-  showing the various options that can that can be set for the package as
-  well as telling the user that they need to obtain a key from the HUD User
-  website. 
+    showing the various options that can that can be set for the package as
+    well as telling the user that they need to obtain a key from the HUD User
+    website. 
 
 ### BUG FIXES
 
   * Fixed state level FMR data returning two copies of very_low and not low
[…]
