### 2022-06-03 b362916
Adding sys variables to get in add user agent and cache
