### 2022-06-03 8f04e28
renaming caching function to start with hud instead of rhud, to make it more consistent with rest of functions in package
