+++ man/hud_set_cache_dir.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/hud_cache.R
-\name{rhud_set_cache_dir}
-\alias{rhud_set_cache_dir}
-\title{rhud_set_cache_dir}
+\name{hud_set_cache_dir}
+\alias{hud_set_cache_dir}
+\title{hud_set_cache_dir}
 \usage{
-rhud_set_cache_dir(path = "NULL", in_wkdir = FALSE, in_home = FALSE)
+hud_set_cache_dir(path = "NULL", in_wkdir = FALSE, in_home = FALSE)
 }
 \arguments{
 \item{path}{The full path to the desired cache directory. Only one can be
@@ -32,7 +32,14 @@ forward slashes when specifying your cache directory's path in R.
 }
 \examples{
 \dontrun{
-rhud_set_cache_dir("./an/example/path", in_wkdir = TRUE, in_home = TRUE)
+hud_set_cache_dir("./an/example/path", in_wkdir = TRUE, in_home = TRUE)
 
+}
+}
+\seealso{
+\itemize{
+\item \code{\link[=hud_get_cache_dir]{hud_get_cache_dir()}}
+\item \code{\link[=hud_set_cache_dir]{hud_set_cache_dir()}}
+\item \code{\link[=hud_clear_cache]{hud_clear_cache()}}
 }
 }
