+++ man/hud_get_cache_dir.Rd
@@ -1,0 +1,30 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/hud_cache.R
+\name{hud_get_cache_dir}
+\alias{hud_get_cache_dir}
+\title{hud_get_cache_dir}
+\usage{
+hud_get_cache_dir()
+}
+\value{
+A character vector with path to cached files. If none is set,
+will default to R temp session directory
+}
+\description{
+Get the path rhud is using to store cached files.
+}
+\examples{
+\dontrun{
+library(rhud)
+
+hud_get_cache_dir()
+
+}
+}
+\seealso{
+\itemize{
+\item \code{\link[=hud_get_cache_dir]{hud_get_cache_dir()}}
+\item \code{\link[=hud_set_cache_dir]{hud_set_cache_dir()}}
+\item \code{\link[=hud_clear_cache]{hud_clear_cache()}}
+}
+}
