+++ README.Rmd
@@ -216,9 +216,9 @@ also provide miscellaneous supplemental APIs under them.
   
   | Caching |
   |------------------------------------------|
-  | `rhud_set_cache_dir()` |
-  | `rhud_get_cache_dir()``|
-  | `rhud_clear_cache()` | 
+  | `hud_set_cache_dir()` |
+  | `hud_get_cache_dir()``|
+  | `hud_clear_cache()` | 
 
   - Utilities
   
