+++ man/hud_clear_cache.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/hud_cache.R
-\name{rhud_clear_cache}
-\alias{rhud_clear_cache}
-\title{rhud_clear_cache}
+\name{hud_clear_cache}
+\alias{hud_clear_cache}
+\title{hud_clear_cache}
 \usage{
-rhud_clear_cache()
+hud_clear_cache()
 }
 \description{
 Remove cached data from the caching directory that is used to
@@ -14,7 +14,15 @@ have set another directory to use.
 }
 \examples{
 \dontrun{
-rhud_clear_cache()
 
+hud_clear_cache()
+
+}
+}
+\seealso{
+\itemize{
+\item \code{\link[=hud_get_cache_dir]{hud_get_cache_dir()}}
+\item \code{\link[=hud_set_cache_dir]{hud_set_cache_dir()}}
+\item \code{\link[=hud_clear_cache]{hud_clear_cache()}}
 }
 }
