+++ man/rhud-exports.Rd
@@ -1,8 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/zzz.R
-\name{rhud-exports}
-\alias{rhud-exports}
-\title{rhud exported operators}
-\description{
-rhud exported operators
-}
