+++ README.md
@@ -218,10 +218,10 @@ The HUD also provide miscellaneous supplemental APIs under them.
 
 -   Caching
 
-| Caching                                                 |
-|---------------------------------------------------------|
-| `rhud_set_cache_dir()`                                  |
-| ``` rhud_get_cache_dir()``|   | ```rhud_clear_cache()\` |
+| Caching                                               |
+|-------------------------------------------------------|
+| `hud_set_cache_dir()`                                 |
+| ``` hud_get_cache_dir()``|   | ```hud_clear_cache()\` |
 
 -   Utilities
 
