+++ NAMESPACE
@@ -24,6 +24,7 @@ export(hud_chas_nation)
 export(hud_chas_state)
 export(hud_chas_state_mcd)
 export(hud_chas_state_place)
+export(hud_clear_cache)
 export(hud_cw)
 export(hud_cw_cbsa_zip)
 export(hud_cw_cbsadiv_zip)
@@ -42,6 +43,7 @@ export(hud_fmr_county_zip)
 export(hud_fmr_metroarea_zip)
 export(hud_fmr_state_counties)
 export(hud_fmr_state_metroareas)
+export(hud_get_cache_dir)
 export(hud_get_key)
 export(hud_get_user_agent)
 export(hud_il)
@@ -49,14 +51,13 @@ export(hud_nation_states_territories)
 export(hud_rec_cw_yr)
 export(hud_rec_fmr_yr)
 export(hud_rec_il_yr)
+export(hud_set_cache_dir)
 export(hud_set_key)
 export(hud_set_user_agent)
 export(hud_state_counties)
 export(hud_state_metropolitan)
 export(hud_state_minor_civil_divisions)
 export(hud_state_places)
-export(rhud_clear_cache)
-export(rhud_set_cache_dir)
 export(rhud_website)
 export(trt_in_z)
 export(z_in_cbsa)
