### 2022-06-03 0bf39ee
Adding more details on how queries are constructed for all functions. ie combinbations of query + year + quarter.
