### 2022-06-03 7a9a23f
Adding newline to start and end of linewrites in key, cache, and user agent writes to system variables. Fixing in_home=TRUE writing to working directory .Rprofile.
