### 2022-06-03 22cff18
Adding more details to geo in geo infix operators.
