+++ man/grapes-ctysb_in_z-grapes.Rd
@@ -7,9 +7,12 @@
 countysub \%ctysb_in_z\% zip
 }
 \arguments{
-\item{zip}{The zip to determine overlap with countysub}
+\item{countysub}{The countysub to determine overlap with zip}
 
-\item{ctysb}{The countysub to determine overlap with zip}
+\item{zip}{The zip to determine overlap with countysub}
+}
+\value{
+If countysub(s) exist in the zip(s) specified, then TRUE is returned.
 }
 \description{
 Given a countysub and zip code, determine if they overlap
@@ -19,6 +22,10 @@ any residential, business, other, or total addresses reside in both.
 This means that it is possible that certain geoids(lhs) are not counted even
 though their boundaries intersect the queried geoids(rhs). This is likely
 because addresses do not lie in their intersecting region.
+
+Infix operators default to most recent year and quarter of which the
+crosswalk files are available. For access to the year and quarter parameter,
+take a look at their non-infix versions.
 }
 \examples{
 \dontrun{
