### 2022-06-03 49463de
Adding more details to geo in geo infix operators.
