### 2022-06-03 72e0142
Adding more details on how caching works. Also fixed some bugs when switching caching folders
