+++ R/hud_cache.R
@@ -17,6 +17,11 @@
 #' Windows users: please note that you'll need to use double-backslashes or
 #' forward slashes when specifying your cache directory's path in R.
 #'
+#' * This will create a folder called rhud_cache in specified path
+#'
+#' * If setting a new directory, the previous cache will not be cleaned. In
+#'   a future update there will be a garbage collection option.
+#'
 #' @param path The full path to the desired cache directory. Only one can be
 #'  set at a time. If no path is specified, it will use the temp directory
 #'  for R. The temp directory is not persistent.
@@ -34,7 +39,7 @@
 #' hud_set_cache_dir("./an/example/path", in_wkdir = TRUE, in_home = TRUE)
 #'
 #' }
-hud_set_cache_dir <- function(path = "NULL",
+hud_set_cache_dir <- function(path,
                               in_wkdir = FALSE,
                               in_home = FALSE
                               ) {
@@ -43,6 +48,12 @@ hud_set_cache_dir <- function(path = "NULL",
 
   # TODO: First check for valid path formatting by regex
 
+  if (missing(path)) {
+    path <- paste(tempdir(), "//rhud_cache", sep = "")
+    message("* Missing path, will set to the working session temp directory.")
+  } else {
+    path <- paste(path, "//rhud_cache", sep = "")
+  }
 
   if (!is.character(path) || length(path) != 1) {
     stop(paste("Make sure argument key is of type",
@@ -59,6 +70,7 @@ hud_set_cache_dir <- function(path = "NULL",
   }
 
   Sys.setenv("RHUD_CACHE_DIR" = path)
+  suppressWarnings(R.cache::setCacheRootPath(path))
   message("* Setting the RHUD_CACHE_DIR variable for the working session.")
 
   if (in_wkdir) {
