### 2022-06-03 8c44471
fixing cache functions in start up message
