+++ R/zzz.R
@@ -35,9 +35,9 @@ NULL
                               "* To get tibbles instead of dataframes, ",
                               "use `options(rhud_use_tibble = TRUE)`\n",
                               "* Caching is enabled by default. To set a new cache ",
-                              "directory use rhud_set_cache_dir().\n",
+                              "directory use hud_set_cache_dir().\n",
                               "* To delete cached ",
-                              "items, use rhud_clear_cache()\n",
+                              "items, use hud_clear_cache()\n",
                               "* Set these in your R script for single session ",
                               "or in .Rprofile for persistence.\n",
                               "* For more information on how to use this package, ",
