### 2022-06-03 74ba81c
Added instructions to set key in attach message
