### 2022-06-04 3d51e3e
Fully developed the non-infix versions of the geo in geo family of functions. Fixed issues with setting cache using key argument instead of path. Fixed issues with certain cw functions not cleansingproperly because of added primarygeoid and secondary geoid needed to work. Moved startup messages to onAttach.
