+++ R/zzz.R
@@ -47,6 +47,15 @@ NULL
                               "------------------------------------------------\n",
                               sep = ""))
 
+  if (Sys.getenv("RHUD_CACHE_DIR") == "" ||
+      Sys.getenv("RHUD_CACHE_DIR") == "NULL") {
+    packageStartupMessage("\n* Setting rhud cache to R temp directory.")
+
+  } else {
+    packageStartupMessage(paste("\n* Setting rhud cache to",
+                                Sys.getenv("RHUD_CACHE_DIR"),
+                                "directory."))
+  }
 }
 
 
@@ -60,12 +69,8 @@ NULL
 
   if(Sys.getenv("RHUD_CACHE_DIR") == "" || Sys.getenv("RHUD_CACHE_DIR") == "NULL") {
     suppressWarnings(R.cache::setCacheRootPath(paste(tempdir(),"//", "rhud_cache", sep ="")))
-    packageStartupMessage("\n* Setting rhud cache to R temp directory.")
   } else {
     suppressWarnings(R.cache::setCacheRootPath(Sys.getenv("RHUD_CACHE_DIR")))
-    packageStartupMessage(paste("\n* Setting rhud cache to",
-                                Sys.getenv("RHUD_CACHE_DIR"),
-                                "directory."))
   }
 
 }
