+++ R/hud_geo_is_in_geo.R
@@ -29,14 +29,15 @@
 #' }
 z_in_trt <- function(zip, tract, year, quarter) {
 
-  if (missing(year)) {
+  if (missing(year) || missing(quarter)) {
     args <- hud_rec_cw_yr()
-    year = args[1]
-  }
+    if (missing(year)) {
+      year = args[1]
+    }
 
-  if (missing(quarter)) {
-    args <- hud_rec_cw_yr()
-    quarter = args[2]
+    if (missing(quarter)) {
+      quarter = args[2]
+    }
   }
 
   args <- hud_rec_cw_yr()
@@ -104,25 +105,38 @@ z_in_trt <- function(zip, tract, year, quarter) {
 #' }
 z_in_cty <- function(zip, county, year, quarter) {
 
-  if (missing(year)) {
+  if (missing(year) || missing(quarter)) {
     args <- hud_rec_cw_yr()
-    year = args[1]
+    if (missing(year)) {
+      year = args[1]
+    }
+
+    if (missing(quarter)) {
+      quarter = args[2]
+    }
   }
 
-  if (missing(quarter)) {
-    args <- hud_rec_cw_yr()
-    quarter = args[2]
+  county <- geo_is_infix_rhs_cleansing(query = county, "county")
+
+  res <- c()
+  for (i in seq_len(length(zip))) {
+
+    queried <- geo_is_infix_query_and_get_warnings(query = zip[i],
+                                                   f = hud_cw_zip_county,
+                                                   year = args[1],
+                                                   quarter = args[2],
+                                                   querytype = "zip")
+
+    if (any(queried %in% as.character(county))) {
+
+
+      res <- c(res, TRUE)
+    } else {
+      res <- c(res, FALSE)
+    }
   }
 
-  if (as.character(county) %in%
-      as.character(suppressMessages(hud_cw_zip_county(zip,
-                                                      minimal = TRUE,
-                                                      year = year,
-                                                      quarter = quarter)))) {
-    return(TRUE)
-  }
-
-  return(FALSE)
+  return(res)
 }
 
 
@@ -159,25 +173,41 @@ z_in_cty <- function(zip, county, year, quarter) {
 #' }
 z_in_cbsa <- function(zip, cbsa, year, quarter) {
 
-  if (missing(year)) {
+  if (missing(year) || missing(quarter)) {
     args <- hud_rec_cw_yr()
-    year = args[1]
+    if (missing(year)) {
+      year = args[1]
+    }
+
+    if (missing(quarter)) {
+      quarter = args[2]
+    }
   }
 
-  if (missing(quarter)) {
-    args <- hud_rec_cw_yr()
-    quarter = args[2]
+  args <- hud_rec_cw_yr()
+
+
+  cbsa <- geo_is_infix_rhs_cleansing(query = cbsa, "cbsa")
+
+  res <- c()
+  for (i in seq_len(length(zip))) {
+
+    queried <- geo_is_infix_query_and_get_warnings(query = zip[i],
+                                                   f = hud_cw_zip_cbsa,
+                                                   year = args[1],
+                                                   quarter = args[2],
+                                                   querytype = "zip")
+
+    if (any(queried %in% as.character(cbsa))) {
+
+
+      res <- c(res, TRUE)
+    } else {
+      res <- c(res, FALSE)
+    }
   }
 
-  if (as.character(cbsa) %in%
-      as.character(suppressMessages(hud_cw_zip_cbsa(zip,
-                                                      minimal = TRUE,
-                                                      year = year,
-                                                      quarter = quarter)))) {
-    return(TRUE)
-  }
-
-  return(FALSE)
+  return(res)
 }
 
 
@@ -214,25 +244,38 @@ z_in_cbsa <- function(zip, cbsa, year, quarter) {
 #' }
[…]
