+++ tests/testthat/test-test-non_infix_cw.R
@@ -1,170 +1,205 @@
-test_that("test %z_in_trt%", {
+test_that("test z_in_trt", {
+  skip_if(Sys.getenv("HUD_KEY") == "")
   # When numerics have leading zeros, it might not work...
 
-  # Invalid zipcode 34321 should throw warning....
+  # Invalid zipcode 34321 should throw warning...
+
   expect_warning(expect_equal(z_in_trt(zip = c(35213, 34321),
                                        tract = c("01073010801")),
                               c(TRUE, FALSE)))
 
-  expect_equal(c(77032, 77396) %z_in_trt% c(24033800608), c(FALSE, FALSE))
+  expect_equal(z_in_trt(zip = c(77032, 77396), tract = c(24033800608)),
+               c(FALSE, FALSE))
 
-  expect_equal(c(20774, 20772) %z_in_trt% c(24033800608), c(TRUE, TRUE))
+  expect_equal(z_in_trt(zip = c(20774, 20772), tract = c(24033800608)),
+               c(TRUE, TRUE))
 
 
-  expect_error(c(20774, 1) %z_in_trt% c(24033800608))
+  expect_error(z_in_trt(zip = c(20774, 1), tract = c(24033800608)))
 
 
-  expect_error(c(20774, 20772) %z_in_trt% c(240338008))
+  expect_error(z_in_trt(zip = c(20774, 20772), tract = c(240338008)))
+
 })
 
 
-test_that("test %z_in_cty%", {
-  expect_equal(71052 %z_in_cty% 22031, TRUE)
+test_that("test z_in_cty", {
+  skip_if(Sys.getenv("HUD_KEY") == "")
 
-  expect_equal(c(71049, 71052, 71419, 71027) %z_in_cty% 22031, c(TRUE,
-                                                                 TRUE,
-                                                                 TRUE, TRUE))
+  expect_equal(z_in_cty(zip = 71052, county = 22031), TRUE)
 
-  expect_error(c(20774, 1) %z_in_cty% c(24033800))
+  expect_equal(z_in_cty(zip = c(71049, 71052, 71419, 71027), county = 22031),
+               c(TRUE, TRUE, TRUE, TRUE))
+
+  expect_error(z_in_cty(zip = c(20774, 1), county = c(24033800)))
 
 
-  expect_error(c(20774, 20772) %z_in_cty% c(240338008))
+  expect_error(z_in_cty(zip = c(20774, 20772), county = c(240338008)))
 })
 
 
-test_that("test %z_in_cbsa%", {
+test_that("test z_in_cbsa", {
+  skip_if(Sys.getenv("HUD_KEY") == "")
 
-  expect_equal(71052 %z_in_cbsa% 43340, TRUE)
+  expect_equal(z_in_cbsa(zip = 71052, cbsa = 43340), TRUE)
 
-  expect_equal(98569 %z_in_cbsa% 10140, TRUE)
+  expect_equal(z_in_cbsa(zip = 98569, cbsa = 10140), TRUE)
 
-  expect_equal(c(98520, 98541, 98526, 98571, 98559, 98583) %z_in_cbsa% 10140,
+  expect_equal(z_in_cbsa(zip = c(98520, 98541, 98526, 98571, 98559, 98583),
+                         cbsa = 10140),
                c(TRUE, TRUE, TRUE, TRUE, TRUE, TRUE))
 
-  expect_error(c(98520, 12) %z_in_cbsa% c(10140))
[…]

[…]
