### 2022-06-04 feaff39
Fixing warnings for tests in the misc functions that should not throw errors. Fixing issues in seealso documentation in get recent data functions. Fixed seealso functions in the geo in geo infix operators. Fixed documenation in helpers for the geo in geo functions. Fixed issuesin cw input check cleansing for making sure queries are in proper year and quarter range not taking into account multiple year inputs.
