+++ NEWS.md
@@ -1,4 +1,5 @@
 rhud 0.4.0.9000 (6/10/2022)
+============================
 
 ### NEW FEATURES:
   
