### 2022-06-04 824766d
Removing WIP tag on NEWS.md items for completed tasks
