### 2022-06-04 bf0216a
fixing formatting issue in functions displayed on readme
