+++ man/hud_rec_il_yr-open-paren-close-paren.Rd
@@ -24,6 +24,6 @@ hud_rec_il_yr()
 \itemize{
 \item \code{\link[=hud_rec_cw_yr]{hud_rec_cw_yr()}}
 \item \code{\link[=hud_rec_fmr_yr]{hud_rec_fmr_yr()}}
-\item \link{hud_rec_il_yr}
+\item \code{\link[=hud_rec_il_yr]{hud_rec_il_yr()}}
 }
 }
