### 2022-06-04 9cdb713
Adding return value documentation for geo in geo function calls.
