+++ man/z_in_cd.Rd
@@ -15,6 +15,9 @@ z_in_cd(zip, cd, year, quarter)
 
 \item{quarter}{The quarter of the crosswalk files.}
 }
+\value{
+If zip(s) exist in the cd(s) specified, then TRUE is returned.
+}
 \description{
 Given a zip code and a congressional district, determine
 if they overlap using the crosswalk files. Overlap will be described if
