+++ man/z_in_cbsadiv.Rd
@@ -15,6 +15,9 @@ z_in_cbsadiv(zip, cbsadiv, year, quarter)
 
 \item{quarter}{The quarter of the crosswalk files.}
 }
+\value{
+If zip(s) exist in the cbsadiv(s) specified, then TRUE is returned.
+}
 \description{
 Given a zip code and a cbsadiv, determine if they overlap
 using the crosswalk files. Overlap will be described if
