+++ man/cbsa_in_z.Rd
@@ -15,6 +15,9 @@ cbsa_in_z(cbsa, zip, year, quarter)
 
 \item{quarter}{The quarter of the crosswalk files.}
 }
+\value{
+If cbsa(s) exist in the zip(s) specified, then TRUE is returned.
+}
 \description{
 Given a cbsa and a zip, determine if they overlap
 using the crosswalk files. Overlap will be described if
