+++ man/hud_rec_fmr_yr.Rd
@@ -6,6 +6,10 @@
 \usage{
 hud_rec_fmr_yr()
 }
+\value{
+The most recent year for the fmr files for state, county, and
+metroarea queries.
+}
 \description{
 Ping the Fair Markets Rent API provided by HUD User to
 determine the most recently released files. This will only ping
