+++ man/cd_in_z.Rd
@@ -15,6 +15,9 @@ cd_in_z(cd, zip, year, quarter)
 
 \item{quarter}{The quarter of the crosswalk files.}
 }
+\value{
+If cd(s) exist in the zip(s) specified, then TRUE is returned.
+}
 \description{
 Given a congressional district and a zip, determine if they
 overlap using the crosswalk files. Overlap will be described if
