+++ man/z_in_cty.Rd
@@ -15,6 +15,9 @@ z_in_cty(zip, county, year, quarter)
 
 \item{quarter}{The quarter of the crosswalk files.}
 }
+\value{
+If zip(s) exist in the county(s) specified, then TRUE is returned.
+}
 \description{
 Given a zip code and a county, determine if they overlap
 using the crosswalk files. Overlap will be described if
