+++ man/z_in_ctysb.Rd
@@ -1,9 +1,8 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/hud_geo_is_in_geo.R
-\name{z_in_ctysub}
-\alias{z_in_ctysub}
+\name{z_in_ctysb}
 \alias{z_in_ctysb}
-\title{z_in_ctysub}
+\title{z_in_ctysb}
 \usage{
 z_in_ctysb(zip, countysub, year, quarter)
 }
@@ -27,7 +26,7 @@ any residential, business, other, or total addresses reside in both.
 \examples{
 \dontrun{
 
-z_in_ctysub(zip = 35213, countysub = 0107390324, year = 2019, quarter = 2)
+z_in_ctysb(zip = 35213, countysub = 0107390324, year = 2019, quarter = 2)
 
 }
 }
