### 2022-06-05 9fb3159
Fixing spelling in functions present in readme.
