### 2022-06-06 ff23063
Added documentation to new infix operators to have desc above needing to set the HUD_KEY variable. Added a check to make sure user does not have an empty HUD_KEY variable for new infix operator. Added argument key to new functions to get most recent year and quarter of data. Changed all RHS geoid of infix operators and their non-infix to check to use internal cw_input_check instead of another check function.
