+++ R/hud_geo_is_in_geo.R
@@ -7,6 +7,8 @@
 #' @param tract The tract to determine overlap with zip
 #' @param year The year of the crosswalk files.
 #' @param quarter The quarter of the crosswalk files.
+#' @param key The API key for this user. You must go to HUD and sign up
+#'   for an account and request for an API key.
 #' @returns If zip(s) exist in the tract(s) specified, then TRUE is returned.
 #' @seealso
 #' * [rhud::z_in_trt()]
@@ -28,7 +30,8 @@
 #' z_in_trt(zip = 71052, tract = 22031950600, year = 2019, quarter = 2)
 #'
 #' }
-z_in_trt <- function(zip, tract, year, quarter) {
+z_in_trt <- function(zip, tract, year, quarter,
+                     key = Sys.getenv("HUD_KEY")) {
 
   if (missing(year) || missing(quarter)) {
     args <- hud_rec_cw_yr()
@@ -41,15 +44,25 @@ z_in_trt <- function(zip, tract, year, quarter) {
     }
   }
 
-  args <- hud_rec_cw_yr()
-
   # TODO: We might want to allow using names also..
   # There is a bit of overhead cost for doing individual queries because each
   # zip will need individual calls to hud_cw_zip_tract... Could optimize by
   # using internal functions...
 
   # Need to validate tract..
-  tract <- geo_is_infix_rhs_cleansing(query = tract, "tract")
+  cleaned <- cw_input_check_cleansing(primary_geoid = "tract",
+                                      secondary_geoid = "zip",
+                                      query = tract, year = year,
+                                      quarter = quarter,
+                                      Sys.getenv("HUD_KEY"))
+
+
+  tract <- cleaned[1]
+  year <- cleaned[2]
+  quarter <- cleaned[3]
+  key <- cleaned[4]
+
+  if (any(nchar(tract) != 11)) stop("\nInputted tract(s) are not all of length 11.")
 
   res <- c()
 
@@ -59,7 +72,9 @@ z_in_trt <- function(zip, tract, year, quarter) {
                                                    f = hud_cw_zip_tract,
                                                    year = year,
                                                    quarter = quarter,
-                                                   querytype = "zip")
+                                                   querytype = "zip",
+                                                   key = key
+                                                   )
 
     if (any(queried %in% as.character(tract))) {
       res <- c(res, TRUE)
@@ -84,6 +99,8 @@ z_in_trt <- function(zip, tract, year, quarter) {
 #' @param county The county to determine overlap with zip
 #' @param year The year of the crosswalk files.
 #' @param quarter The quarter of the crosswalk files.
+#' @param key The API key for this user. You must go to HUD and sign up
+#'   for an account and request for an API key.
 #' @seealso
 #' * [rhud::z_in_trt()]
 #' * [rhud::z_in_cty()]
@@ -105,7 +122,8 @@ z_in_trt <- function(zip, tract, year, quarter) {
 #' z_in_cty(zip = 71052, county = 22031, year = 2019, quarter = 2)
 #'
 #' }
-z_in_cty <- function(zip, county, year, quarter) {
+z_in_cty <- function(zip, county, year, quarter,
+                     key = Sys.getenv("HUD_KEY")) {
 
   if (missing(year) || missing(quarter)) {
     args <- hud_rec_cw_yr()
@@ -118,16 +136,30 @@ z_in_cty <- function(zip, county, year, quarter) {
     }
   }
 
-  county <- geo_is_infix_rhs_cleansing(query = county, "county")
+  cleaned <- cw_input_check_cleansing(primary_geoid = "county",
+                                      secondary_geoid = "zip",
+                                      query = county, year = year,
+                                      quarter = quarter,
+                                      Sys.getenv("HUD_KEY"))
+
+
+  county <- cleaned[1]
+  year <- cleaned[2]
[…]
