### 2022-06-06 7ee95dd
Adding plurality to geoids mentioned in documentation.Made years and quarters params in non-infix geo in geo to be more consistent to that described in crosswalk functions.
