+++ R/hud_geo_is_infix_geo.R
@@ -3,7 +3,7 @@
 #' @description To use this function, the HUD_KEY system environment variable
 #'   must be set first: use hud_set_key("the_key") to do this.
 #'
-#'   Given a zip code and a tract, determine if they overlap
+#'   Given zip code(s) and tract(s), determine if they overlap
 #'   using the crosswalk files. Overlap will be described if
 #'   any residential, business, other, or total addresses reside in both.
 #'
@@ -15,8 +15,8 @@
 #'   crosswalk files are available. For modifying the year and quarter parameter
 #'   take a look at their non-infix versions.
 #'
-#' @param zip The zip to determine overlap with tract
-#' @param tract The tract to determine overlap with zip
+#' @param zip The zip(s) to determine overlap with tract(s)
+#' @param tract The tract(s) to determine overlap with zip(s)
 #' @returns If zip(s) exist in the tract(s) specified, then TRUE is returned.
 #' @export
 #' @examples
@@ -85,7 +85,7 @@
 #' @description To use this function, the HUD_KEY system environment
 #'   variable must be set first: use hud_set_key("the_key") to do this.
 #'
-#'   Given a zip code and a county, determine if they overlap
+#'   Given zip code(s) and county(s), determine if they overlap
 #'   using the crosswalk files. Overlap will be described if
 #'   any residential, business, other, or total addresses reside in both.
 #'
@@ -97,8 +97,8 @@
 #'   crosswalk files are available. For modifying the year and quarter parameter,
 #'   take a look at their non-infix versions.
 #'
-#' @param zip The zip to determine overlap with tract
-#' @param county The county to determine overlap with zip
+#' @param zip The zip(s) to determine overlap with county(s)
+#' @param county The county(s) to determine overlap with zip(s)
 #' @returns If zip(s) exist in the county(s) specified, then TRUE is returned.
 #' @export
 #' @examples
@@ -158,7 +158,7 @@
 #' @description To use this function, the HUD_KEY system environment
 #'   variable must be set first: use hud_set_key("the_key") to do this.
 #'
-#'   Given a zip code and a cbsa, determine if they overlap
+#'   Given a zip code(s) and a cbsa(s), determine if they overlap
 #'   using the crosswalk files. Overlap will be described if
 #'   any residential, business, other, or total addresses reside in both.
 #'
@@ -170,8 +170,8 @@
 #'   crosswalk files are available. For modifying the year and quarter parameter,
 #'   take a look at their non-infix versions.
 #'
-#' @param zip The zip to determine overlap with cbsa
-#' @param cbsa The cbsa to determine overlap with zip
+#' @param zip The zip(s) to determine overlap with cbsa(s)
+#' @param cbsa The cbsa(s) to determine overlap with zip(s)
 #' @returns If zip(s) exist in the cbsa(s) specified, then TRUE is returned.
 #' @export
 #' @examples
@@ -232,7 +232,7 @@
 #' @description To use this function, the HUD_KEY system environment
 #'   variable must be set first: use hud_set_key("the_key") to do this.
 #'
-#'   Given a zip code and a cbsadiv, determine if they overlap
+#'   Given zip code(s) and cbsadiv(s), determine if they overlap
 #'   using the crosswalk files. Overlap will be described if
 #'   any residential, business, other, or total addresses reside in both.
 #'
@@ -244,8 +244,8 @@
 #'   crosswalk files are available. For modifying the year and quarter parameter,
 #'   take a look at their non-infix versions.
 #'
-#' @param zip The zip to determine overlap with cbsadiv
-#' @param cbsadiv The cbsadiv to determine overlap with zip
+#' @param zip The zip(s) to determine overlap with cbsadiv(s)
[…]
