+++ R/hud_geo_is_in_geo.R
@@ -1,12 +1,14 @@
 #' @name z_in_trt
 #' @title z_in_trt
-#' @description Given a zip code and a tract, determine if they overlap
+#' @description Given zip code(s) and tract(s), determine if they overlap
 #'   using the crosswalk files. Overlap will be described if
 #'   any residential, business, other, or total addresses reside in both.
-#' @param zip The zip to determine overlap with tract
-#' @param tract The tract to determine overlap with zip
-#' @param year The year of the crosswalk files.
-#' @param quarter The quarter of the crosswalk files.
+#' @param zip The zip(s) to determine overlap with tract(s).
+#' @param tract The tract(s) to determine overlap with zip(s).
+#' @param year Gets the year that this data was recorded. Can specify multiple
+#'   years. Default is the previous year.
+#' @param quarter Gets the quarter of the year that this data was recorded.
+#'   Defaults to the first quarter of the year.
 #' @param key The API key for this user. You must go to HUD and sign up
 #'   for an account and request for an API key.
 #' @returns If zip(s) exist in the tract(s) specified, then TRUE is returned.
@@ -92,13 +94,15 @@ z_in_trt <- function(zip, tract, year, quarter,
 
 #' @name z_in_cty
 #' @title z_in_cty
-#' @description Given a zip code and a county, determine if they overlap
+#' @description Given zip code(s) and county(s), determine if they overlap
 #'   using the crosswalk files. Overlap will be described if
 #'   any residential, business, other, or total addresses reside in both.
-#' @param zip The zip to determine overlap with county
-#' @param county The county to determine overlap with zip
-#' @param year The year of the crosswalk files.
-#' @param quarter The quarter of the crosswalk files.
+#' @param zip The zip(s) to determine overlap with county(s).
+#' @param county The county(s) to determine overlap with zip(s).
+#' @param year Gets the year that this data was recorded. Can specify multiple
+#'   years. Default is the previous year.
+#' @param quarter Gets the quarter of the year that this data was recorded.
+#'   Defaults to the first quarter of the year.
 #' @param key The API key for this user. You must go to HUD and sign up
 #'   for an account and request for an API key.
 #' @seealso
@@ -178,13 +182,15 @@ z_in_cty <- function(zip, county, year, quarter,
 
 #' @name z_in_cbsa
 #' @title z_in_cbsa
-#' @description Given a zip code and a cbsa, determine if they overlap
+#' @description Given zip code(s) and cbsa(s), determine if they overlap
 #'   using the crosswalk files. Overlap will be described if
 #'   any residential, business, other, or total addresses reside in both.
-#' @param zip The zip to determine overlap with cbsa
-#' @param cbsa The cbsa to determine overlap with zip
-#' @param year The year of the crosswalk files.
-#' @param quarter The quarter of the crosswalk files.
+#' @param zip The zip(s) to determine overlap with cbsa(s).
+#' @param cbsa The cbsa(s) to determine overlap with zip(s).
+#' @param year Gets the year that this data was recorded. Can specify multiple
+#'   years. Default is the previous year.
+#' @param quarter Gets the quarter of the year that this data was recorded.
+#'   Defaults to the first quarter of the year.
 #' @param key The API key for this user. You must go to HUD and sign up
 #'   for an account and request for an API key.
 #' @returns If zip(s) exist in the cbsa(s) specified, then TRUE is returned.
@@ -264,13 +270,15 @@ z_in_cbsa <- function(zip, cbsa, year, quarter,
 
 #' @name z_in_cbsadiv
 #' @title z_in_cbsadiv
-#' @description Given a zip code and a cbsadiv, determine if they overlap
+#' @description Given zip code(s) and cbsadiv(s), determine if they overlap
[…]
