+++ man/grapes-cbsadiv_in_z-grapes.Rd
@@ -7,9 +7,9 @@
 cbsadiv \%cbsadiv_in_z\% zip
 }
 \arguments{
-\item{cbsadiv}{The cbsadiv to determine overlap with zip}
+\item{cbsadiv}{The cbsadiv(s) to determine overlap with zip(s).}
 
-\item{zip}{The zip to determine overlap with cbsadiv}
+\item{zip}{The zip(s) to determine overlap with cbsadiv(s).}
 }
 \value{
 If cbsadiv(s) exist in the zip(s) specified, then TRUE is returned.
@@ -18,7 +18,7 @@ If cbsadiv(s) exist in the zip(s) specified, then TRUE is returned.
 To use this function, the HUD_KEY system environment
 variable must be set first: use hud_set_key("the_key") to do this.
 
-Given a cbsadiv and a zip, determine if they overlap
+Given cbsadiv(s) and zip(s), determine if they overlap
 using the crosswalk files. Overlap will be described if
 any residential, business, other, or total addresses reside in both.
 

[…]
