### 2022-06-07 #524 @rtaph
- I have no prior relationship with any of the package authors.

  - The statement is there but I think "the problem it solves" could be made clearer. What I found a bit hard to immediately undestand was _how_ rhud solves the stipulated problem.
- [x] **Installation instructions:** for the development version of package and any non-standard dependencies in README.
  - I was able to install `rhud` without problems using the instructions on the README (however, see below about building it with vignettes).
  - These could be improved on a bit by commenting on the output from examples more. For an unitiated user, the context and applications would be helpful.
  - Some examples in the docs appear broken for internal functions. `devtools::run_examples(run_dontrun = TRUE)` does not work for all examples. See below.
  - Some sections of your `CONTRIBUTING` file appear unfinished.
  - I think you [still need a Zenodo DOI](https://devguide.ropensci.org/building.html#citation-file) in your `CITATION` file.

	- I am able to run `devtools::load_all()` successfully, but `devtools::check()` fails for me in the vignette-building stage. See my comment further down about pre-building your vignettes.
    - There are some small issues detailed below.

---

Good job! I enjoyed going through your package and learning more about this API. On the whole, the package seems useful and I am glad to see that it will make it easier to interact the the HUD API. Thank you for inviting me to offer my review and critique.

From the version I reviewed ([84c8d](https://github.com/etam4260/rhud/tree/84c8d5b978decea4e8e45a8070545f171293207f)), there are areas I suggest improving prior to publication with ROpenSci. This includes opportunities to:
- **tidy-up** code by removing unused functions, addressing straggler `#TODO`s, dropping functions with empty bodies, etc.
- **refactor** blocks of code that are frequently repeated, that can be simplified, or that can be made more human-readable.
- **improve documentation** to make it easier for a newcommer to grasp the nature of `rhud` regardless of their entrypoint into the documentation.

Below is a detailed list of suggestions-- some of which you may already have addressed or thought about. It is a lot to go through, so feel free to disagree and take only what you find valuable. If there are elements that you agree with but which are ambitious, some can be tasked for future development.

**Docs and Communication**

- It took me some time to understand what what _crosswalk_ meant. The word is used as a noun, verb, and adjective, and it was not until reading the full linked journal article that I think I think I understood it. It might be helpful to explain this concept to readers.
- The ROpenSci package guidance says
  > The README, the top-level package docs, vignettes, websites, etc., should all have enough information at the beginning to get a high-level overview of the package and the services/data it connects to, and provide navigation to other relevant pieces of documentation.

  I don't think all vignettes yet provide an easy entry point. I would encourage adding to their text to make it clearer.
- In the `README`, it might be helpful to reference R packages that achieve something similar by alternate means. I'm thinking of GIS analogues like `tongfen` and `areal`.
- There are some un-introduced abbreviations such as "PD&R". I suggest reviewing all the docs to check for other naked abbreviations.
- In general I found the examples from the vignettes a bit hard to follow. There is an opportunity to improve them by adding some commentary on the output. This is probably not needed for all flavours of cross-walk, but commenting on the first example in your vignette could really help.
- In vignettes, I would place lowercase abbreviations either in codeblocks (e.g., `cbsa`) or in uppercase (e.g., "CBSA")
- I would encourage organizing your pkgdown [reference](https://pkgdown.r-lib.org/articles/pkgdown.html#reference) page into sections in the `_pkgdown.yml` to make it easier to navigate. Perhaps grouping them by API?
- I would use `@title` roxygen2 tage to give a descriptive title rather than duplicating the function name.
- I would reorganize your `README` to put the most important information up top, and other information like citation and disclaimers lower down.
- As per the ROpenSci packaging guidance, I would encourage the use of the `@family` roxygen tag to simplify `@seealso` sections.
- I think `cw_do_query_calls` could be better documented in terms of what the function is meant to achieve. I had to look at the actual code to understand it, as the `@description` was not specific enough for me to understand.
- "query: character double integer or numeric":  This mixes data types with data classes. Can we remove `double` here?
- Although not required, it would be useful to have an `@example` for internal functions. This would be helpful for reviewers or future code maintainers even if you have `@noRd`.
- In your `@param` roxygen blocks, I would encourage describing whether certain parameters can accept vectors or scalar values only. Something like `@param state A character vector indicating the state to ...`
- I would also review all your documentation to ensure that the expected class is clearly mentioned in the `@param`. For instance, in `hud_chas` it is not clear (unless you look at the example) if the input needs to be `1`, `"1"`, `"National"`, or `"1 - National"`.
- `Setup.Rmd`:
  - Instead of `hud_{1}_{2}_{3}`, I would use mnemonic names. For example, `hud_{dataset}_{geo}_{resolution}` or `hud_{api}_{from}_{to}`.
- [This blog](https://blog.r-hub.io/2020/06/03/vignettes/#how-to-include-a-compute-intensive--authentication-dependent-vignette) encourages shipping API-wrapping packages with precomputed vignettes. That might be something to consider to allow the package to build faster and without an internet connection and/or API key. I suspect this is also the issue preventing `devtools::check()` to pass on my machine.
- I would suggest not having `Sys.setenv("HUD_KEY" = "q3r2rjimd129fj121jid")` in examples because a) it is not a real key; and b) it will cause failures even if a user has a key because running this line overwrites their own environment variable (I did when I ran `devtools::run_examples(run_dontrun = TRUE)`). If you instead instruct users how to set their `HUD_KEY` persistently iff it is missing, they should be able to run the examples.

**Tests**

- I would add a message to `skip_if()`. For example, to know that tests are being skipped because the `HUD_KEY` is missing. Checking for the `HUD_KEY` is so common that I think it warrants a wrapper `skip_if_no_hud_key()`.
- I would review the name of the unit tests to make them conform to an imperative tense. For example, "test awkward crosswalk" could be changed to "an improper crosswalk request must error" so that it plays nice with the phrase "test that ..."
-  It would be helpful to be able to _quiet_ the printing of the status bar, especially for unit tests. Consider creating an `rhud.quiet` option to silence these messages. This would be particularly handy in unit tests. What I have in mind is something akin to the `usethis.quiet` option in the usethis package.
- From the ROpenSci packaging guidance:
    > testthat has a function skip_on_cran() that you can use to not run tests on CRAN. We recommend using this on all functions that are API calls since they are quite likely to fail on CRAN. These tests should still run on continuous integration.
- I see that you have an empty test `with_mock_dir()`, so I suspect you are dabbling with webmocking and may already know about [this resource](https://books.ropensci.org/http-testing) by Scott Chamberlain and Maëlle Salmon. The book has fantastic guidance on how run some unit tests through moking. For example, `vcr` might be a great solution for testing the API without an internet connection, as it will snapshot real responses as test fixtures. If you feel even more adventurous, you could create a Github Workflow that tests live requests on a schedule (e.g. semi-monthly) using a [Github secret](https://docs.github.com/en/rest/actions/secrets) to store an API key.
- Two tests fail rather than skip when the API key is missing.

**Code**

- `hud_set_key()`:
  - In `vignettes/Setup.Rmd`, it could be useful to encourage users to set their environment key via `usethis::edit_r_environ()`. This would persist across sessions. Working with the renviron file seems cleaner than all the regex and string manipulation in `hud_set_key()`, though the key would be exposed to any package in R. In the medium term, it would be nice to also offer the option to use a `keyring`.
- `cw_input_check_cleansing`:
  - What about returning a named list instead of an unnamed list? This might allow you to just use something like `args$zip` and this would be less prone to breaking in future releases. I would also be more human-readable.
- `hud_get_key()`:
  - I suggest raising an informative error if the variable is not set.
- `.onAttach()`:
  - From the ROpenSci guidance:
    > Only use package startup messages when necessary (function masking for instance). Avoid package startup messages like “This is foobar 2.4-0” or citation guidance because they can be annoying to the user. Rely on documentation for such guidance.

     I would only notify the user on startup if their HUD key is missing, and take away all the messaging about features. If the key is found on package startup, I would provide no message at all.
- `check_is_not_list()`:
  - AFAIK, "numeric" is not an R type but rather a class, so I think `typeof(dataframe[, i]) != "numeric"` is not necessary. You could express the block in question more succinctly as
    ```
    types <- c("character", "integer", "double")
    typeof(dataframe) == "list" && all(apply(dataframe, 2, typeof) %in% types)
    ```
  - I suggest renaming this function to `is_valid_rhud_df()`. It would make it more human-readable in unit tests.
- `numbers_only()` might more aptly be named `digits_only`, since it will return `FALSE` for inputs such as `1.6` and `1.0`.
- You have a lot of repetition in the form of `if (to_tibble == FALSE) {`. I recommend writing a helper function `format_df(x, to_tibble = to_tibble))` that you can call instead of needing the control flow (it would just return a dataframe if `to_tibble = FALSE`). This would make it more readable and less duplicative, and you might be able to logically fit the `as.data.frame(do.call("rbind", list_res))` functionality there too.
- `hud_state_metropolitan`:
  - I find [this block](https://github.com/etam4260/rhud/blob/84c8d5b978decea4e8e45a8070545f171293207f/R/hudmisc.R#L215-L229) difficult to follow. Any way you can improve the readability?
  - [Here](https://github.com/etam4260/rhud/blob/84c8d5b978decea4e8e45a8070545f171293207f/R/hudmisc.R#L207-L211) you can just iterate over state directly, instead of indexing with `i`
- `add_leading_zeros()`:
  - The roxygen2 examples mistakenly still have `hud_*()`. Same goes for  `remove_leading_zeros()`. Running `devtools::run_examples(run_dontrun = TRUE)` can help identify cases like these.
  - I would be explicit in the `@returns` that this will cast input to a character vector. Same goes for  `remove_leading_zeros()`.
  - This function can be refactored for greater readability. I would suggest something along the lines of:
    ```r
    assert_geoid_type(geoid_type) # <- let this raise the error
    num_char <- switch(
        geoid_type,
        "countysub" =10,
        "cd" = 4,
        "tract" = 11,
        5
    )
    fix_geoid(input, num_char = num_char)
    ```
- `fmr_il_input_check_cleansing()`:
  - I believe you don't need `stop(paste(..., sep = ""))` and can directly do `stop(...)`. Applies to other areas in the package too.
- `capitalize()`:
  - This function is always called with the arguments `tolower(state)`. It would be more direct to refactor this with a function `extract_state()` that does both steps in one go, and get rid of `capitalize()`.
- `crosswalk_a_dataset_input_check_cleansing()`:
  - I think it is safer to return a named list and not rely on positional indices.
  - This function might benefit from renaming. I would suggest something along the lines of `cw_data_validate()`.
- `fix_geoid()`:
  - The function says that the first argument can be a column. IMO it would be cleaners if it only accepted vectors.
  - I suggest renaming this to something like `left_pad_zeros()` (as it is actually `geoid`-agnostic) and rename `add_leading_zeros()` to some name with `geoid` in it as it is specific to geoids.
  - I think this function can be refactored using functional programming principles. Alternatively, if you're willing to accept an import, you could use `stringi::stri_pad_left(geoids, num_char, 0)`.
- `misc_do_query_call()`:
  - You repeat `add_headers()` quite a few times, in other places, so I think a helper function to do this step would make sense.
- `chas_do_query_calls()`:
  - You only use `%notin%` infix once, so it might not be worth the trouble of making a function for it, and instead go for the vanilla `!x %in% y`.
  - I would suggest pulling `all_measurements` into a function that simply returns this vector, or use internal data to store a dictionary as a list.
  - I don't think the lone `message("\n")` is ideal. Can you tidy this up?
-  `rhud_website()`:
   -  This function could be [DRY](https://en.wikipedia.org/wiki/Don%27t_repeat_yourself)'d up a bit by simply using a `switch()` on the "start"/"open" piece.

Other code:
- Not all calls to `stop()` have `call. = FALSE` set. Is this intentional?
- This block gets repeated a lot:
    ```r
    if (!curl::has_internet()) stop("\nYou currently do not have internet access.", call. = FALSE)
    ```
  How about writing a helper `assert_internet()` to reduce duplication?

- You could write some helper functions to define the API host instead of repeating it: e.g., `hud_host <- function() "https://www.huduser.gov"`
- I don't think `library(rhud)` is necessary at the top of all the examples. Of course, it does not hurt to have it there, but I think most R developers omit this.

- `httr::content()` calls: Do we know if the APIs will always return UTF-8 content? If so, it would be good to be explicit.

- `error_urls <- c()` might be better as `error_urls <- NULL` or `error_urls <- character(0)`
- instead of using `paste()` to build query strings, I would use `GET`'s other named arguments and/or `modify_url()`. Something like this would be easier to follow and maintain:
    ```r
    config <- add_headers(Authorization = paste("Bearer ", as.character(key)))
    resp <- GET(
        url = "https://www.huduser.gov/hudapi/public/il",
        config =  config,
        path = ifelse(querytype == "state", "statedata", "data"),
        query = list(year = all_queries$year[i]),
        user_agent("https://github.com/etam4260/rhud"),
        timeout(30)
        )
    ```
- I would encourage the use of `tryCatch()` instead of `try()`.
-  `GET()`:
   - I would create a helper function to process the status `httr` reponses and handle different status codes, instead of `try(content(call), silent = TRUE)` and instead of `if ("error" %in% names(cont)) {`.
   - You can also use `httr::warn_for_status()` and `httr::stop_for_status()`.
- Instead of running `timeout(30)` everywhere, I suspect it might be possible to use rely on httr options for this. You may also be interested in `httr::RETRY()` and [this blog post](https://blog.r-hub.io/2020/04/07/retry-wheel/).
- It would be cleaner to set `to_tibble = getOption('rhud_use_tibbles', FALSE)` in the function's signature instead of having all your `missing()` checks in the function body.
- `download_bar()`:
  - Have you considered the [progress](https://github.com/r-lib/progress) package instead of writing this functionality from scratch? It would add a light dependency, but since there are still some `TODO`s in your existing function, it might be worth it.

- Some functions are defined but never used:
  - `decimal_num()`
  - `is.negative()`
  - `fivenumbers()` (also, this function does the opposite of what I expected. Would rename it to `not_*()` if you keep it)
  - `fivenumsthenfournums()`
  - `elevennumbers()`
  - `fournumbers()`
  - `tennumbers()`
  - `hud_get_key()`: Used in examples/vignettes but not in other functions. Probably should be used wherever you have `Sys.getenv("HUD_KEY")`.
- `add_delimiters()`: Is an empty function.
- Some functions have commented code that probably should be addressed/removed before finalization (e.g., `setup.R`)
- There are some remaining style issues:
  - Usage of `<-` instead of `=`
  - Usage of `vapply` instead of `sapply()`
  - Reducing line length to have fewer than 80 chars per line where practical.
  - Using the pattern `if (!to_tibble)` instead of `if (to_tibble == FALSE)` (and analogously for `var == TRUE` expressions).
  - The tidyverse style is to [not use `return()` explicitly at the end of a function](https://style.tidyverse.org/functions.html#return). I see a few places that do not follow that convention.

**Miscellaneous**

- I recommend adding the [spelling](https://github.com/ropensci/spelling) package to you check routine by running `spelling::spell_check_setup()`.
- For future development/maintenance, you may be interested in `fledge`. By using [conventional commits](https://www.conventionalcommits.org/), you can facilitate creation of new releases and changelogs. The package is still under development but I already use it for my own package development.
- You mention in your startup message that you are exploring caching options, yes? Have you considered using Neal Richardson `httpcache` package to cache httr's requests? I have never used it but I think it could be a valuable addition. I tried it by replacing `@import httr` with
  ```r
  #' @import httpcache
  #' @importFrom httr content add_headers timeout user_agent
  ```
  Once you make `httr` imports explicit, it seems to work well.
- An idea might be to write a `rhud_sitrep()` situation report function to print things such as:
  - if the API key is set (and recommend what do do if not),
  - if the API is reachable
    - (I could not find a health-check endpoint so maybe a lightweight API call would be needed?)
  - check remaining API bandwith (if API has such an endpoint),
  - display cache settings (whether or not enabled, and print the cache policy).

  See `usethis::git_sitrep()`, `usethis::proj_sitrep()`, `cli::cli_sitrep()`, and `devtools::dev_sitrep()` for examples.
- A future release could leverage usage of secrets/keyrings to store the API key. See https://httr.r-lib.org/articles/secrets.html
