### 2022-06-07 b7e00fa
Moved reference to hudpy to top of file. Added an example for using tibbles instead of dataframes.
