+++ README.Rmd
@@ -29,6 +29,8 @@ knitr::opts_chunk$set(
 <!-- badges: end -->
 
 
+Are you a python developer? Check out [hudpy](https://github.com/etam4260/hudpy)
+instead.
 
 ## Installation
 
@@ -64,9 +66,16 @@ hud_cw_zip_tract(zip = '35213', year = c('2010'), quarter = c('1'))
 
 ```
 
+## Tibbles vs Dataframes
+To get tibbles instead of data frames, use `options(rhud_use_tibble = TRUE)`
+or set it explicitly using the `to_tibble` argument.
+``` {r, eval = FALSE} 
+options(rhud_use_tibble = TRUE)
 
-Are you a python developer? Check out [hudpy](https://github.com/etam4260/hudpy)
-instead.
+hud_cw_zip_tract(zip = '35213', year = c('2010'), quarter = c('1'),
+                 to_tibble = TRUE)
+
+```
 
 ## Housing and Urban Development in R
 
