### 2022-06-07 486937b
Updating setup vignette to have examples interfacing with caching, getting tibbles, and setting a user agent. Converted all parameters field in vignettes to be in a tabular format. Added details on how to get the most recent year and quarters for files in vignettes. Added details on the infix operators and their non infix versions. The CHAS documentation now includes details on how to query for place and mcd. [no ci]
