+++ vignettes/Comprehensive-Housing-and-Affordability-Strategy.Rmd
@@ -41,42 +41,46 @@ different geographic resolutions. The package also contains an omni function
 which encapsulates thecapabilities of all the main function calls below --
 [omni-function][Using the omni function for querying]
 
-
-1) [nation][CHAS for Nation]
-2) [state][CHAS for State]
-3) [county][CHAS for County]
-4) [mcd][CHAS for MCD] 
-5) [place][CHAS for Place]
+1) [nation][Comprehensive Housing and Affordability Strategy for Nation]
+2) [state][Comprehensive Housing and Affordability Strategy for States]
+3) [county][Comprehensive Housing and Affordability Strategy for County]
+4) [mcd][Comprehensive Housing and Affordability Strategy for MCD] 
+5) [place][Comprehensive Housing and Affordability Strategy for Place]
 
 ### Comprehensive Housing and Affordability Strategy for Nation
 ```{r, eval=FALSE}
   library(rhud)
   hud_chas_nation(year=c("2014-2018"))
 ```
-
+<br>
 ### Comprehensive Housing and Affordability Strategy for States
+
 ```{r, eval=FALSE}
   # Queries for CHAS in a state
   hud_chas_state("Maryland", year=c("2012-2016"))
 ```
-
+<br>
 ### Comprehensive Housing and Affordability Strategy for County
+
 ```{r, eval=FALSE}
   # Queries a county in Virginia
   hud_chas_county("51999", year=c("2014-2018"))
 ```
-
+<br>
 ### Comprehensive Housing and Affordability Strategy for MCD
+
 ```{r, eval=FALSE}
   # Queries for CHAS for all mcds in Virginia
-  hud_chas_state_mcd("VA", "94087")
+  hud_chas_state_mcd("VA", year=c("2014-2018"))
 ```
-
+<br>
 ### Comprehensive Housing and Affordability Strategy for Place
+
 ```{r, eval=FALSE}
   # Queries for CHAS for all places in state fips code 51.
-  hud_chas_state_place("51", 48996)
+  hud_chas_state_place("51", year=c("2014-2018"))
 ```
+<br>
 
 ### Using the omni function for querying
 
@@ -93,6 +97,37 @@ details on how to use this.
   hud_chas(type = 1)
 
   hud_chas(type = 2, 51)
+```
+
+<br><br><br>
+
+### Querying for Geographic Identifers
+
+### US Counties 
+This is an example to show you how to query for all counties in MD,
+Virginia, and California, respectively.
+
+```{r, eval=FALSE}
+hud_state_counties("MD")
+
+hud_state_counties("Virginia")
+
+hud_state_counties("6")
+```
+<br>
+
+### US Minor Civil Divisions
+This is an example to show you how to query for minor civil divisions in
+Virginia and California.
+```{r, eval=FALSE}
+hud_state_minor_civil_divisions(c("VA", "CA"))
+```
+<br>
+
+### US Places
+This is an example to show you how to query for places in Wyoming and Michigan
+```{r, eval=FALSE}
+hud_state_places(c("VA", "CA"))
 ```
 
 <br><br><br>
