+++ vignettes/Covid-19-and-Evictions-Crosswalk.Rmd
@@ -7,147 +7,91 @@ vignette: >
   %\VignetteEncoding{UTF-8}
 ---
 
+<i>Sample provided by <b>Shuyu Jin (geojsy@umd.edu)</b></i>
+
+
 ```{r, include = FALSE}
 knitr::opts_chunk$set(
-  eval = FALSE,
   collapse = TRUE,
-  comment = "#>"
+  comment = "#>",
+  eval = FALSE
 )
 ```
 
-<i>Sample provided by <b>Shuyu Jin</b> (geojsy@umd.edu)</i>
+### Overview
+In the United States, housing advocates and policymakers are concerned that a sudden disruption in household finances caused by the economic crisis related to the COVID-19 pandemic will lead to a wave of evictions by tenants unable to pay their rents [(*Cowin, Martin, and Stevens, 2020*)](https://www.clevelandfed.org/en/newsroom-and-events/publications/community-development-briefs/db-20200717-measuring-evictions-during-the-covid-19-crisis.aspx). 
 
-## Overview
+Viewing [*2016 eviction rankings*](https://evictionlab.org/rankings/#/evictions?r=United%20States&a=0&d=evictionRate&lang=en) for locations across the U.S., Richmond has maintained a high eviction rate over the past few years. To understand how the pandemic is affecting evictions, I collected weekly eviction filing data for Richmond, which are provided by [*Princeton’s Eviction Lab*](https://evictionlab.org/eviction-tracking/). 
 
-In the United States, housing advocates and policymakers are concerned that
-a sudden disruption in household finances caused by the economic crisis related
-to the COVID-19 pandemic will lead to a wave of evictions by tenants unable
-to pay their rents [(*Cowin, Martin, and Stevens, 2020*)](https://www.clevelandfed.org/en/newsroom-and-events/publications/community-development-briefs/db-20200717-measuring-evictions-during-the-covid-19-crisis.aspx). This will
-certainly have a dramatic impact on the policies and funding allocated to 
-areas of the US.
+The mapping below uses ggplot and the crosswalk files to display COVID-19 cases as well as evictions based on zip codes and census tracts.
 
-Viewing [*2016 eviction rankings*](https://evictionlab.org/rankings/#/evictions?r=United%20States&a=0&d=evictionRate&lang=en) for locations across the U.S., Richmond has maintained a 
-high eviction rate over the past few years. To understand how the pandemic is 
-affecting evictions, I collected weekly eviction filing data for Richmond, which are
-provided by [*Princeton’s Eviction Lab*](https://evictionlab.org/eviction-tracking/). 
 
-The mapping below uses ggplot and the crosswalk files to 
-display COVID-19 cases as well as evictions based on zip codes and
-census tracts.
-
-<br>
-
-## Loading in the necessary libraries
-
+### Loading in the necessary libraries
 ```{r message=FALSE}
 library(rhud)
 library(sf)
 library(dplyr)
 library(scales)
 library(ggplot2)
-library(readr)
-library(lubridate)
-
 options (digits=20)
 ```
-
-<br>
-
 To improve performance, use tigris caching.
 
 ```{r message=FALSE}
 options(tigris_use_cache = TRUE)
 ```
 
-## Reading in data
+
+### Reading in data
 
 ```{r message=FALSE}
-
-# Need to move these into the data folder in the R package and 
[…]
