### 2022-06-09 #524 @etam4260
Hi @khueyama @rtaph

Thanks for all the feedback! Definitely a lot to go through and digest. I am hoping to talk to the other rhud package contributors to get some thoughts. I will try to get back in a few days.
