### 2022-06-09 c7547db
update eviction case study HTML
