### 2022-06-09 4585084
TODO added for rhud_website function. Fixed some typos in vignettes.
