### 2022-06-09 d6f08f8
Added dependency htmlwidgets to description file. Fixed issues with vignette indexing causing rmd check issues.
