+++ man/rhud_website.Rd
@@ -14,12 +14,11 @@ rhud_website(website = c("github-pages", "github"))
 }}
 }
 \description{
-Quickly get documentation for the hudr package by opening up
+Quickly get documentation for the rhud package by opening up
 the websites associated with it. Currently supports Unix and Windows OS.
 }
 \examples{
 \dontrun{
-library(rhud)
 rhud_website("github-pages")
 rhud_website("github")
 rhud_website()
