+++ DESCRIPTION
@@ -35,7 +35,8 @@ Suggests:
     ggplot2,
     readr,
     lubridate,
-    htmltools
+    htmltools,
+    htmlwidgets
 Imports: 
     httr,
     curl,
