### 2022-06-11 a7c8cd7
using browseURL instead of OS specificcommands for getting to the documentation website from R.[no ci]
