+++ R/hud_website_docs.R
@@ -18,18 +18,6 @@ rhud_website <- function(website = c("github-pages", "github")) {
   if(!curl::has_internet()) stop("You currently do not have internet access.")
   # Check for internet
 
-  github_pages <- "https://etam4260.github.io/rhud/"
-  github <- "https://github.com/etam4260/rhud"
-
-  # TODO: Change this to browseURL()
-
-  if (.Platform$OS.type == "unix") {
-    if ("github" %in% website) system(paste("open", github, sep = " "))
-    if ("github-pages" %in% website) system(paste("open", github_pages,
-                                                 sep = " "))
-  } else if (.Platform$OS.type == "windows") {
-    if ("github" %in% website) system(paste("start", github, sep = " "))
-    if ("github-pages" %in% website) system(paste("start", github_pages,
-                                                 sep = " "))
-  }
+  if ("github" %in% website) browseURL("https://github.com/etam4260/rhud")
+  if ("github-pages" %in% website) browseURL("https://etam4260.github.io/rhud/")
 }
