### 2022-06-11 b8815f8
Removing unused functions and files. Updated DESCRIPTION file to include utils package for browseURL function.[no ci]
