+++ R/hud_split_geoid.R
@@ -1,85 +1,0 @@
-# Auxiliary functions for splitting GEOIDS into their constituent parts
-# as per the US Census definition of them...
-
-#' @name split_zip
-#' @title split_zip
-#' @param zip A vector of zip codes
-#' @returns Returns a data frame containing the
-#'   decomposed elements of zip codes.
-#' @noRd
-#' @noMd
-split_zip <- function(zip) {
-
-}
-
-
-#' @name split_county
-#' @title split_county
-#' @param county A vector of county fip codes
-#' @returns Returns a data frame containing the
-#'   decomposed elements of county fip codes.
-#' @noRd
-#' @noMd
-split_county <- function(county) {
-
-}
-
-
-#' @name split_cbsa
-#' @title split_cbsa
-#' @param zip A vector of cbsa codes
-#' @returns Returns a data frame containing the
-#'   decomposed elements of cbsa codes.
-#' @noRd
-#' @noMd
-split_cbsa <- function(cbsa) {
-
-}
-
-
-#' @name split_cbsadiv
-#' @title split_cbsadiv
-#' @param zip A vector of cbsadiv codes
-#' @returns Returns a data frame containing the
-#'   decomposed elements of cbsadiv codes.
-#' @noRd
-#' @noMd
-split_cbsadiv <- function(cbsadiv) {
-
-}
-
-
-#' @name split_cd
-#' @title split_cd
-#' @param zip A vector of cd codes
-#' @returns Returns a data frame containing the
-#'   decomposed elements of a cd codes.
-#' @noRd
-#' @noMd
-split_cd <- function(cd) {
-
-}
-
-
-#' @name split_countysub
-#' @title split_countysub
-#' @param zip A vector of countysub codes
-#' @returns Returns a data frame containing the
-#'   decomposed elements of countysub codes.
-#' @noRd
-#' @noMd
-split_countysub <- function(countysub) {
-
-}
-
-
-#' @name split_tract
-#' @title split_tract
-#' @param zip A vector of tract codes
-#' @returns Returns a data frame containing the
-#'   decomposed elements of a tract code.
-#' @noRd
-#' @noMd
-split_tract <- function(tract) {
-
-}
