### 2022-06-15 407fc1f
Fixing examples in functions so that they should all pass. For cw from cbsadiv to zip or zip to cbsadiv, it seems that the HUD APIs are not working for these queries. Should contact HUD IT.
