+++ R/hud_omni_functions.R
@@ -89,10 +89,6 @@
 #' * [rhud::hud_cw()]
 #' @examples
 #' \dontrun{
-#' library(rhud)
-#'
-#' Sys.setenv("HUD_KEY" = "q3r2rjimd129fj121jid")
-#'
 #' hud_cw(type = 1, query = "35213", year = c("2010", "2011"),
 #'    quarter = c("1"))
 #'
@@ -118,7 +114,7 @@
 #'    quarter = c("1", "2"))
 #'
 #' hud_cw(type = 9, query = "10380", year = c("2017"),
-#'    quarter = c("1", "2", "3"))
+#'    quarter = c("4"))
 #'
 #' hud_cw(type = 10, query = "2202", year = c("2010", "2011"),
 #'    quarter = c("4", "3"))
@@ -284,9 +280,6 @@ hud_cw <- function(type, query,
 #'
 #' @examples
 #' \dontrun{
-#' library(rhud)
-#'
-#' Sys.setenv("HUD_KEY" = "q3r2rjimd129fj121jid")
 #'
 #' hud_fmr("VA", year=c(2021))
 #'
@@ -429,9 +422,6 @@ hud_fmr <- function(query, year = format(Sys.Date() - 365, "%Y"),
 #'
 #' @examples
 #' \dontrun{
-#' library(rhud)
-#'
-#' Sys.setenv("HUD_KEY" = "q3r2rjimd129fj121jid")
 #'
 #' hud_il("VA", year=c(2021))
 #'
@@ -576,10 +566,6 @@ hud_il <- function(query, year = format(Sys.Date() - 365, "%Y"),
 #'   https://www.huduser.gov/portal/dataset/chas-api.html
 #' @examples
 #' \dontrun{
-#' library(rhud)
-#'
-#' Sys.setenv("HUD_KEY" = "q3r2rjimd129fj121jid")
-#'
 #' hud_chas(1)
 #'
 #' hud_chas('2', state_id = '56')
