+++ R/hud_geo_is_infix_geo.R
@@ -251,6 +251,8 @@
 #' @examples
 #' \dontrun{
 #'
+#' # HUD API is broken for year 2021, quarter 3 for this query
+#'
 #' 71052 %z_in_cbsadiv% 43340
 #'
 #' }
@@ -697,6 +699,8 @@
 #' @export
 #' @examples
 #' \dontrun{
+#'
+#' # HUD API is broken for year 2021, quarter 3.
 #'
 #' 43340 %z_in_cbsadiv% 71052
 #'
