+++ man/hud_fmr_state_counties.Rd
@@ -36,10 +36,6 @@ at a county resolution for all counties in this state.
 }
 \examples{
 \dontrun{
-library(rhud)
-
-Sys.setenv("HUD_KEY" = "q3r2rjimd129fj121jid")
-
 hud_fmr_state_counties("VA", year = c(2021))
 
 hud_fmr_state_counties("Alabama", year = c(2021))

[…]
