+++ NAMESPACE
@@ -70,3 +70,4 @@ import(R.cache)
 import(httr)
 import(tibble)
 importFrom(curl,has_internet)
+importFrom(utils,browseURL)
