### 2022-06-15 d834da5
Updating readme to include zenodo badge. Adding update in NEWS.md about changes in function examples.[no ci]
