### 2022-06-15 2928816
shortening lines of code to 80 characters or less in R/ file. Renamed test files to not repeat test twice. (Fixing most issues found by lintr for the package) [no ci]
