+++ R/hud_geo_is_in_geo.R
@@ -38,11 +38,11 @@ z_in_trt <- function(zip, tract, year, quarter,
   if (missing(year) || missing(quarter)) {
     args <- hud_rec_cw_yr()
     if (missing(year)) {
-      year = args[1]
+      year <- args[1]
     }
 
     if (missing(quarter)) {
-      quarter = args[2]
+      quarter <- args[2]
     }
   }
 
@@ -64,7 +64,9 @@ z_in_trt <- function(zip, tract, year, quarter,
   quarter <- cleaned[3]
   key <- cleaned[4]
 
-  if (any(nchar(tract) != 11)) stop("\nInputted tract(s) are not all of length 11.")
+  if (any(nchar(tract) != 11)) {
+    stop("\nInputted tract(s) are not all of length 11.", call. = FALSE)
+  }
 
   res <- c()
 
@@ -132,11 +134,11 @@ z_in_cty <- function(zip, county, year, quarter,
   if (missing(year) || missing(quarter)) {
     args <- hud_rec_cw_yr()
     if (missing(year)) {
-      year = args[1]
+      year <- args[1]
     }
 
     if (missing(quarter)) {
-      quarter = args[2]
+      quarter <- args[2]
     }
   }
 
@@ -152,7 +154,9 @@ z_in_cty <- function(zip, county, year, quarter,
   quarter <- cleaned[3]
   key <- cleaned[4]
 
-  if (any(nchar(county) != 5)) stop("\nInputted county(s) are not all of length 5.")
+  if (any(nchar(county) != 5)) {
+    stop("\nInputted county(s) are not all of length 5.", call. = FALSE)
+  }
 
   res <- c()
   for (i in seq_len(length(zip))) {
@@ -220,11 +224,11 @@ z_in_cbsa <- function(zip, cbsa, year, quarter,
   if (missing(year) || missing(quarter)) {
     args <- hud_rec_cw_yr()
     if (missing(year)) {
-      year = args[1]
+      year <- args[1]
     }
 
     if (missing(quarter)) {
-      quarter = args[2]
+      quarter <- args[2]
     }
   }
 
@@ -240,7 +244,9 @@ z_in_cbsa <- function(zip, cbsa, year, quarter,
   quarter <- cleaned[3]
   key <- cleaned[4]
 
-  if (any(nchar(cbsa) != 5)) stop("\nInputted cbsa(s) are not all of length 5.")
+  if (any(nchar(cbsa) != 5)) {
+    stop("\nInputted cbsa(s) are not all of length 5.", call. = FALSE)
+  }
 
   res <- c()
   for (i in seq_len(length(zip))) {
@@ -309,11 +315,11 @@ z_in_cbsadiv <- function(zip, cbsadiv, year, quarter,
   if (missing(year) || missing(quarter)) {
     args <- hud_rec_cw_yr()
     if (missing(year)) {
-      year = args[1]
+      year <- args[1]
     }
 
     if (missing(quarter)) {
-      quarter = args[2]
+      quarter <- args[2]
     }
   }
 
@@ -329,7 +335,9 @@ z_in_cbsadiv <- function(zip, cbsadiv, year, quarter,
   quarter <- cleaned[3]
   key <- cleaned[4]
 
-  if (any(nchar(cbsadiv) != 5)) stop("\nInputted cbsadiv(s) are not all of length 5.")
+  if (any(nchar(cbsadiv) != 5)) {
+    stop("\nInputted cbsadiv(s) are not all of length 5.", call. = FALSE)
+  }
 
   res <- c()
   for (i in seq_len(length(zip))) {
@@ -383,7 +391,8 @@ z_in_cbsadiv <- function(zip, cbsadiv, year, quarter,
 #' * [rhud::cbsadiv_in_z()]
[…]
