+++ tests/testthat/test-cw_a_dataset.R
@@ -93,7 +93,7 @@ test_that("test simple crosswalk for zip", {
 })
 
 
-test_that("test awkward crosswalk",{
+test_that("test awkward crosswalk", {
   skip_if(Sys.getenv("HUD_KEY") == "")
 
   sample <- data.frame(mes = c(1232, 1232, 1232), zip = c(21206, 21206, 21206))

[…]
