### 2022-06-16 7c3a5d0
removing naked abbreviations in roxygen tags[no ci]
