### 2022-06-16 9da92ab
removing naked abbreviations in roxygen tags[no ci]
