### 2022-06-16 #524 @rtaph
Hi @etam4260

Great to see so much activity!

> I'm not exactly sure what else needs to go in the CONTRIBUTING file.

IMO the elements you have in the CONTRIBUTING file are sufficient. I think I just saw "This outlines how to propose changes to rhud..." and thought the ellipses indicated this might be an unfinished sentence.

>  I'm not sure what else would make the vignettes clearer

I have two suggestions:
1. Adding a few sentences up top to explain that rhud solves the task of converting data associated to ZIP codes to other well-known geographic regions such as counties.
2. Printing the output from the examples. At the moment, some do, and some don't.
