### 2022-06-16 629a2ed
Removing used cache testing. Adding skip tests when hud_key is missing for some tests.
