+++ R/zzz.R
@@ -60,16 +60,11 @@ NULL
   } else {
     packageStartupMessage(paste("\n* Setting rhud cache to",
                                 Sys.getenv("RHUD_CACHE_DIR"),
-                                "directory."))
+                                "directory.\n"))
   }
 }
 
 
-# When package is loaded, if the user had declared rhud_cache_dir, then
-# implicitly they would want caching removing the need to make use have to
-# turn it on and off. That function saves the variables that the user wants
-# as system variables then this uses it to initial the caching directory when
-# user loads the package.
 .onLoad <- function(libname, pkgname) {
   # Use evalWithMemoization to cache the data properly.
 
