+++ R/hud_loading_bar.R
@@ -1,11 +1,12 @@
 #' @name download_bar
-#' @title download_bar
+#' @title Query Download Bar
 #' @description A simple loading bar for the number of queries completed:
 #'   prints the loading bar to the R console.
-#' @param done The number of operations done.
-#' @param total The number of total operations needed.
-#' @param current The current item being worked on: in this case the url.
-#' @param error The errored out URLs.
+#' @param done A number: the number of operations.
+#' @param total A number: the number of total operations needed.
+#' @param current The current item being worked on: in most cases, the url
+#'   endpoint
+#' @param error A number: the total number of errorred out urls.
 #' @param percentage Can supply a percentage instead of done and total.
 #' @noRd
 #' @noMd
@@ -23,26 +24,11 @@ download_bar <- function(done = NULL, total = NULL,
     remain_bars <- paste(rep("-", floor(remain_perc * 50)), collapse = "")
   }
 
-  # TODO: For the num_done/total we might want it to show
-  # the number of successful queries instead... or maybe even show
-  # the errorred ones as a separate number.
-  # "Error: 6"
-
-  # TODO: Show the ETA for download time. (num api calls * 60)/2000
   url <- substr(current,
                 regexec("https://www.huduser.gov/hudapi/public/(.*)",
                         current)[[1]][2],
                 nchar(current)
                 )
-
-  # TODO: Need to add current working item when downloading. How best to show
-  # that?? Might show the last part of the url...
-
-  #[cw, firstgeoid, secondgeoid, year, quarter]
-  #[fmr, firstgeoid, secondgeoid, year]
-  #[il, firstgeoid, year]
-  #[chas, firstgeoid, year]
-  #[misc, firstgeoid, secondgeoid]
 
   # Create the entire loading bar
   loading <- paste("Downloading:\t",
