### 2022-06-17 4001e95
[no ci] Added more descriptive title roxygen tag for functions.
