+++ R/hud_check_for_numbers_only.R
@@ -1,8 +1,8 @@
 #' @name numbers_only
-#' @title numbers_only
-#' @description Detect whether string contains only numbers.
-#' @param x A string.
-#' @returns TRUE if string is all numbers, FALSE if not.
+#' @title All Numbers in Character vector
+#' @description Detect whether character vector contains only numbers.
+#' @param x A character vector.
+#' @returns TRUE if character vector is all numbers, FALSE if not.
 #' @noRd
 #' @noMd
 numbers_only <- function(x) !grepl("\\D", x)
