+++ R/hud_website_docs.R
@@ -2,7 +2,7 @@
 #' @importFrom utils browseURL
 #'
 #' @name rhud_website
-#' @title rhud_website
+#' @title Access RHUD Web Resources
 #' @description Quickly get documentation for the rhud package by opening up
 #'   the websites associated with it.
 #' @param website The websites available.
