+++ man/hud_get_cache_dir.Rd
@@ -2,7 +2,7 @@
 % Please edit documentation in R/hud_cache.R
 \name{hud_get_cache_dir}
 \alias{hud_get_cache_dir}
-\title{hud_get_cache_dir}
+\title{Get RHUD Cache Directory}
 \usage{
 hud_get_cache_dir()
 }

[…]
