+++ vignettes/Covid-19-and-Evictions-Crosswalk.Rmd
@@ -1,5 +1,5 @@
 ---
-title: "Covid-19-and-Evictions-Crosswalk"
+title: "Covid 19 and Evictions Crosswalk"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{Covid-19-and-Evictions-Crosswalk}
@@ -13,17 +13,17 @@ vignette: >
 ```{r, include = FALSE}
 knitr::opts_chunk$set(
   collapse = TRUE,
-  comment = "#>",
-  eval = FALSE
+  comment = "#>"
 )
 ```
 
 ### Overview
-In the United States, housing advocates and policymakers are concerned that a sudden disruption in household finances caused by the economic crisis related to the COVID-19 pandemic will lead to a wave of evictions by tenants unable to pay their rents [(*Cowin, Martin, and Stevens, 2020*)](https://www.clevelandfed.org/en/newsroom-and-events/publications/community-development-briefs/db-20200717-measuring-evictions-during-the-covid-19-crisis.aspx). 
 
-Viewing [*2016 eviction rankings*](https://evictionlab.org/rankings/#/evictions?r=United%20States&a=0&d=evictionRate&lang=en) for locations across the U.S., Richmond has maintained a high eviction rate over the past few years. To understand how the pandemic is affecting evictions, I collected weekly eviction filing data for Richmond, which are provided by [*Princeton’s Eviction Lab*](https://evictionlab.org/eviction-tracking/). 
+ZIP Codes are widely used for delivering service, mapping and other spatial analyses. However, there is a problematic effect since ZIP Codes’ boundaries are not created for analytical purposes. This effect is notorious for distorting policy-related analysis and misrepresentative results through inaccurate patterns displayed on the map [(Wilson and Din, 2018)](https://www.huduser.gov/portal/periodicals/cityscpe/vol20num2/ch16.pdf).
 
-The mapping below uses ggplot and the crosswalk files to display COVID-19 cases as well as evictions based on zip codes and census tracts.
+The HUD ZIP Code crosswalk files aim to allocate address locations from ZIP Codes to other geographies or vice versa. Specifically, merging aggregated data at the ZIP Code level with existing data sets to get more informative data for analysis [(Din and Wilson, 2020)](https://www.huduser.gov/portal/periodicals/cityscpe/vol22num1/ch12.pdf). The following example demonstrates how to use the crosswalk function to display COVID-19 cases and evictions based on zip codes and census tracts. 
+
+Weekly eviction filing data for Richmond, provided by [Princeton’s Eviction Lab](https://evictionlab.org/eviction-tracking/), and the COVID-19 data set, downloaded from [Virginia Open Data Portal](https://data.virginia.gov/Government/VDH-COVID-19-PublicUseDataset-ZIPCode/8bkr-zfqv/data).
 
 
 ### Loading in the necessary libraries
@@ -50,8 +50,8 @@ richmond_monthly_2020_2021 <- read_csv("Data/richmond_monthly_2020_2021.csv")
 VA_COVID_19_PublicUseDataset_ZIPCode <- read_csv("Data/VA-COVID-19-PublicUseDataset-ZIPCode.csv")
 
 # Using tigris package to download Richmond shapefiles
-richmond_shapefile_zctas = tigris::zctas(cb = FALSE, starts_with = "23")
-richmond_shapefile_tracts = tigris::tracts(state = "51", county = "760", cb = TRUE)
+richmond_shapefile_zctas <- tigris::zctas(cb = FALSE, starts_with = "23")
[…]
