+++ vignettes/Visualizing-FMR-and-IL.Rmd
@@ -1,11 +1,12 @@
 ---
-title: "Visualizing-FMR-and-IL"
+title: "Visualizing Fair Market Rents and Income Limits"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{Comparing-FMR-and-IL}
+  %\VignetteIndexEntry{Visualizing-FMR-and-IL}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
+
 <i>Sample provided by <b>Shuyu Jin (geojsy@umd.edu)</b></i>
 
 ```{r, include = FALSE}
@@ -16,7 +17,7 @@ knitr::opts_chunk$set(
 ```
 
 ## Overview
-The US Department of Housing and Urban Development maintains both [*FMR*](https://www.hud.gov/program_offices/public_indian_housing/programs/hcv/landlord/fmr) (Fair Markets Rent) and [*IL*](https://www.huduser.gov/portal/datasets/il.html) (Income Limits) programs which are crucial for determining eligibility for assistance dollars in the US. Provided below is an example of querying both FMR and IL as well as mapping them using the leaflet R library.
+The US Department of Housing and Urban Development maintains both [Fair Market Rents](https://www.hud.gov/program_offices/public_indian_housing/programs/hcv/landlord/fmr) (FMR) and [Income Limits](https://www.huduser.gov/portal/datasets/il.html) (IL) programs which are crucial for determining eligibility for assistance dollars in the US. Provided below is an example of querying both FMR and IL as well as mapping them using the leaflet R library.
 
 
 ## Loading in the necessary libraries
@@ -41,29 +42,29 @@ Read in the Fair Markets Rent for counties in the year 2021 as well as the shape
 files needed to map.
 
 ```{r message=FALSE}
-md_counties_2021_fmr = hud_fmr_state_counties(state = 'MD', year = '2021') 
+md_counties_2021_fmr <- hud_fmr_state_counties(state = 'MD', year = '2021') 
 ```
 
 Get county data for the state of Maryland.
 
 ```{r message=FALSE}
-md_counties = hud_state_counties("MD")
+md_counties <- hud_state_counties("MD")
 ```
 
 Query for every county and get their Income Limits data using the rhud package.
 
 ```{r message=FALSE}
-md_counties_2021_il = hud_il(md_counties$fips_code)
+md_counties_2021_il <- hud_il(md_counties$fips_code)
 
 #Find unique column names and keep only unique column names
-unique_names = unique(colnames(md_counties_2021_il))
-md_counties_2021_il_unique = md_counties_2021_il[unique_names]
+unique_names <- unique(colnames(md_counties_2021_il))
+md_counties_2021_il_unique <- md_counties_2021_il[unique_names]
 ```
 
 Download the shape files for counties in the US and obtain consistent projection.
 
 ```{r message=FALSE}
-county_shapefile = tigris::counties(state = NULL, cb = FALSE,
+county_shapefile <- tigris::counties(state = NULL, cb = FALSE,
                                        resolution = "500k", year = 2021) %>% sf::st_transform('+proj=longlat +datum=WGS84')
 
 ```
@@ -71,9 +72,9 @@ county_shapefile = tigris::counties(state = NULL, cb = FALSE,
 Filter the counties to those only in Maryland.
 
 ```{r message=FALSE}
-md_county_shapefile = county_shapefile[county_shapefile$STATEFP == 24, ]
+md_county_shapefile <- county_shapefile[county_shapefile$STATEFP == 24, ]
 
-md_county_shapefile$fullFips = paste0(as.character(md_county_shapefile$STATEFP),
[…]
