+++ vignettes/Setup.Rmd
@@ -55,9 +55,10 @@ Sys.getenv("HUD_KEY")
 
 To set the key...
 
+You need to go to https://www.huduser.gov/hudapi/public/register?comingfrom=1 and register for a key and replace it here.
+
 ```{r, eval=FALSE}
-Sys.setenv("HUD_KEY" = "2yeuduhq72ueajk")
-# "2yeuduhq72ueajk" is just a sample key. You need to go to https://www.huduser.gov/hudapi/public/register?comingfrom=1 and register for a key and replace it here.
+Sys.setenv("HUD_KEY" = "YOUR KEY HERE")
 Sys.getenv("HUD_KEY")
 ```
 
@@ -67,7 +68,7 @@ intuitive. You can also tell RStudio to remember the key by setting `in_wkdir` a
 .Rprofile.
 
 ```{r, eval=FALSE}
-hud_set_key("2yeuduhq72ueajk", in_home = TRUE)
+hud_set_key("OUR KEY HERE", in_home = TRUE)
 ```
 
 To check whether rhud can gain access to this environment variable...
@@ -147,30 +148,30 @@ these function calls should easily help you determine what you are querying
 and what data is returned.
 
 The general syntax for these functions within the package follow this pattern
-where {1},{2}, and {3} are placeholders:
+where {dataset},{geography}, and {resolution} are placeholders:
 
 
-hud\_{1}\_{2}\_{3}
+hud\_{dataset}\_{geography}\_{resolution}
 
 
-{1}: This symbol represents the dataset to query for. So for example if you want
+{dataset}: This symbol represents the dataset to query for. So for example if you want
 fair markets rent data, this will be 'fmr'. If you want data from the crosswalk
 files, this will be 'cw'.
 
-{2}: This symbol represents the geographic resolution to query for. 
+{geography}: This symbol represents the geographic resolution to query for. 
 For example, if the function requires that you input state(s), then this 
 will be named as 'state.' If the function requires that you input
 county(s) then this will be 'county'.
 
-{3}: This symbol represented the output geographic resolution. So for example,
+{resolution}: This symbol represented the output geographic resolution. So for example,
 if we want data at a zip code level, then this will have 'zip'.
 
-For those that do not have the {2} and {3}, these are omni functions which 
-are capable of performing all queries under {1}, but these functions are
+For those that do not have the {geography} and {resolution}, these are omni functions which 
+are capable of performing all queries under {dataset}, but these functions are
 harder to understand.
 
-For those that do not have the {3}, it inherits the {2}. This means that 
-if {2} is a state input, then the data described by the output data is also
+For those that do not have the {resolution}, it inherits the {geography}. This means that 
+if {geography} is a state input, then the data described by the output data is also
 a state.
 
 A quick example:
