### 2022-07-20 4d4c7ce
Formatting the entire website: Reorganziced the homepage
