+++ vignettes/USPS-Crosswalk.Rmd
@@ -14,7 +14,7 @@ knitr::opts_chunk$set(
 )
 ```
 
-## Program Overview
+### Program Overview
 
 According to (US Department of Housing and Urban Development, n.d.), the
 difficulty of linking United States Postal Service (USPS) ZIP codes to Census
@@ -55,20 +55,21 @@ These journal articles describe the problem and proposed solution in more detail
 
 <br>
 
-## Census Geographies
+### Census Geographies
 
 This chart provided by the US Census Bureau gives a good example of the
 relationships among different geographies. The crosswalk files only support a
 subset of these.
 
 ```{r, echo=FALSE, fig.cap="Hierarchy of Census Geographic Entities", out.width = '100%'}
-knitr::include_graphics("./Hierarchy_of_Census_Geographic_Entities.001.jpeg")
+# knitr::include_graphics("./Hierarchy_of_Census_Geographic_Entities.001.jpeg")
 ```
 
+# <img src='./Hierarchy_of_Census_Geographic_Entities.001.jpeg' align="center" width = "100%"/>
 
 <br>
 
-## Examples
+### Examples
 
 There are 12 main function calls for the crosswalk files:
 the package also contains an omni function which encapsulates the 
@@ -110,7 +111,7 @@ These are basic examples which shows you to query the Crosswalk API. Before
 looking at the outputted data I RECOMMEND first taking a look at [the parameters][Parameters]
 as well as [return data][Returns] located at the bottom of the page.
 
-### Crosswalk zipcode to census tract
+#### Crosswalk zipcode to census tract
 ```{r, eval = if(Sys.getenv("HUD_KEY") == "") FALSE else TRUE}
 library(rhud)
 options (digits=4)
@@ -119,67 +120,67 @@ hud_cw_zip_tract(zip = '35213', year = c('2010'), quarter = c('1'))
 ```
 <br>
 
-### Crosswalk zipcode to county fip
+#### Crosswalk zipcode to county fip
 ```{r, eval = if(Sys.getenv("HUD_KEY") == "") FALSE else TRUE}
 hud_cw_zip_county(zip = 35213, year = c('2020'), quarter = c('2'))
 ```
 <br>
 
-### Crosswalk zipcode to core base statistical area (cbsa)
+#### Crosswalk zipcode to core base statistical area (cbsa)
 ```{r, eval = if(Sys.getenv("HUD_KEY") == "")  FALSE else TRUE}
 hud_cw_zip_cbsa(zip = 35213, year = c('2011'), quarter = c('3'))
 ```
 <br>
 
-### Crosswalk zipcode to core based statistical area division (cbsadiv)
+#### Crosswalk zipcode to core based statistical area division (cbsadiv)
 ```{r, eval = if(Sys.getenv("HUD_KEY") == "")  FALSE else TRUE}
 hud_cw_zip_cbsadiv(zip = '22031', year = c('2019'), quarter = c('4'))
 ```
 <br>
 
-### Crosswalk zipcode to congressional district (cd)
+#### Crosswalk zipcode to congressional district (cd)
 ```{r, eval = if(Sys.getenv("HUD_KEY") == "")  FALSE else TRUE}
 hud_cw_zip_cd(zip = '35213', year = c(2011), quarter = c(1))
 ```
 <br>
 
-### Crosswalk census tract to zipcode
+#### Crosswalk census tract to zipcode
 ```{r, eval = if(Sys.getenv("HUD_KEY") == "")  FALSE else TRUE}
 hud_cw_tract_zip(tract = 48201223100, year = c('2017'), quarter = c('1'))
 ```
 <br>
 
-### Crosswalk county fip into zipcode 
+#### Crosswalk county fip into zipcode 
[…]
