+++ R/hud_do_query_calls.R
@@ -8,73 +8,56 @@
 #' @description Helper function for making the query calls to
 #'   Comprehensive Housing and Affordability Strategy (CHAS)
 #'   API endpoint as well as concatenating all response objects to
-#'   be returned to the user.
-#' @param urls The urls to query for.
-#' @param key The key obtained from HUD
+#'   be returned to the user. Collect error urls and warn users and cache the
+#'   queries and show the download bar.
+#' @param urls A character vector: the urls to query for.
+#' @param key A character vector of length one with the key obtained from HUD
 #'   (US Department of Housing and Urban Development)
 #'   USER website.
-#' @param to_tibble If TRUE, return as tibble else dataframe.
+#' @param to_tibble A logical: if TRUE, return as tibble else dataframe.
 #' @returns A tibble or dataframe of all response bodies.
 #' @noRd
 #' @noMd
 chas_do_query_calls <- function(urls, key, to_tibble) {
   # Form all query calls...
   list_res <- c()
-  error_urls <- c()
-
-  `%notin%` <- Negate(`%in%`)
+  error_urls <- character(0)
 
   # These measurements are hardcoded in, but a more effective method might
-  # be to systematic checks to find all unique columns names from all CHAS
+  # be systematic checks to find all unique columns names from all CHAS
   # datasets.
-  all_measurements <- c("geoname", "sumlevel", "year",
-                        "A1", "A2", "A3", "A4", "A5", "A6", "A7", "A8", "A9",
-                        "A10", "A11", "A12", "A13", "A14", "A15", "A16", "A17",
-                        "A18",
-                        "B1", "B2", "B3", "B4", "B5", "B6", "B7",
-                        "B8", "B9",
-                        "C1", "C2", "C3", "C4", "C5", "C6",
-                        "D1", "D2", "D3",
-                        "D4", "D5", "D6", "D7", "D8", "D9", "D10", "D11", "D12",
-                        "E1", "E2", "E3", "E5", "E6", "E7", "E9", "E10", "E11",
-                        "E13", "E14", "E15", "E17", "E18", "E19", "E21", "E22",
-                        "E23",
-                        "F1", "F2", "F3", "F5", "F6", "F7", "F9", "F10", "F11",
-                        "F13", "F14", "F15", "F17", "F18", "F19", "F21", "F22",
-                        "F23",
-                        "G1", "G2", "G3", "G5", "G6", "G7", "G9", "G10", "G11",
-                        "G13", "G14", "G15", "G17", "G18", "G19",
-                        "H1", "H2", "H4", "H5", "H7", "H8", "H10", "H11", "H13",
-                        "H14", "H16",
-                        "I1", "I2", "I4", "I5", "I7", "I8", "I10", "I11",
-                        "I13", "I14", "I16",
-                        "J1", "J2", "J4", "J5", "J7", "J8", "J10",
-                        "J11", "J13", "J14", "J16")
 
+  all_measurements <- hud_chas_col_names()
 
   for (i in seq_len(length(urls))) {
 
     url <- urls[i]
+
     call <- R.cache::memoizedCall(make_query_calls, url, key)
+    processed_code <- process_status_codes(call)
 
-    cont <- try(content(call), silent = TRUE)
+    if (!is.null(processed_code)) {
 
-    if ("error" %in% names(cont) || length(cont) == 0) {
-      # Need to output a single error message instead of a bunch when
-      # something bad occurs. Append to list of errored urlss.
-      error_urls <- c(error_urls, url)
+      error_urls <- list(c(processed_code, url))
+
     } else {
-      not_measured <- all_measurements[all_measurements %notin%
+      cont <- try(content(call), silent = TRUE)
+
+      not_measured <- all_measurements[!all_measurements %in%
                                          names(unlist(cont[[1]]))]
[…]
