+++ tests/testthat/test-fmr.R
@@ -1,5 +1,6 @@
-test_that("Fair Markets Rent State Queries", {
-  skip_if(Sys.getenv("HUD_KEY") == "")
+test_that("hud_fmr() Simple State Queries", {
+  skip_if_no_key()
+
   va <- hud_fmr("VA", year = c(2021))
   expect_true(length(va) >= 1)
   md <- hud_fmr("MD", year = c("2021"))
@@ -10,8 +11,9 @@ test_that("Fair Markets Rent State Queries", {
   expect_true(length(al) >= 1)
 })
 
-test_that("Fair Markets Rent County Queries", {
-  skip_if(Sys.getenv("HUD_KEY") == "")
+test_that("hud_fmr() Simple County Queries", {
+  skip_if_no_key()
+
   c1 <- hud_fmr("5100199999", year = c(2021))
   expect_true(nrow(c1) >= 1)
   c2 <- hud_fmr("5100199999", year = c("2021"))
@@ -27,8 +29,9 @@ test_that("Fair Markets Rent County Queries", {
 })
 
 
-test_that("Small Area FMR Queries", {
-  skip_if(Sys.getenv("HUD_KEY") == "")
+test_that("hud_fmr() Simple Metroarea Queries", {
+  skip_if_no_key()
+
   sa1 <- hud_fmr("METRO47900M47900", year = c(2018))
   expect_true(nrow(sa1) >= 1)
   sa2 <- hud_fmr("METRO29180N22001", year = c(2019))
@@ -42,8 +45,9 @@ test_that("Small Area FMR Queries", {
 })
 
 
-test_that("Fair Markets Rent Different Years", {
-  skip_if(Sys.getenv("HUD_KEY") == "")
+test_that("hud_fmr() Different Years Queries", {
+  skip_if_no_key()
+
   y1 <- hud_fmr("AL", year = c(2020))
   expect_true(length(y1) >= 1)
   y2 <- hud_fmr("AL", year = c(2021, 2020, 2019, 2018))
