+++ R/zzz.R
@@ -27,40 +27,47 @@ NULL
 # to set the default year, quarter, and whether they want dataframes or
 # tibbles when querying data.
 .onAttach <- function(libname, pkgname) {
-  packageStartupMessage(paste("\nrhud:\n",
-                              "---------------------------------------------",
-                              "---------------------------------------------",
-                              "\n* To begin, first obtain a key ",
-                              "from HUD User website at `https://www.huduser.",
-                              "gov/hudapi/public/login`\n",
-                              "* To set the key, use hud_set_key",
-                              "(\"your-key\")\n",
-                              "* To get tibbles instead of dataframes, ",
-                              "use `options(rhud_use_tibble = TRUE)`\n",
-                              "* Caching is enabled by default. To set a new ",
-                              "cache ",
-                              "directory use hud_set_cache_dir().\n",
-                              "* To delete cached ",
-                              "items, use hud_clear_cache()\n",
-                              "* Set these in your R script for single ",
-                              "session ",
-                              "or in .Rprofile for persistence.\n",
-                              "* For more information on how to use this ",
-                              "package, ",
-                              "visit `https://etam4260.github.io/rhud/",
-                              "index.html`\n",
-                              "--------------------------------------------",
-                              "--------------------------------------------",
-                              sep = ""))
 
-  if (Sys.getenv("RHUD_CACHE_DIR") == "" ||
-      Sys.getenv("RHUD_CACHE_DIR") == "NULL") {
-    packageStartupMessage("\n* Setting rhud cache to R temp directory.")
+  # If HUD_KEY is already set, then they likely don't need the start up message
 
-  } else {
-    packageStartupMessage(paste("\n* Setting rhud cache to",
-                                Sys.getenv("RHUD_CACHE_DIR"),
-                                "directory.\n"))
+  if (Sys.getenv("HUD_KEY") == "") {
+    packageStartupMessage(paste("\nrhud:\n",
+                                "---------------------------------------------",
+                                "---------------------------------------------",
+                                "\n* To begin, first obtain a key ",
+                                "from HUD User website at `https://www.huduser.",
+                                "gov/hudapi/public/login`\n",
+                                "* To set the key, use hud_set_key",
+                                "(\"your-key\")\n",
+                                "* To get tibbles instead of dataframes, ",
+                                "use `options(rhud_use_tibble = TRUE)`\n",
+                                "* To silence messages from downloads, use",
+                                " `options(rhud_quiet_loads = TRUE)`\n",
+                                "* Caching is enabled by default. To set a new ",
+                                "cache ",
+                                "directory use hud_set_cache_dir().\n",
+                                "* To delete cached ",
+                                "items, use hud_clear_cache()\n",
+                                "* Set these in your R script for single ",
+                                "session ",
+                                "or in .Rprofile for persistence.\n",
+                                "* For more information on how to use this ",
+                                "package, ",
+                                "visit `https://etam4260.github.io/rhud/",
+                                "index.html`\n",
+                                "--------------------------------------------",
+                                "--------------------------------------------",
+                                sep = ""))
+
+    if (Sys.getenv("RHUD_CACHE_DIR") == "" ||
+        Sys.getenv("RHUD_CACHE_DIR") == "NULL") {
+      packageStartupMessage("\n* Setting rhud cache to R temp directory.")
+
+    } else {
+      packageStartupMessage(paste("\n* Setting rhud cache to",
+                                  Sys.getenv("RHUD_CACHE_DIR"),
[…]
