+++ tests/testthat/test-key.R
@@ -1,3 +1,0 @@
-test_that("test", {
-  expect_equal(2 * 2, 4)
-})

[…]
