+++ vignettes/Fair-Markets-Rent.Rmd
@@ -1,294 +1,0 @@
----
-title: "Fair Market Rents"
-output: rmarkdown::html_vignette
-vignette: >
-  %\VignetteIndexEntry{Fair-Market-Rents}
-  %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
----
-
-```{r, include = FALSE}
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>"
-)
-```
-
-### Program Overview
-
-According to (Fair Market Rents | HUD USER), Fair Market Rents (FMRs) are used
-to calculate payment standard amounts for the Housing Choice Voucher program,
-initial renewal rents for some expiring project-based Section 8 contracts,
-initial rents for housing assistance payment (HAP) contracts in the Moderate
-Rehabilitation Single Room Occupancy program (Mod Rehab), rent ceilings for
-rental units in both the HOME Investment Partnerships program and the Emergency
-Solution Grants program, and rent ceilings for rental units in both the HOME
-Investment Partnerships program and the Emergency Solution Grant.
-
-
-# <img src='./fmr_pic.jpg' href="https://creativecommons.org/licenses/by/2.0/" title="CC BY 2.0" align="right" width="40%"/>
-
-
-Furthermore, the United States Department of Housing and Urban Development (HUD)
-calculates FMRs for metropolitan areas defined by the Office of Management and
-Budget (OMB), some HUD-designated subdivisions of OMB metropolitan areas, and
-each non-metropolitan county on an annual basis. FMRs must be posted at least 30
-days before they go into effect, and they must go into force at the beginning of
-the federal fiscal year (usually October 1). Fair Market Rentals are estimates
-of the 40th percentile gross rents for standard quality units within a
-metropolitan region or non-metropolitan county, as described in 24 CFR 888.113.
-
-
-### Examples
-
-These are basic examples to show you how to query the Fair Markets Rent API. Before
-looking at the examples I RECOMMEND first looking at [the parameters][Parameters]
-as well as [return data][Returns] located at the bottom of the page.
-
-**Disclaimer:** The output tables are too large to be displayed here so we
-choose to leave them out.
-
-Querying a county or cbsa code that are considered small areas will 
-return data at a zip code level. If they are not small areas, then it will 
-give a singular measurement for that entire county or cbsa. 
-
-Lets say we are working with county level data -- when mixing them together 
-(counties that are small areas and not small areas) those counties that are
-not small areas will have an NA in the zip_code, 
-meaning this measurement is defined at a county level. Those counties that 
-are small areas will return zip code data in the zip_code field.
-
-#### State Level Fair Markets Rent
-This is an example which shows you how to query the Fair Markets Rent API for a
-state. You can use the state abbreviation, state name, or state fips code. It
-will return all counties in the state and their Fair Markets Rents as well
-as all metroareas in the state and their Fair Markets Rent.
-```{r, eval=FALSE}
-# These functions gets FMR data for the state of Virginia in 2021.
-hud_fmr(query = 'VA', year = '2021')
-
-hud_fmr(query= "Virginia", year = '2021')
-
-hud_fmr(query= "51", year = '2021')
-```
[…]
