+++ vignettes/Visualizing-FMR-and-IL.Rmd
@@ -1,5 +1,5 @@
 ---
-title: "Visualizing Fair Market Rents and Income Limits"
+title: "Visualizing-FMR-and-IL"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{Visualizing-FMR-and-IL}
