### 2022-07-27 cadaff0
Add an index that lists all functions
