+++ vignettes/Fair-Market-Rents.Rmd
@@ -26,11 +26,9 @@ rental units in both the HOME Investment Partnerships program and the Emergency 
 Furthermore, the United States Department of Housing and Urban Development (HUD)
 calculates FMRs for metropolitan areas defined by the Office of Management and
 Budget (OMB), some HUD-designated subdivisions of OMB metropolitan areas, and
-each non-metropolitan county on an annual basis. FMRs must be posted at least 30
-days before they go into effect, and they must go into force at the beginning of
-the federal fiscal year (usually October 1). Fair Market Rentals are estimates
-of the 40th percentile gross rents for standard quality units within a
-metropolitan region or non-metropolitan county, as described in 24 CFR 888.113.
+each non-metropolitan county on an annual basis. 
+
+FMRs must be posted at least 30 days before they go into effect, and they must go into force at the beginning of the federal fiscal year (usually October 1). Fair Market Rentals are estimates of the 40th percentile gross rents for standard quality units within a metropolitan region or non-metropolitan county, as described in 24 CFR 888.113.
 
 
 ### Examples
