+++ _pkgdown.yml
@@ -1,3 +1,5 @@
+template:
+  bootstrap: 5
 navbar:
   structure:
     left:  [intro, setup, articles, casestudy, reference, changelog]
@@ -30,6 +32,9 @@ navbar:
    changelog:
     text: Changelog
     href: news/index.html
+   search:
+    exclude: ['news/index.html']
+
 
 reference:
 - title: All functions
