### 2022-07-27 98f3056
Add Bootstrap 5 template to built a search bar
