### 2022-07-28 #524 @etam4260
Hi @jhollist, I am still currently working on the changes -- mostly DRYing it up.  @rtaph as for the feedback that you mentioned below Other Code/Miscellaneous. I agree with most of it, at ~ maybe 75% done with those changes.
