+++ R/hud_omni_functions.R
@@ -177,12 +177,17 @@ hud_cw <- function(type, query,
   lhgeoid <- strsplit(alltypes[as.integer(type)], "-")[[1]][1]
   rhgeoid <- strsplit(alltypes[as.integer(type)], "-")[[1]][2]
 
-  args <- cw_input_check_cleansing(lhgeoid, rhgeoid, query, year, quarter, key)
+  args <- cw_input_check_cleansing(lhgeoid,
+                                   rhgeoid,
+                                   query,
+                                   year,
+                                   quarter,
+                                   key)
 
-  query <- args[[1]]
-  year <- args[[2]]
-  quarter <- args[[3]]
-  key <- args[[4]]
+  query <- args$query
+  year <- args$year
+  quarter <- args$query
+  key <- args$key
 
 
   # Need to make sure query is a zip code of 5 digits.
@@ -302,108 +307,18 @@ hud_fmr <- function(query, year = format(Sys.Date() - 365, "%Y"),
                     key = Sys.getenv("HUD_KEY"),
                     to_tibble = getOption("rhud_use_tibble", FALSE)) {
 
-  res <- NULL
   is_internet_available()
 
   args <- fmr_il_input_check_cleansing(query, year, key)
-  query <- args[[1]]
-  year <- args[[2]]
-  key <- args[[3]]
-  querytype <- args[[4]]
+  query <- args$query
+  year <- args$year
+  key <- args$key
+  querytype <- args$querytype
 
   allqueries <- expand.grid(query = query, year = year,
                             stringsAsFactors = FALSE)
 
-  fmr_do_query_call()
-
-
-  # If state query, will provide data at metro and county level
-  if (querytype == "state") {
-    # Merge county level data with metroarea data.
-    # Create all combinations of query and year...
-    allqueries <- expand.grid(query = query, year = year,
-                              stringsAsFactors = FALSE)
-    error_urls <- c()
-
-    list_county_res <- c()
-    list_metroarea_res <- c()
-
-    for (i in seq_len(nrow(allqueries))) {
-      # Build the urls for querying the data.
-
-      urls <- paste(get_hud_host_name(),
-                    "fmr/",
-                    "statedata/",
-                    allqueries$query[i], "?year=", allqueries$year[i], sep = "")
-
-      call <- R.cache::memoizedCall(make_query_calls, urls, key)
-
-      cont <- try(content(call), silent = TRUE)
-
-      if ("error" %in% names(cont)) {
-        error_urls <- c(error_urls, urls)
-      } else {
-        res_county <- as.data.frame(do.call(rbind, cont$data$counties))
-        res_metroareas <- as.data.frame(do.call(rbind, cont$data$metroareas))
-
-        res_county$query <- allqueries$query[i]
-        res_county$year <- allqueries$year[i]
-
-        res_metroareas$query <- allqueries$query[i]
-        res_metroareas$year <- allqueries$year[i]
-
-        list_county_res[[i]] <- res_county
-        list_metroarea_res[[i]] <- res_metroareas
-      }
-
-      download_bar(done = i, total = nrow(allqueries),
-                   current = urls, error = length(error_urls))
-    }
-
-
-    if (length(error_urls) != 0) {
-      # Spit out error messages to user after all
-      # queries are done.
-      warning(paste("Could not find data for queries: \n\n",
-                    paste(paste("*", error_urls, sep = " "), collapse = "\n"),
-                    "\n\nIt is possible that your key maybe invalid or ",
-                    "there isn't any data for these parameters, ",
-                    "If you think this is wrong please ",
-                    "report it at https://github.com/etam4260/rhud/issues.",
[…]
