+++ tests/testthat/test-il.R
@@ -11,10 +11,10 @@ test_that("hud_il() Income Limits Simple State Queries", {
   expect_true(nrow(al) == 1)
 
   # Check that columns returned are of type vectors
-  expect_true(check_is_not_list(va))
-  expect_true(check_is_not_list(md))
-  expect_true(check_is_not_list(ca))
-  expect_true(check_is_not_list(al))
+  expect_true(is_valid_rhud_df(va))
+  expect_true(is_valid_rhud_df(md))
+  expect_true(is_valid_rhud_df(ca))
+  expect_true(is_valid_rhud_df(al))
 
   # Garbage collection
   va <- NULL
@@ -25,18 +25,18 @@ test_that("hud_il() Income Limits Simple State Queries", {
   # Make multiple state queries
   mult_state <- hud_il(c("AL", "MD", "CA"), c(2018, 2017, 2019))
   expect_true(nrow(mult_state) == 9)
-  expect_true(check_is_not_list(mult_state))
+  expect_true(is_valid_rhud_df(mult_state))
 
   # Make multiple year queries
   mult_year <- hud_il(c("OH"), c(2020, 2019, 2018))
   expect_true(nrow(mult_year) == 3)
-  expect_true(check_is_not_list(mult_year))
+  expect_true(is_valid_rhud_df(mult_year))
 
   # Although HUD dataset classifies PR as puerto rico and a state:
   # it is not a state and the API doesn"t allow you to get data for this.
   # Same for DC. Should expect an error if trying to grab this data.
-  expect_error(hud_il("PR"))
-  expect_error(hud_il("DC"))
+  expect_warning(hud_il("PR"))
+  expect_warning(hud_il("DC"))
 
   # Try to query for all states... filter out US territories...
   hudstates <- hud_nation_states_territories()
@@ -45,17 +45,17 @@ test_that("hud_il() Income Limits Simple State Queries", {
 
   all_states_full <- hud_il(hudstates$state_name)
   expect_true(nrow(all_states_full) >= 1)
-  expect_true(check_is_not_list(all_states_full))
+  expect_true(is_valid_rhud_df(all_states_full))
 
   # Try to query for all state codes?
   all_state_abbr <- hud_il(hudstates$state_code)
   expect_true(nrow(all_state_abbr) >= 1)
-  expect_true(check_is_not_list(all_state_abbr))
+  expect_true(is_valid_rhud_df(all_state_abbr))
 
   # Try to query for all state name?
   all_state_num <- hud_il(hudstates$state_num)
   expect_true(nrow(all_state_num) >= 1)
-  expect_true(check_is_not_list(all_state_num))
+  expect_true(is_valid_rhud_df(all_state_num))
 
   # Check if they are identical according to R...
   expect_true(identical(all_states_full, all_state_abbr) &&
@@ -67,24 +67,24 @@ test_that("hud_il() Income Limits Simple County Queries", {
 
   c1 <- hud_il("5100199999", year = c(2021))
   expect_true(nrow(c1) == 1)
-  expect_true(check_is_not_list(c1))
+  expect_true(is_valid_rhud_df(c1))
   c1 <- NULL
 
   c2 <- hud_il("5100199999", year = c("2020"))
   expect_true(nrow(c2) == 1)
-  expect_true(check_is_not_list(c2))
+  expect_true(is_valid_rhud_df(c2))
   c2 <- NULL
 
   c3 <- hud_il("5151099999", year = c(2019))
   expect_true(nrow(c3) == 1)
[…]
