+++ tests/testthat/test-decomp_chas.R
@@ -6,12 +6,12 @@ test_that("hud_chas_nation() Simple Examples and Errors", {
   one_year <- hud_chas_nation()
   # Just expect one row to return... weak check.
   expect_true(nrow(one_year) == 1)
-  expect_true(check_is_not_list(one_year))
+  expect_true(is_valid_rhud_df(one_year))
 
   # Try querying multiple years from the nations
   two_years <- hud_chas_nation(year = c("2014-2018", "2012-2016"))
   expect_true(nrow(two_years) == 2)
-  expect_true(check_is_not_list(two_years))
+  expect_true(is_valid_rhud_df(two_years))
 
   # Try querying a year which is not allowed... should expect warning
   expect_error(hud_chas_nation(year = c("2017-2022", "2321-142131")))
@@ -46,10 +46,7 @@ test_that("hud_chas_state() Simple Examples, All States, and Errors", {
   test <- hud_chas_state(state = "CA")
   expect_true(nrow(test) == 1)
 
-  test <- hud_chas_state(state = "VIRGINIA")
-  expect_true(nrow(test) == 1)
-
-  test <- hud_chas_state(state = "vIRGINIa")
+  test <- hud_chas_state(state = "Virginia")
   expect_true(nrow(test) == 1)
 
   # Try querying for all states in nation
@@ -60,12 +57,12 @@ test_that("hud_chas_state() Simple Examples, All States, and Errors", {
   # Try to query for all state codes?
   all_state_abbr <- hud_chas_state(hudstates$state_code)
   expect_true(nrow(all_state_abbr) >= 1)
-  expect_true(check_is_not_list(all_state_abbr))
+  expect_true(is_valid_rhud_df(all_state_abbr))
 
   # Try to query for all state name?
   all_state_num <- hud_chas_state(hudstates$state_num)
   expect_true(nrow(all_state_num) >= 1)
-  expect_true(check_is_not_list(all_state_num))
+  expect_true(is_valid_rhud_df(all_state_num))
 
   # Check if they are identical according to R...
   expect_true(identical(all_state_num, all_state_abbr))
@@ -96,7 +93,7 @@ test_that("hud_chas_county() Simple Examples, Maryland Counties, and Error", {
   # Only use the first 5 numbers in fips code.
   all_md <- hud_chas_county(county = substr(all_md_counties$fips_code, 0, 5))
   expect_true(nrow(all_md) >= 1)
-  expect_true(check_is_not_list(all_md))
+  expect_true(is_valid_rhud_df(all_md))
 })
 
 test_that("hud_chas_state_mcd() For MD", {
@@ -106,5 +103,5 @@ test_that("hud_chas_state_mcd() For MD", {
   all_md <- hud_chas_state_mcd("md")
 
   expect_true(nrow(all_md) >= 1)
-  expect_true(check_is_not_list(all_md))
+  expect_true(is_valid_rhud_df(all_md))
 })
