+++ _pkgdown.yml
@@ -1,3 +1,4 @@
+url: https://etam4260.github.io/rhud
 template:
   bootstrap: 5
 navbar:
