+++ README.Rmd
@@ -228,8 +228,11 @@ also provide miscellaneous supplemental APIs under them.
 ### Contributors
 
 * Emmet Tam(https://github.com/etam4260)[emmet_tam@yahoo.com]
+
 * Allison Reilly[areilly2@umd.edu]
+
 * Hamed Ghaedi[hghaedi@terpmail.umd.edu]
+
 * Shuyu Jin(https://github.com/geojsy)[geojsy@umd.edu]
 
 
