### 2022-08-24 53e995b
Adding testthat as a import dependency for the package. Updating work in progress items in (NEWS.md) Fixing issues with global variable naming and dependencies in R cmd check.
