+++ NAMESPACE
@@ -70,6 +70,7 @@ export(z_in_ctysb)
 export(z_in_trt)
 import(R.cache)
 import(httr)
+import(testthat)
 import(tibble)
 importFrom(curl,has_internet)
 importFrom(utils,browseURL)
