### 2022-08-24 #524 @etam4260
Hi, @rtaph @jhollist @khueyama
as a follow up, I think most of internal changes/DRYing is done (there is definitely more DRYING that can be done, such as the CORE api mentioned by @khueyama, but have not gotten around to implementing it), here are the few items left:

* [This blog](https://blog.r-hub.io/2020/06/03/vignettes/#how-to-include-a-compute-intensive--authentication-dependent-vignette) encourages shipping API-wrapping packages with precomputed vignettes. That might be something to consider to allow the package to build faster and without an internet connection and/or API key.

* I see that you have an empty test with_mock_dir(), so I suspect you are dabbling with webmocking and may already know about [this resource](https://books.ropensci.org/http-testing) by Scott Chamberlain and Maëlle Salmon. The book has fantastic guidance on how run some unit tests through moking. For example, vcr might be a great solution for testing the API without an internet connection, as it will snapshot real responses as test fixtures. If you feel even more adventurous, you could create a Github Workflow that tests live requests on a schedule (e.g. semi-monthly) using a [Github secret](https://docs.github.com/en/rest/actions/secrets) to store an API key

* In vignettes/Setup.Rmd, it could be useful to encourage users to set their environment key via usethis::edit_r_environ(). This would persist across sessions. Working with the renviron file seems cleaner than all the regex and string manipulation in hud_set_key(), though the key would be exposed to any package in R. In the medium term, it would be nice to also offer the option to use a keyring.

* Instead of using paste() to build query strings, I would use GET's other named arguments and/or modify_url(). Something like this would be easier to follow and maintain:

Furthermore, I know the use of RETRY was mentioned in one of the reviews. This is something getting worked on, but I am assuming this should include a getOption for the number of retries that a user would want.

@rtaph as for your suggestion on printing the output from the example, do you mean using print() to show in console?

Also do you recommend separating the package from the website documentation as it looks like it is throwing NOTEs related to the /doc folder size? Otherwise are there any other issues?
