### 2022-08-24 c33e703
Updating DESCRIPTION file to include Rafael Hellwig and Kyle Ueyama as reviewers of the package. Updated NEWS.md to include reviewers in changelog. Updating CITATION.cff and codemeta.json. Short description added to README.Rmd to include main functionality of USPS Crosswalk files within the rhud package. [no ci]
