+++ CITATION.cff
@@ -19,7 +19,7 @@ authors:
   email: emmet_tam@yahoo.com
 preferred-citation:
   type: manual
-  title: 'rhud: A R Interface to the HUD (US Department of Housing and Urban Development)
+  title: 'rhud: A R Interface for the HUD (US Department of Housing and Urban Development)
     APIs'
   authors:
   - family-names: Tam
@@ -48,6 +48,25 @@ keywords:
 - hud
 - income-limits
 references:
+- type: manual
+  title: 'rhud: A R interface for accessing HUD (US Department of Housing and Urban
+    Development) APIs'
+  authors:
+  - family-names: Tam
+    given-names: Emmet
+  - family-names: Reilly
+    given-names: Allison
+  - family-names: Ghaedi
+    given-names: Hamed
+  - family-names: Jin
+    given-names: Shuyu
+  year: '2022'
+  notes: 'To cite package ''rhud'' in publications use:'
+  publisher:
+    name: Zenodo
+  version: 0.3.0.9000
+  doi: 10.5281/zenodo.6644503
+  url: https://doi.org/10.5281/zenodo.6644503
 - type: software
   title: covr
   abstract: 'covr: Test Coverage for Packages'
@@ -118,17 +137,6 @@ references:
     orcid: https://orcid.org/0000-0003-3925-190X
   year: '2022'
   url: https://CRAN.R-project.org/package=rmarkdown
-- type: software
-  title: testthat
-  abstract: 'testthat: Unit Testing for R'
-  notes: Suggests
-  authors:
-  - family-names: Wickham
-    given-names: Hadley
-    email: hadley@rstudio.com
-  year: '2022'
-  url: https://CRAN.R-project.org/package=testthat
-  version: '>= 3.0.0'
 - type: software
   title: devtools
   abstract: 'devtools: Tools to Make Developing R Packages Easier'
@@ -316,6 +324,28 @@ references:
   year: '2022'
   url: https://CRAN.R-project.org/package=htmltools
 - type: software
+  title: htmlwidgets
+  abstract: 'htmlwidgets: HTML Widgets for R'
+  notes: Suggests
+  authors:
+  - family-names: Vaidyanathan
+    given-names: Ramnath
+  - family-names: Xie
+    given-names: Yihui
+  - family-names: Allaire
+    given-names: JJ
+  - family-names: Cheng
+    given-names: Joe
+    email: joe@rstudio.com
+  - family-names: Sievert
+    given-names: Carson
+    email: carson@rstudio.com
+    orcid: https://orcid.org/0000-0002-4958-2844
+  - family-names: Russell
+    given-names: Kenton
+  year: '2022'
+  url: https://CRAN.R-project.org/package=htmlwidgets
+- type: software
   title: httr
   abstract: 'httr: Tools for Working with URLs and HTTP'
   notes: Imports
@@ -337,39 +367,6 @@ references:
   year: '2022'
   url: https://CRAN.R-project.org/package=curl
 - type: software
-  title: zoo
-  abstract: 'zoo: S3 Infrastructure for Regular and Irregular Time Series (Z''s Ordered
-    Observations)'
-  notes: Imports
-  authors:
-  - family-names: Zeileis
-    given-names: Achim
-    email: Achim.Zeileis@R-project.org
-    orcid: https://orcid.org/0000-0003-0918-3766
-  - family-names: Grothendieck
-    given-names: Gabor
-    email: ggrothendieck@gmail.com
-  - family-names: Ryan
-    given-names: Jeffrey A.
-    email: jeff.a.ryan@gmail.com
-  year: '2022'
-  url: https://CRAN.R-project.org/package=zoo
-- type: software
-  title: rio
[…]
