### 2022-08-25 #524 @rtaph
Thanks for the update, @etam4260 ! It is totally reasonable to leave some suggestions for future consideration.

I will try to carve out time this weekend to look at your latest changes.

In the meantime:
1. Re: _"do you mean using print() to show in console?"_
    There are some vignettes where you call a function as an example but don't show the output. For instance, in the [state-level fair market rents](https://etam4260.github.io/rhud/articles/Fair-Market-Rents.html#state-level-fair-market-rents), we don't know what the result of `hud_fmr(query = 'VA', year = '2021')` is unless we run it ourselves in an R session. I am guessing maybe this was because the output is quite large to print? Even so, I think it would be very helpful  to show _some_ of the structure of the output. Something like
	  ``` r
	  tab <- rhud::hud_fmr(query = 'VA', year = '2021')
	  head(tab[[1]])
	  #>
	  #>   town_name      county_name
	  #> 1      NULL  Accomack County
	  #> 2      NULL Albemarle County
	  #> 3      NULL  Alexandria city
	  #> 4      NULL Alleghany County
	  #> 5      NULL    Amelia County
	  #> 6      NULL   Amherst County
	  #>                                                                     metro_name
	  #> 1                                                          Accomack County, VA
	  #> 2                                       Charlottesville, VA HUD Metro FMR Area
	  #> 3                 Washington-Arlington-Alexandria, DC-VA-MD HUD Metro FMR Area
	  #> 4 Alleghany County-Clifton Forge city-Covington city, VA HUD Nonmetro FMR Area
	  #> 5                                                             Richmond, VA MSA
	  #> 6                                                            Lynchburg, VA MSA
	  #>    fips_code Efficiency One-Bedroom Two-Bedroom Three-Bedroom Four-Bedroom
	  #> 1 5100199999        481         602         713           947          967
	  #> 2 5100399999        949        1077        1266          1575         1965
	  #> 3 5151099999       1513        1548        1765          2263         2742
	  #> 4 5100599999        495         558         735           958         1142
	  #> 5 5100799999        993        1020        1163          1538         1840
	  #> 6 5100999999        633         660         784          1053         1241
	  #>   FMR Percentile statename statecode smallarea_status query year
	  #> 1             40  Virginia        VA                0    VA 2021
	  #> 2             40  Virginia        VA                0    VA 2021
	  #> 3             40  Virginia        VA                1    VA 2021
	  #> 4             40  Virginia        VA                0    VA 2021
	  #> 5             40  Virginia        VA                0    VA 2021
	  #> 6             40  Virginia        VA                0    VA 2021
	  ```

	or

	  ``` r
	  str(rhud::hud_fmr(query = 'VA', year = '2021'))
	  #>
	  #> List of 2
	  #>  $ counties  :'data.frame':  133 obs. of  15 variables:
	  #>   ..$ town_name       : chr [1:133] "NULL" "NULL" "NULL" "NULL" ...
	  #>   ..$ county_name     : chr [1:133] "Accomack County" "Albemarle County" "Alexandria city" "Alleghany County" ...
	  #>   ..$ metro_name      : chr [1:133] "Accomack County, VA" "Charlottesville, VA HUD Metro FMR Area" "Washington-Arlington-Alexandria, DC-VA-MD HUD Metro FMR Area" "Alleghany County-Clifton Forge city-Covington city, VA HUD Nonmetro FMR Area" ...
	  #>   ..$ fips_code       : chr [1:133] "5100199999" "5100399999" "5151099999" "5100599999" ...
	  #>   ..$ Efficiency      : chr [1:133] "481" "949" "1513" "495" ...
	  #>   ..$ One-Bedroom     : chr [1:133] "602" "1077" "1548" "558" ...
	  #>   ..$ Two-Bedroom     : chr [1:133] "713" "1266" "1765" "735" ...
	  #>   ..$ Three-Bedroom   : chr [1:133] "947" "1575" "2263" "958" ...
	  #>   ..$ Four-Bedroom    : chr [1:133] "967" "1965" "2742" "1142" ...
	  #>   ..$ FMR Percentile  : chr [1:133] "40" "40" "40" "40" ...
	  #>   ..$ statename       : chr [1:133] "Virginia" "Virginia" "Virginia" "Virginia" ...
	  #>   ..$ statecode       : chr [1:133] "VA" "VA" "VA" "VA" ...
	  #>   ..$ smallarea_status: chr [1:133] "0" "0" "1" "0" ...
	  #>   ..$ query           : chr [1:133] "VA" "VA" "VA" "VA" ...
	  #>   ..$ year            : chr [1:133] "2021" "2021" "2021" "2021" ...
	  #>  $ metroareas:'data.frame':  19 obs. of  13 variables:
	  #>   ..$ metro_name      : chr [1:19] "Blacksburg-Christiansburg-Radford, VA HUD Metro FMR Area" "Buckingham County, VA HUD Metro FMR Area" "Charlottesville, VA HUD Metro FMR Area" "Culpeper County, VA HUD Metro FMR Area" ...
	  #>   ..$ code            : chr [1:19] "METRO13980M13980" "METRO16820N51029" "METRO16820M16820" "METRO47900N51047" ...
	  #>   ..$ Efficiency      : chr [1:19] "795" "564" "949" "788" ...
	  #>   ..$ One-Bedroom     : chr [1:19] "858" "652" "1077" "794" ...
	  #>   ..$ Two-Bedroom     : chr [1:19] "978" "743" "1266" "1046" ...
	  #>   ..$ Three-Bedroom   : chr [1:19] "1400" "1007" "1575" "1439" ...
	  #>   ..$ Four-Bedroom    : chr [1:19] "1693" "1140" "1965" "1811" ...
	  #>   ..$ FMR Percentile  : chr [1:19] "40" "40" "40" "40" ...
	  #>   ..$ statename       : chr [1:19] "Virginia" "Virginia" "Virginia" "Virginia" ...
	  #>   ..$ statecode       : chr [1:19] "VA" "VA" "VA" "VA" ...
	  #>   ..$ smallarea_status: chr [1:19] "0" "0" "0" "0" ...
	  #>   ..$ query           : chr [1:19] "VA" "VA" "VA" "VA" ...
	  #>   ..$ year            : chr [1:19] "2021" "2021" "2021" "2021" ...
	  ```

2. Re: _"do you recommend separating the package from the website documentation?"_
	That is a great question. I don't know what the best practice is when the `/docs` throw a NOTE due to size. @jhollist , do you have an opinion?
