### 2022-08-25 #524 @mpadge
> Re: "do you recommend separating the package from the website documentation?"
> That is a great question. I don't know what the best practice is when the /docs throw a NOTE due to size.

@rtaph You don't really need to worry, because as soon as the package is transferred the docs will be built by the internal rOpenSci doc server, and you can just remove the whole /docs folder anyway.  See the [*Packaging* chapter of *Dev Guide*](https://devguide.ropensci.org/building.html#website) for details.
