### 2022-08-27 b229ff3
Fixing bug where all error messages are not printed when there are more than 1.
