### 2022-08-27 9f40b17
Fixing option(rhud_quiet_loads) inverted logic.
