### 2022-08-27 1086fd0
Adding missed logic for checking NULL dataframe when renaming column in hud_cw_zip_cbsadiv() function. [no ci]
