+++ man/grapes-z_in_cbsadiv-grapes.Rd
@@ -35,9 +35,7 @@ take a look at their non-infix versions.
 \examples{
 \dontrun{
 
-# HUD API is broken for year 2021, quarter 3 for this query
-
-71052 \%z_in_cbsadiv\% 43340
+"03884" \%z_in_cbsadiv\% 40484
 
 }
 }
