### 2022-08-27 c80f168
hud_rec...() functions now return named lists vs named vectors. Internal changes to use named arguments returnfrom these functions rather than indexing in geo in geo functions
