+++ R/hud_geo_is_in_geo.R
@@ -41,16 +41,10 @@
 z_in_trt <- function(zip, tract, year, quarter,
                      key = Sys.getenv("HUD_KEY")) {
 
-  if (missing(year) || missing(quarter)) {
-    args <- hud_rec_cw_yr()
-    if (missing(year)) {
-      year <- args[1]
-    }
-
-    if (missing(quarter)) {
-      quarter <- args[2]
-    }
-  }
+  year <- geo_is_in_geo_most_rec_year_quarter(year = year,
+                                              quarter = quarter)$year
+  quarter <- geo_is_in_geo_most_rec_year_quarter(year = year,
+                                                 quarter = quarter)$quarter
 
   # Need to validate tract..
   cleaned <- cw_input_check_cleansing(primary_geoid = "tract",
@@ -138,16 +132,10 @@ z_in_trt <- function(zip, tract, year, quarter,
 z_in_cty <- function(zip, county, year, quarter,
                      key = Sys.getenv("HUD_KEY")) {
 
-  if (missing(year) || missing(quarter)) {
-    args <- hud_rec_cw_yr()
-    if (missing(year)) {
-      year <- args[1]
-    }
-
-    if (missing(quarter)) {
-      quarter <- args[2]
-    }
-  }
+  year <- geo_is_in_geo_most_rec_year_quarter(year = year,
+                                              quarter = quarter)$year
+  quarter <- geo_is_in_geo_most_rec_year_quarter(year = year,
+                                                 quarter = quarter)$quarter
 
   cleaned <- cw_input_check_cleansing(primary_geoid = "county",
                                       secondary_geoid = "zip",
@@ -234,16 +222,10 @@ z_in_cty <- function(zip, county, year, quarter,
 z_in_cbsa <- function(zip, cbsa, year, quarter,
                       key = Sys.getenv("HUD_KEY")) {
 
-  if (missing(year) || missing(quarter)) {
-    args <- hud_rec_cw_yr()
-    if (missing(year)) {
-      year <- args[1]
-    }
-
-    if (missing(quarter)) {
-      quarter <- args[2]
-    }
-  }
+  year <- geo_is_in_geo_most_rec_year_quarter(year = year,
+                                              quarter = quarter)$year
+  quarter <- geo_is_in_geo_most_rec_year_quarter(year = year,
+                                                 quarter = quarter)$quarter
 
   cleaned <- cw_input_check_cleansing(primary_geoid = "cbsa",
                                       secondary_geoid = "zip",
@@ -331,16 +313,10 @@ z_in_cbsa <- function(zip, cbsa, year, quarter,
 z_in_cbsadiv <- function(zip, cbsadiv, year, quarter,
                          key = Sys.getenv("HUD_KEY")) {
 
-  if (missing(year) || missing(quarter)) {
-    args <- hud_rec_cw_yr()
-    if (missing(year)) {
-      year <- args[1]
-    }
-
-    if (missing(quarter)) {
-      quarter <- args[2]
-    }
-  }
+  year <- geo_is_in_geo_most_rec_year_quarter(year = year,
+                                              quarter = quarter)$year
+  quarter <- geo_is_in_geo_most_rec_year_quarter(year = year,
+                                                 quarter = quarter)$quarter
 
   cleaned <- cw_input_check_cleansing(primary_geoid = "cbsadiv",
                                       secondary_geoid = "zip",
@@ -428,16 +404,10 @@ z_in_cbsadiv <- function(zip, cbsadiv, year, quarter,
 z_in_ctysb <- function(zip, countysub, year, quarter,
                        key = Sys.getenv("HUD_KEY")) {
 
-  if (missing(year) || missing(quarter)) {
[…]
