### 2022-08-27 9b7c674
Added feature where users can now set the number of RETRYs done before a query has failed. Updated documentation for the package description and startup message to include the new option.
