+++ R/zzz.R
@@ -11,12 +11,19 @@
 #' rhud is a R library that allows users to easily download data from the
 #' US Department of Housing and Urban Development APIs @@ HUD User.
 #' The data is defaulted to return measurements for the prior year
-#' (the year 365 days ago from the current system date). For details about
-#' this package as well as details about its python counterpart,
-#' visit https://github.com/etam4260/rhud
+#' (the year 365 days ago from the current system date) for most functions.
+#' For details about this package as well as details about its python
+#' counterpart, visit https://github.com/etam4260/rhud
 #'
 #' The default return type of data is a dataframe.
 #' To get tibbles instead, use: options(rhud_use_tibble = TRUE)
+#'
+#' To quiet the download bar, use: options(rhud_quiet_loads = TRUE)
+#'
+#' This interface uses RETRY over GET requests. By default, rhud will ping a
+#' URL 'twice' at most for a successful response. To set the number of retries
+#' used, use options(rhud_num_retrys = 1), where in this case the
+#' number '1' is a placeholder.
 #'
 #' @author Emmet Tam
 NULL
@@ -35,7 +42,7 @@ NULL
                                 "---------------------------------------------",
                                 "---------------------------------------------",
                                 "\n* To begin, first obtain a key ",
-                              "from HUD User website at `https://www.huduser.",
+                              "from HUD USER website at `https://www.huduser.",
                                 "gov/hudapi/public/login`\n",
                                 "* To set the key, use hud_set_key",
                                 "(\"your-key\")\n",
@@ -43,6 +50,10 @@ NULL
                                 "use `options(rhud_use_tibble = TRUE)`\n",
                                 "* To silence messages from downloads, use",
                                 " `options(rhud_quiet_loads = TRUE)`\n",
+                                "* To set the number of retries ",
+                                "used, use options(rhud_num_retrys = 1)\n",
+                                "* By default, rhud will ping a URL 'twice' at ",
+                                "most for a successful response.\n",
                                "* Caching is enabled by default. To set a new ",
                                 "cache ",
                                 "directory use hud_set_cache_dir().\n",
