### 2022-08-27 b1ec7f4
renaming hud_num_retrys to hud_num_retries and updating associated documentation + adding it to the NEWS.md file
