### 2022-08-27 #524 @etam4260
Thanks for the info @mpadge and good to know. Do you know if this is the same case for CRAN?

@rtaph Additional info: I think my testing suite is still a little bit incomplete as of now. I need to go back and make sure I test exact outputs versus determining if I get some output. In my case, the API for USPS Crosswalk API changed, but my tests didn't catch that. Also not at that 75% test coverage mark just yet 😅
