### 2022-08-28 3de4fc5
Adding tests for hud_rec...() functions. Fixing roxygen name for hud_rec_il_yr. Reducing straggling lines with > 80 characters. Renamed some files in /R to be more intuitive.
