+++ R/hud_crosswalk.R
@@ -1,28 +1,88 @@
-#' @name hud_cw_zip_tract
-#' @title Crosswalk Zip to Tract
-#' @description This function queries the USPS Crosswalks API provided by US
-#'   Department of Housing and Urban Development (HUD USER).
-#'   This returns the crosswalk for
-#'   zip to tract.
-#' @param zip A character or numeric vector: 5 digit
-#'   (United States Postal Service) USPS zipcode
-#'   of the data to retrieve. E.g. 22031 for type
-#'   1 to 5 and 11 .
-#' @param year A character or numeric vector: gets the year that this data
-#'   was recorded. Can specify multiple
-#'   years. Default is the previous year.
-#' @param quarter A character or numeric vector: gets the quarter of the year
-#'   that this data was recorded.
+#' @import tibble
+
+
+#' @name crosswalk
+#' @title Crosswalk a Dataset
+#' @description Using the US Housing and Urban Development USPS Crosswalk files,
+#'  crosswalk an entire dataset.
+#'  Currently supported crosswalks:
+#'   1) zip-tract
+#'   2) zip-county
+#'   3) zip-cbsa
+#'   4) zip-cbsadiv (Available 4th Quarter 2017 onwards)
+#'   5) zip-cd
+#'   6) tract-zip
+#'   7) county-zip
+#'   8) cbsa-zip
+#'   9) cbsadiv-zip (Available 4th Quarter 2017 onwards)
+#'   10) cd-zip
+#'   11) zip-countysub (Available 2nd Quarter 2018 onwards)
+#'   12) countysub-zip (Available 2nd Quarter 2018 onwards)
+#' @param data A dataframe or tibble with rows describing measurements at a zip,
+#'   county, county subdivision (countysub), congressional district (cd),
+#'   census tract, core base statistical area (cbsa), or core based
+#'   statistical area division (cbsadiv) geographic identifier.
+#'   1) zip
+#'   2) tract
+#'   3) county
+#'   4) countysub
+#'   5) cbsa
+#'   6) cbsadiv
+#'   7) cd
+#' @param geoid A character vector describing the current geoid
+#'   that the dataset is described in: must be
+#'   zip, county, countysub, cd,
+#'   tract, cbsa, or cbsadiv geographic identifier.
+#'   1) zip
+#'   2) tract
+#'   3) county
+#'   4) countysub
+#'   5) cbsa
+#'   6) cbsadiv
+#'   7) cd
+#' @param geoid_col A character or numeric vector of length one:
+#'   the column containing the geographic identifier; must be
+#'   zip, county, county subdivision (countysub), congressional district (cd),
+#'   census tract, core base statistical area (cbsa), and core based
+#'   statistical area division (cbsadiv) geographic identifier.
+#'   Supply either the name of the column or the index.
+#'   All elements in this column must be numbers only at the proper length.
+#'   For example, zip codes must be 5 digit numbers.
+#' @param cw_geoid A character vector of length one: the geoid to crosswalk
+#'   the dataset to; must be zip, county, county subdivision (countysub),
+#'   congressional district (cd),
+#'   census tract, core base statistical area (cbsa), or core based
+#'   statistical area division (cbsadiv) geoid.
+#'   1) zip
+#'   2) tract
+#'   3) county
+#'   4) countysub
+#'   5) cbsa
+#'   6) cbsadiv
+#'   7) cd
+#' @param cw_geoid_col A character or numeric vector: the columns in the dataset
+#'   to distribute according to method ratio.
+#'   If method is empty, no allocation method will be applied --
+#'   the crosswalk file will just be merged to the dataset.
+#'   All elements in these columns must be numbers only.
[…]
