+++ R/hud_decomp_crosswalk.R
@@ -1,0 +1,1231 @@
+#' @name hud_cw_zip_tract
+#' @title Crosswalk Zip to Tract
+#' @description This function queries the USPS Crosswalks API provided by US
+#'   Department of Housing and Urban Development (HUD USER).
+#'   This returns the crosswalk for
+#'   zip to tract.
+#' @param zip A character or numeric vector: 5 digit
+#'   (United States Postal Service) USPS zipcode
+#'   of the data to retrieve. E.g. 22031 for type
+#'   1 to 5 and 11 .
+#' @param year A character or numeric vector: gets the year that this data
+#'   was recorded. Can specify multiple
+#'   years. Default is the previous year.
+#' @param quarter A character or numeric vector: gets the quarter of the year
+#'   that this data was recorded.
+#'   Defaults to the first quarter of the year.
+#' @param minimal A logical: return just the crosswalked geoids if TRUE.
+#'   Otherwise, return
+#'   all fields. This does not remove duplicates.
+#' @param key A character vector of length one with the key obtained from HUD
+#'   (US Department of Housing and Urban Development)
+#'   USER website.
+#' @param to_tibble A logical: if TRUE, return the data in a tibble format
+#'   rather than a data frame.
+#' @keywords Crosswalks API
+#' @seealso
+#' * [rhud::crosswalk()]
+#' * [rhud::hud_cw_zip_tract()]
+#' * [rhud::hud_cw_zip_county()]
+#' * [rhud::hud_cw_zip_cbsa()]
+#' * [rhud::hud_cw_zip_cbsadiv()]
+#' * [rhud::hud_cw_zip_countysub()]
+#' * [rhud::hud_cw_zip_cd()]
+#' * [rhud::hud_cw_tract_zip()]
+#' * [rhud::hud_cw_county_zip()]
+#' * [rhud::hud_cw_cbsa_zip()]
+#' * [rhud::hud_cw_cbsadiv_zip()]
+#' * [rhud::hud_cw_cd_zip()]
+#' * [rhud::hud_cw_countysub_zip()]
+#' * [rhud::hud_cw()]
+#' @export
+#' @returns This function returns a dataframe containing crosswalk data for
+#'   zip to tract for all combinations of "zip", "year", and "quarter"
+#'   inputs.
+#'
+#'   These measurements include res-ratio, bus-ratio,
+#'   oth-ratio, tot-ratio. For more details on these measurements, visit
+#'   https://www.huduser.gov/portal/dataset/uspszip-api.html
+#'
+#' @examples
+#' \dontrun{
+#'
+#' hud_cw_zip_tract(zip = '35213', year = c('2010'), quarter = c('1'))
+#'
+#' hud_cw_zip_tract(zip = '35213', year = c('2010'), quarter = c('1'),
+#'    minimal = TRUE)
+#' }
+hud_cw_zip_tract <- function(zip, year = format(Sys.Date() - 365, "%Y"),
+                             quarter = 1, minimal = FALSE,
+                             key = Sys.getenv("HUD_KEY"),
+                             to_tibble = getOption("rhud_use_tibble", FALSE)) {
+
+  is_internet_available()
+
+  res <- NULL
+  primary_geoid <- "zip"
+  secondary_geoid <- "tract"
+
+  args <- cw_input_check_cleansing(primary_geoid, secondary_geoid, zip,
+                                   year, quarter, key)
+
+  zip <- args$query
+  year <- args$year
+  quarter <- args$quarter
+  key <- args$key
+
+  if (any(nchar(zip) != 5)) stop("\nZipcode inputs are not all of length 5",
+                                 call. = FALSE)
+
+  all_queries <- expand.grid(query = zip, year = year,
+                            quarter = quarter, stringsAsFactors = FALSE)
+
[…]
