### 2022-08-28 e2437d3
adding test for rhud_sitrep() function [no ci]
