+++ tests/testthat/test-sitrep.R
@@ -1,0 +1,3 @@
+test_that("rhud_sitrep() Output", {
+  expect_message(rhud_sitrep(), regexp = NULL)
+})
