### 2022-08-31 7f7e7b9
Fixing issue causing warnings to not print properly when rhud_quiet_loads is true. Refactored crosswalk() function to utilize switch statements over nested ifs.
