+++ R/hud_crosswalk.R
@@ -119,12 +119,17 @@
 #'           year = 2018, quarter = 1)
 #'
 #' }
-crosswalk <- function(data, geoid, geoid_col, cw_geoid, cw_geoid_col = NA,
+crosswalk <- function(data,
+                      geoid,
+                      geoid_col,
+                      cw_geoid,
+                      cw_geoid_col = NA,
                       method = NA,
                       year = format(Sys.Date() - 365, "%Y"),
                       quarter = 1,
                       key = Sys.getenv("HUD_KEY"),
                       to_tibble = getOption("rhud_use_tibble", FALSE)) {
+
   is_internet_available()
   result <- NULL
 
@@ -145,54 +150,45 @@ crosswalk <- function(data, geoid, geoid_col, cw_geoid, cw_geoid_col = NA,
   quarter <- args$quarter
   key <- args$key
 
+  cw_data <- switch(
+    geoid,
 
-  if (geoid == "zip" && cw_geoid %in% c("county", "countysub", "tract",
-                                       "cbsa", "cbsadiv", "cd")) {
-    if (cw_geoid == "county") {
-      cw_data <- hud_cw_zip_county(data[, geoid_col], year = year,
-                                   quarter = quarter, key = key)
-    } else if (cw_geoid == "countysub") {
-      cw_data <- hud_cw_zip_countysub(data[, geoid_col], year = year,
+    "zip" = switch(
+      cw_geoid,
+      "county" =    hud_cw_zip_county(data[, geoid_col], year = year,
+                                      quarter = quarter, key = key),
+      "countysub" = hud_cw_zip_countysub(data[, geoid_col], year = year,
+                                      quarter = quarter, key = key),
+      "cd" =        hud_cw_zip_cd(data[, geoid_col], year = year,
+                                      quarter = quarter, key = key),
+      "tract" =     hud_cw_zip_tract(data[, geoid_col], year = year,
+                                      quarter = quarter, key = key),
+      "cbsa" =      hud_cw_zip_cbsa(data[, geoid_col], year = year,
+                                      quarter = quarter, key = key),
+      "cbsadiv" =   hud_cw_zip_cbsadiv(data[, geoid_col], year = year,
                                       quarter = quarter, key = key)
-    } else if (cw_geoid == "cd") {
-      cw_data <- hud_cw_zip_cd(data[, geoid_col], year = year,
-                               quarter = quarter, key = key)
-    } else if (cw_geoid == "tract") {
-      cw_data <- hud_cw_zip_tract(data[, geoid_col], year = year,
-                                  quarter = quarter, key = key)
-    } else if (cw_geoid == "cbsa") {
-      cw_data <- hud_cw_zip_cbsa(data[, geoid_col], year = year,
-                                 quarter = quarter, key = key)
-    } else if (cw_geoid == "cbsadiv") {
-      cw_data <- hud_cw_zip_cbsadiv(data[, geoid_col], year = year,
-                                    quarter = quarter, key = key)
-    }
-  } else if (geoid == "county" && cw_geoid == "zip") {
-    cw_data <- hud_cw_county_zip(data[, geoid_col], year = year,
-                                 quarter = quarter, key = key)
-  } else if (geoid == "countysub"  && cw_geoid == "zip") {
-    cw_data <- hud_cw_countysub_zip(data[, geoid_col], year = year,
-                                    quarter = quarter, key = key)
-  } else if (geoid == "cd"  && cw_geoid == "zip") {
-    cw_data <- hud_cw_cd_zip(data[, geoid_col], year = year,
-                             quarter = quarter, key = key)
[…]
