### 2022-09-02 9818a8b
Added smart_rbind function to deal with datasets with overlapping columnnames, but may have others in them too. Updated some test cases in test-chas, test-cw, and test-cw_a_dataset to test exact number of columns.
