### 2022-09-03 8e399c9
Fixing issues in tests where a cw test tested length on empty object. Fixed chas test where variables were references incorrectly.
