+++ tests/testthat/test-cw.R
@@ -9,38 +9,38 @@ test_that("hud_cw() for All Types", {
   # Lets try a ZIP code in Alabama for now for 1 -> 5 as well as 11.
   zip_tract <- hud_cw(type = 1, query = "35213",
                       year = c("2010", "2011"), quarter = c("1"))
-  expect_equal(length(zip_tract), 9)
+  expect_equal(length(zip_tract), 8)
 
   expect_true(nrow(zip_tract) >= 1)
 
   zip_county <- hud_cw(type = "2", query = "35213",
                        year = c("2016", "2020"), quarter = c("2"))
-  expect_equal(length(zip_county), 9)
+  expect_equal(length(zip_county), 8)
 
   expect_true(nrow(zip_county) >= 1)
 
   zip_cbsa <- hud_cw(type = 3, query = 35213,
                      year = c("2012", "2011"), quarter = c("3"))
-  expect_equal(length(zip_cbsa), 9)
+  expect_equal(length(zip_cbsa), 8)
 
   expect_true(nrow(zip_cbsa) >= 1)
 
   zip_cbsadiv <- hud_cw(type = 4, query = "22031",
                         year = c("2017", "2019"), quarter = c("4"))
-  expect_equal(length(zip_cbsadiv), 9)
+  expect_equal(length(zip_cbsadiv), 8)
 
   expect_true(nrow(zip_cbsadiv) >= 1)
 
   zip_cd <- hud_cw(type = "5", query = "35213",
                    year = c(2011, "2012"), quarter = c("1", "2"))
-  expect_equal(length(zip_cd), 9)
+  expect_equal(length(zip_cd), 8)
 
   expect_true(nrow(zip_cd) >= 1)
 
   tract_zip <- hud_cw(type = 6, query = "48201223100",
                       year = c("2017", "2010"),
                       quarter = c("1", "2", "3"))
-  expect_equal(length(tract_zip), 9)
+  expect_equal(length(tract_zip), 8)
 
   expect_true(nrow(tract_zip) >= 1)
 
@@ -48,7 +48,7 @@ test_that("hud_cw() for All Types", {
   county_zip <- hud_cw(type = 7, query = "22031",
                        year = c("2010", "2011"),
                        quarter = c("1", "2", "3", "4"))
-  expect_equal(length(county_zip), 9)
+  expect_equal(length(county_zip), 8)
 
   expect_true(nrow(county_zip) >= 1)
 
@@ -56,7 +56,7 @@ test_that("hud_cw() for All Types", {
   county_zip <- hud_cw(type = 7,
                        query = substr(hud_state_counties("md")$fips_code, 1, 5),
                        year = c("2010"), quarter = c("1"))
-  expect_equal(length(county_zip), 9)
+  expect_equal(length(county_zip), 8)
 
   expect_true(nrow(county_zip) >= 1)
 
@@ -64,7 +64,7 @@ test_that("hud_cw() for All Types", {
   # CBSA defines Micropolitan and Metropolitan
   cbsa_zip <- hud_cw(type = 8, query = "10140",
                      year = c("2010", "2011"), quarter = c("2", "1"))
-  expect_equal(length(cbsa_zip), 9)
+  expect_equal(length(cbsa_zip), 8)
 
   expect_true(nrow(cbsa_zip) >= 1)
 
@@ -72,20 +72,20 @@ test_that("hud_cw() for All Types", {
   # metropolitan areas.
   cbsadiv_zip <- hud_cw(type = 9, query = "10380",
                         year = c("2017"), quarter = c("4"))
-  expect_equal(length(cbsadiv_zip), 9)
+  expect_equal(length(cbsadiv_zip), 8)
 
[…]
