+++ tests/testthat/test-chas.R
@@ -41,22 +41,22 @@ test_that("hud_chas() for Different Years", {
   # 2006-2010
 
   y1 <- hud_chas(1, year = c("2014-2018", "2013-2017"))
-  expect_true(length(city) == 132)
+  expect_true(length(y1) == 132)
   expect_true(nrow(y1) >= 1)
 
   y2 <- hud_chas(1, year = c("2013-2017"))
-  expect_true(length(city) == 132)
+  expect_true(length(y2) == 132)
   expect_true(nrow(y2) >= 1)
 
   y3 <- hud_chas(1, year = c("2008-2012", "2007-2011", "2014-2018"))
-  expect_true(length(city) == 132)
+  expect_true(length(y3) == 132)
   expect_true(nrow(y3) >= 1)
 
   y4 <- hud_chas(1, year = c("2012-2016", "2007-2011"))
-  expect_true(length(city) == 132)
+  expect_true(length(y4) == 132)
   expect_true(nrow(y4) >= 1)
 
   y5 <- hud_chas(1, year = c("2014-2018", "2013-2017"))
-  expect_true(length(city) == 132)
+  expect_true(length(y5) == 132)
   expect_true(nrow(y5) >= 1)
 })
