### 2022-09-06 #524 @jhollist
@etam4260 Just now getting back to this!

CRAN won't see your docs folder as you currently have it listed in your .Rbuildignore file.  That is the way you want it.  The documentation website will be independent of anything that happens on CRAN.  I think I would follow @mpadge advice and remove the docs folder.  rOpenSci will take care of building that for you.

At least that is what I can tell from digging around a bit.  I haven't used pkdgown sites much myself so don't have a ton of personal experience here.
