### 2022-10-01 683cc76
Adjusting test conditions for crosswalk function calls to include exact dataframe rows and list length, for valid queries.
