+++ tests/testthat/test-decomp_cw.R
@@ -1,7 +1,7 @@
 test_that("hud_cw_zip_tract() Simple Example and Error", {
   skip_if_no_key()
   res <- hud_cw_zip_tract(zip = "35213", year = c("2010"), quarter = c("1"))
-  expect_true(nrow(res) >= 1)
+  expect_true(nrow(res) == 9)
 
   expect_error(hud_cw_zip_tract(zip = "3521334",
                                 year = c("2010"), quarter = c("1")))
@@ -10,13 +10,13 @@ test_that("hud_cw_zip_tract() Simple Example and Error", {
 
   res <- hud_cw_zip_tract(zip = "35213", year = c("2010"),
                           quarter = c("1"), minimal = TRUE)
-  expect_true(is.vector(res) && length(res) >= 1)
+  expect_true(is.vector(res) && length(res) == 9)
 })
 
 test_that("hud_cw_zip_county() Simple Example and Error", {
   skip_if_no_key()
   res <- hud_cw_zip_county(zip = 35213, year = c("2020"), quarter = c("2"))
-  expect_true(nrow(res) >= 1)
+  expect_true(nrow(res) == 1)
 
   expect_error(hud_cw_zip_county(zip = "3521334",
                                  year = c("2010"), quarter = c("1")))
@@ -25,13 +25,13 @@ test_that("hud_cw_zip_county() Simple Example and Error", {
 
   res <- hud_cw_zip_county(zip = "35213", year = c("2010"),
                            quarter = c("1"), minimal = TRUE)
-  expect_true(is.vector(res) && length(res) >= 1)
+  expect_true(is.vector(res) && length(res) == 1)
 })
 
 test_that("hud_cw_zip_cbsa() Simple Example and Error", {
   skip_if_no_key()
   res <- hud_cw_zip_cbsa(zip = 35213, year = c("2020"), quarter = c("2"))
-  expect_true(nrow(res) >= 1)
+  expect_true(nrow(res) == 1)
 
   expect_error(hud_cw_zip_cbsa(zip = "3521334",
                                year = c("2010"), quarter = c("1")))
@@ -40,13 +40,13 @@ test_that("hud_cw_zip_cbsa() Simple Example and Error", {
 
   res <- hud_cw_zip_cbsa(zip = "35213", year = c("2010"),
                          quarter = c("1"), minimal = TRUE)
-  expect_true(is.vector(res) && length(res) >= 1)
+  expect_true(is.vector(res) && length(res) == 1)
 })
 
 test_that("hud_cw_zip_cbsadiv() Simple Example and Error", {
   skip_if_no_key()
   res <- hud_cw_zip_cbsadiv(zip = 22031, year = c("2020"), quarter = c("2"))
-  expect_true(nrow(res) >= 1)
+  expect_true(nrow(res) == 1)
 
   expect_error(hud_cw_zip_cbsadiv(zip = "3521334",
                                   year = c("2010"), quarter = c("1")))
@@ -55,13 +55,13 @@ test_that("hud_cw_zip_cbsadiv() Simple Example and Error", {
 
   res <- hud_cw_zip_cbsadiv(zip = "22031", year = c("2020"),
                             quarter = c("2"), minimal = TRUE)
-  expect_true(is.vector(res) && length(res) >= 1)
+  expect_true(is.vector(res) && length(res) == 1)
 })
 
 test_that("hud_cw_zip_cd() Simple Example and Error", {
   skip_if_no_key()
   res <- hud_cw_zip_cd(zip = 35213, year = c("2020"), quarter = c("2"))
-  expect_true(nrow(res) >= 1)
+  expect_true(nrow(res) == 8)
 
   expect_error(hud_cw_zip_cd(zip = "3521334", year = c("2010"),
[…]
