### 2022-10-01 44697cd
Adjusting test conditions for decomposed fair markets rent function calls to include exact dataframe rows and list length, for valid queries.
