### 2022-10-01 1ac6f7e
Adjusting test conditions for non decomposed crosswalk function calls to include exact dataframe rows and list length, for valid queries.
