### 2022-10-01 2d6c6cd
Adjusting test conditions for non decomposed fair markets rent function calls to include exact dataframe rows and list length, for valid queries.
