### 2022-11-01 fa99a28
Removing example of cache directory setting as it might affect testing/mess with user's files/folders
