### 2022-11-01 c99ae50
Added new 429 error handling for query calls, causing issues with missing data because of too many requests. Adjusted RETRY verb to include a 60 second wait time after a 429 error when server wants client to wait.
