+++ R/hud_pkg_env.R
@@ -1,2 +1,3 @@
 # Create a package environment for storing state level data.
+
 pkg.env <- new.env(parent = emptyenv())
