### 2024-10-28 #524 @adamhsparks
Hi @etam4260, are you still interested in completing these changes or should I close this issue?
